% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_existing_tzar_run.R
\name{act_on_loaded_existing_tzar_run}
\alias{act_on_loaded_existing_tzar_run}
\title{Execute a single action based on tzar repetition values in project.yaml}
\usage{
act_on_loaded_existing_tzar_run(tgt_filename_or_dirname_for_scp, parameters)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}
}
\value{
Returns nothing.
}
\description{
This is the main workhorse function for bdpg.  It's intended to generate
a single bdproblem of a chosen type (COR, WRAP, APP) or a single reserve
selector run based on inputs specified in the yaml file and its
repetitions section.  In general, tzar will spawn a process that calls this
function across a set of problems, e.g., generating 10 apparent problems
for each correct problem whose path is given in an input file.
}
