% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_error_to_spp_occupancy_data.R
\name{apply_const_error_to_spp_occupancy_data}
\alias{apply_const_error_to_spp_occupancy_data}
\title{Apply constant error to spp occ data}
\usage{
apply_const_error_to_spp_occupancy_data(
  cor_num_PUs,
  cor_num_spp,
  cor_bpm,
  FP_rates_matrix,
  FN_rates_matrix
)
}
\arguments{
\item{cor_num_PUs}{integer}

\item{cor_num_spp}{integer}

\item{cor_bpm}{matrix}

\item{FP_rates_matrix}{float matrix of False Positive error rates for each
species/planning unit combination}

\item{FN_rates_matrix}{float matrix of False Negative error rates for each
species/planning unit combination}
}
\value{
Returns ret_vals_from_apply_errors list
}
\description{
Apply constant error to spp occ data
}
