% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_9a_create_Xu_graph.R
\name{assert_edge_list_does_not_violate_assumptions}
\alias{assert_edge_list_does_not_violate_assumptions}
\title{Assert edge list does not violate assumptions}
\usage{
assert_edge_list_does_not_violate_assumptions(edge_list,
  first_row_of_intergroup_links, nodes)
}
\arguments{
\item{edge_list}{two column integer matrix of node IDs with one row for
each edge and columns for the 2 ends of the edge}

\item{first_row_of_intergroup_links}{integer row number of first row
containing intergroup links}

\item{nodes}{data frame containing integer node_ID, integer group_ID, and boolean dependent_set_member}
}
\value{
Returns a boolean
}
\description{
Assert edge list does not violate assumptions
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{cur_row}{
\preformatted{
cur_row :  int 814
}}
\subsection{edge_list}{
\preformatted{
edge_list :  int [1:814, 1:2] 1 3 5 7 9 11 13 15 17 19 ...
}}
\subsection{edge_list_error}{
\preformatted{
edge_list_error :  logi FALSE
}}
\subsection{first_row_of_intergroup_links}{
\preformatted{
first_row_of_intergroup_links :  num 62
}}
\subsection{from_node}{
\preformatted{
from_node :  Named int 54
}}
\subsection{nodes}{
\preformatted{
nodes : 'data.frame':	122 obs. of  3 variables:
 $ node_ID             : int  1 2 3 4 5 6 7 8 9 10 ...
 $ group_ID            : num  1 1 2 2 3 3 4 4 5 5 ...
 $ dependent_set_member: logi  FALSE TRUE FALSE TRUE FALSE TRUE ...
}}
\subsection{num_rows_in_edge_list}{
\preformatted{
num_rows_in_edge_list :  int 814
}}
\subsection{to_node}{
\preformatted{
to_node :  Named int 112
}}
}

