% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{build_PU_spp_pair_indices_from_occ_matrix}
\alias{build_PU_spp_pair_indices_from_occ_matrix}
\title{Build PU spp pair indices from occ matrix}
\usage{
build_PU_spp_pair_indices_from_occ_matrix(occ_matrix, num_PUs, num_spp)
}
\arguments{
\item{occ_matrix}{matrix}

\item{num_PUs}{integer}

\item{num_spp}{integer}
}
\value{
Returns PU_spp_pair_indices data frame
}
\description{
Build planning unit vs. species indices from occupancy matrix.
}
