% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{build_topdir_name}
\alias{build_topdir_name}
\title{Build top directory name}
\usage{
build_topdir_name(obj)
}
\arguments{
\item{obj}{an object such as a Xu_bd_problem or an RSrun}
}
\value{
Returns a directory name (not a path)
}
\description{
Builds a directory name for the highest level directory for a bdpg-related
object such as a reserve selection problem.
}
