% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14a_compute_marxan_solution_scores.R
\name{choose_a_best_solution_ID_according_to_marxan}
\alias{choose_a_best_solution_ID_according_to_marxan}
\title{Choose a best solution ID according to marxan}
\usage{
choose_a_best_solution_ID_according_to_marxan(num_marxan_solutions,
  marxan_solutions_matrix, marxan_best_df_sorted_as_vector)
}
\description{
Marxan returns a best solution, but I have not been able to find
 anyplace where it tells you what solution number it was.
 Since you can have multiple identical solutions, there may be
 multiple solution IDs that could be identified as the best solution.
 I need any one of them to use to get the corresponding solution
 vector.
 I will arbitrarily choose the first one in the list of vectors that
 selected the same PUs as the vector that marxan returned.
}
