% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{close_open_subsection_if_necessary}
\alias{close_open_subsection_if_necessary}
\title{Write closing brackets for previous subsection if open brackets remain}
\usage{
close_open_subsection_if_necessary(prev_state)
}
\arguments{
\item{prev_state}{character string indicating the previous state, e.g.,
"state__block_start"}
}
\value{
Returns nothing
}
\description{
Variables are written out as subsections and that means they start with
brackets.  Finding a function name means that the var subsection brackets
need to be closed and this function does that by just spitting out two
brackets.
}
