% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{combine_2_bdprobs}
\alias{combine_2_bdprobs}
\title{FUNCTION_TITLE}
\usage{
combine_2_bdprobs(bdprob_1, bdprob_2, parameters, bdpg_error_codes, integerize)
}
\arguments{
\item{bdprob_1}{DESCRIPTION.}

\item{bdprob_2}{DESCRIPTION.}

\item{parameters}{DESCRIPTION.}

\item{bdpg_error_codes}{DESCRIPTION.}

\item{integerize}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

