% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob_APP.R
\name{compute_and_save_dist_and_network_metrics_for_prob}
\alias{compute_and_save_dist_and_network_metrics_for_prob}
\title{Compute and save the distribution and network metrics for the problem.}
\usage{
compute_and_save_dist_and_network_metrics_for_prob(Xu_bdprob_APP, Xu_bdprob_COR,
  starting_dir, parameters)
}
\arguments{
\item{Xu_bdprob_APP}{an apparent Xu_bd_problem}

\item{Xu_bdprob_COR}{a Xu_bd_problem}

\item{starting_dir}{a character string}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}
}
\value{
an apparent Xu_bd_problem
}
\description{
Compute and save the distribution and network metrics for the problem.
}
