% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{compute_and_set_obj_checksum}
\alias{compute_and_set_obj_checksum}
\title{Convenience function for computing the object's checksum and setting that slot}
\usage{
compute_and_set_obj_checksum(obj, base_outdir = ".")
}
\arguments{
\item{obj}{object whose checksum is to be computed}

\item{base_outdir}{directory where temporary file for use in checksum
computation will be written and then deleted}
}
\value{
the same object that was passed in except that its checksum is
    now set
}
\description{
This function computes the checksum and sets the checksum slot in the object,
then returns the modified object.  It just packages those up for convenience.
}
