% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14b_compute_solution_scores.R
\name{compute_and_verify_rep_scores_wrt}
\alias{compute_and_verify_rep_scores_wrt}
\title{Compute representation scores for candidate solution}
\usage{
compute_and_verify_rep_scores_wrt(ref_spp_occ_matrix, cand_sol_PU_IDs,
  spp_rep_targets, num_spp)
}
\arguments{
\item{ref_spp_occ_matrix}{reference species occupancy matrix, e.g.,
correct or apparent species occupancy matrix}

\item{cand_sol_PU_IDs}{vector of only the planning unit IDs that are
included in the candidate solution}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}

\item{num_spp}{integer number of species in the problem}
}
\value{
list of solution vector scores
}
\description{
Scores are computed with respect to a reference spp occupancy matrix
(i.e., COR or APP)
}
\details{
The species occupancy matrix is referred to as the "ref_spp_occ_matrix" to
indicate that we're computing scores with respect to a given reference, i.e.,
either correct or apparent, rather than always computing the correct score.

This is to allow us to see the difference between how well a method is
actually doing and how it might appear to be doing when its performance is
measured against data of unknown correctness, which is how nearly all results
are presented in the literature.
}
