% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_11b_network_measures_using_igraph_package.R
\name{compute_igraph_related_network_measures}
\alias{compute_igraph_related_network_measures}
\title{Compute igraph-related network measures}
\usage{
compute_igraph_related_network_measures(rsprob, top_dir)
}
\arguments{
\item{rsprob}{a reserve selection problem}

\item{top_dir}{character string}
}
\description{
Compute igraph-related network measures
}
