% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14a_compute_marxan_solution_scores.R
\name{compute_marxan_solution_scores}
\alias{compute_marxan_solution_scores}
\title{Compute marxan solution scores}
\usage{
compute_marxan_solution_scores(spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU,
  marxan_solution_PU_IDs, targets, num_spp, marxan_solutions_matrix,
  cur_solution_num, marxan_solution_scores, cor_PU_costs, total_landscape_cost)
}
\arguments{
\item{spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU}{matrix}

\item{marxan_solution_PU_IDs}{numeric vector}

\item{targets}{numeric vector}

\item{num_spp}{integer}

\item{marxan_solutions_matrix}{matrix}

\item{cur_solution_num}{integer}

\item{marxan_solution_scores}{data frame}

\item{cor_PU_costs}{numeric vector}

\item{total_landscape_cost}{numeric}
}
\value{
Returns marxan_solution_scores data frame
}
\description{
Compute marxan solution scores
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{cor_PU_costs}{
\preformatted{
cor_PU_costs :  num [1:407] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{cur_cost}{
\preformatted{
cur_cost :  num 61
}}
\subsection{cur_frac_of_all_spp_meeting_their_target}{
\preformatted{
cur_frac_of_all_spp_meeting_their_target :  num 1
}}
\subsection{cur_rep_fractions}{
\preformatted{
cur_rep_fractions :  num [1:1277] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{cur_solution_num}{
\preformatted{
cur_solution_num :  int 1
}}
\subsection{cur_solution_PUs}{
\preformatted{
cur_solution_PUs :  int [1:61] 2 4 6 8 10 12 14 16 18 20 ...
}}
\subsection{marxan_solution_PU_IDs}{
\preformatted{
marxan_solution_PU_IDs :  int [1:61] 2 4 6 8 10 12 14 16 18 20 ...
}}
\subsection{marxan_solution_scores}{
\preformatted{
marxan_solution_scores : 'data.frame':	4 obs. of  3 variables:
 $ solution_num  : int  1 2 3 4
 $ representation: num  1 0 0 0
 $ cost          : num  0.15 0 0 0
}}
\subsection{marxan_solutions_matrix}{
\preformatted{
marxan_solutions_matrix :  num [1:4, 1:407] 0 0 0 0 1 1 1 1 0 0 ...
}}
\subsection{num_spp}{
\preformatted{
num_spp :  int 1277
}}
\subsection{spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU}{
\preformatted{
spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU :  num [1:1277, 1:407] 1 0 0 0 0 0 0 0 0 0 ...
}}
\subsection{targets}{
\preformatted{
targets :  num [1:1277] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{total_landscape_cost}{
\preformatted{
total_landscape_cost :  num 407
}}
}

