% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14a_compute_marxan_solution_scores.R
\name{compute_marxan_solution_scores_wrt_COR_or_APP_reps_and_costs}
\alias{compute_marxan_solution_scores_wrt_COR_or_APP_reps_and_costs}
\title{Compute rep and cost values of marxan solutions with respect to COR or APP}
\usage{
compute_marxan_solution_scores_wrt_COR_or_APP_reps_and_costs(given_bpm,
  given_PU_costs, num_marxan_solutions, marxan_solutions_matrix, targets,
  num_spp)
}
\arguments{
\item{given_bpm}{blah}

\item{given_PU_costs}{blah}

\item{num_marxan_solutions}{blah}

\item{marxan_solutions_matrix}{blah}

\item{targets}{blah}

\item{num_spp}{blah}
}
\value{
blah
}
\description{
Go through the solutions matrix and compute the representation
 and cost values for each solution that marxan found in its different
 restarts.
}
