% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14b_compute_solution_scores.R
\name{compute_num_spp_covered_by_solution}
\alias{compute_num_spp_covered_by_solution}
\title{Compute number of species whose targets are met by a candidate solution}
\usage{
compute_num_spp_covered_by_solution(cand_sol_PU_IDs, bpm, spp_rep_targets)
}
\arguments{
\item{cand_sol_PU_IDs}{vector of only the planning unit IDs that are
included in the candidate solution}

\item{bpm}{bipartite matrix; integer matrix with one row for each species
and one column for each planning unit.  Each matrix entry specifies
whether that species occupies that planning unit; 1 indicates the
species does occupy the planning unit and 0 indicates it does not.
Same as occ_matrix.}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}
}
\value{
number of species whose targets are met by the candidate solution
}
\description{
Compute number of species whose targets are met by a candidate solution
}
