% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_error_to_spp_occupancy_data.R
\name{compute_realized_error_rates}
\alias{compute_realized_error_rates}
\title{Compute realized error rates in apparent problem
 Since errors are generated stochastically, the realized error rates
 in the apparent spp vs PU matrix are unlikely to exactly match
 the error generator's target rates, so measure the error rates
 that actually resulted after adding error.}
\usage{
compute_realized_error_rates(cor_bpm, app_bpm, target_FP_rate = NA,
  target_FN_rate = NA)
}
\description{
Compute realized error rates in apparent problem
 Since errors are generated stochastically, the realized error rates
 in the apparent spp vs PU matrix are unlikely to exactly match
 the error generator's target rates, so measure the error rates
 that actually resulted after adding error.
}
