% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_10b_compute_solution_rep_levels_and_costs.R
\name{compute_solution_cost}
\alias{compute_solution_cost}
\title{Compute cost of given solution vector of PUs to include}
\usage{
compute_solution_cost(PU_set_to_test, PU_costs)
}
\arguments{
\item{PU_set_to_test}{integer vector of planning unit IDs where each ID
specifies a planning unit to include in the set to be tested}

\item{PU_costs}{numeric vector of planning unit costs}
}
\value{
Returns numeric cost of the given solution
}
\description{
Given a set of planning units to include in a solution, compute the total
cost for that set.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{PU_costs}{
\preformatted{
PU_costs :  num [1:122] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{PU_set_to_test}{
\preformatted{
PU_set_to_test :  int [1:61] 2 4 6 8 10 12 14 16 18 20 ...
}}
\subsection{solution_cost}{
\preformatted{
solution_cost :  num 61
}}
}

