% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_10c_build_adj_and_cooccurrence_matrices.R
\name{create_adj_matrix_with_spp_rows_vs_PU_cols}
\alias{create_adj_matrix_with_spp_rows_vs_PU_cols}
\title{Create adjacency matrix with species rows vs planning unit columns}
\usage{
create_adj_matrix_with_spp_rows_vs_PU_cols(num_spp, num_PUs,
  PU_spp_pair_indices, PU_costs, spp_col_name, PU_col_name, dependent_node_IDs,
  correct_solution_vector_is_known)
}
\arguments{
\item{num_spp}{integer number of species in the problem}

\item{num_PUs}{integer number of planning units}

\item{PU_spp_pair_indices}{2 column data frame of PU IDs vs species IDs,
where each row identifies a the ID of a given species that
occurs on the given planning unit}

\item{PU_costs}{numeric vector of planning unit costs}

\item{spp_col_name}{character string giving species column name in data frames}

\item{PU_col_name}{character string giving planning unit column name in data frames}

\item{dependent_node_IDs}{integer vector of IDs of planning units contained
in a constructed correct solution to a Xu problem}

\item{correct_solution_vector_is_known}{boolean flag indicating whether
a correct optimal solution vector is known for the problem (as opposed
for example, to only knowing the correct cost)}
}
\value{
Returns bpm; integer matrix with one row for each species and one
    column for each planning unit.  Each matrix entry specifies whether
    that species occupies that planning unit; 1 indicates the species
    does occupy the planning unit and 0 indicates it does not.
}
\description{
Create numeric adjacency matrix with one row for each species and one
column for each planning unit and each matrix entry specifying whether
that species occupies that planning unit.  A 1 indicates the species
does occupy the planning unit and 0 indicates it does not.
}
