% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob.R
\name{create_allowable_size_Xu_problem_from_scratch}
\alias{create_allowable_size_Xu_problem_from_scratch}
\title{Generate a Xu problem from scratch but insure it's less than a given size}
\usage{
create_allowable_size_Xu_problem_from_scratch(max_allowed_num_spp, parameters,
  integerize, default_num_prob_size_retries_allowed = 20)
}
\arguments{
\item{max_allowed_num_spp}{integer maximum number of species allowed in the
problem (to keep from generating problems that are too large when trying
to keep the run-time down)}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}

\item{default_num_prob_size_retries_allowed}{integer number of retries of
problem creation to try to find one no larger than a given number of
species}
}
\value{
Returns a PU_spp_pair_info_class object or quits if number of
    allowed retries is exceeded
}
\description{
Generate a Xu biodiversity problem based on 4 input control parameters
rather than reading it from a file.
}
