% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_15_create_master_output_structure.R
\name{create_APP_master_output_structure}
\alias{create_APP_master_output_structure}
\title{Create a master output structure for all problem attributes and reserve
selection results when evaluating an APPARENT problem.}
\usage{
create_APP_master_output_structure(marxan_control_values, marxan_output_values,
  parameters, COR_bd_prob, APP_bd_prob, APP_marxan_run)
}
\description{
Create a master output structure for all problem attributes and reserve
selection results when evaluating an APPARENT problem.
}
