% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob_APP.R
\name{create_APP_prob_info_by_adding_error_to_spp_occ_data}
\alias{create_APP_prob_info_by_adding_error_to_spp_occ_data}
\title{Add error to the species occupancy data and save to APP_prob_info structure.}
\usage{
create_APP_prob_info_by_adding_error_to_spp_occ_data(
  Xu_bdprob_COR,
  Xu_bdprob_APP,
  parameters,
  gen_cost_errors,
  gen_FP_FN_errors,
  ret_vals_from_build_const_err = NULL,
  ret_vals_from_apply_cost_errors = NULL
)
}
\arguments{
\item{Xu_bdprob_COR}{a Xu_bd_problem}

\item{Xu_bdprob_APP}{an apparent Xu_bd_problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{gen_cost_errors}{boolean flag indicating whether to add cost error;
TRUE implies add cost error and FALSE implies don't add cost error}

\item{gen_FP_FN_errors}{boolean flag indicating whether to add false positives
and/or false negatives;
TRUE implies add FP/FN error and FALSE implies don't add FP/FN error}

\item{ret_vals_from_build_const_err}{list of return values from building
constant error values if already done, NULL if not already done}

\item{ret_vals_from_apply_cost_errors}{list of return values from building
cost error values if already done, NULL if not already done}
}
\value{
an apparent Xu_bd_problem
}
\description{
Add error to the species occupancy data and save to APP_prob_info structure.
}
