% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_15_create_master_output_structure.R
\name{create_COR_master_output_structure}
\alias{create_COR_master_output_structure}
\title{Create a master output structure for all problem attributes and reserve
selection results when evaluating a CORRECT problem.}
\usage{
create_COR_master_output_structure(COR_bd_prob, marxan_control_values,
  marxan_output_values, parameters)
}
\description{
When evaluating performance on a CORRECT problem, pass that CORRECT problem
in as both the CORRECT and the APPARENT problem.  The code to create the
master output structure puts out various APPARENT values and this dummying
in of the CORRECT for the APPARENT allows the function to work identically
for CORRECT and APPARENT data.
}

