% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob_APP.R
\name{create_dirs_for_APP_prob}
\alias{create_dirs_for_APP_prob}
\title{Create directories for this apparent problem.}
\usage{
create_dirs_for_APP_prob(
  Xu_bdprob_APP,
  basic_or_wrapped_or_comb_str,
  starting_dir
)
}
\arguments{
\item{Xu_bdprob_APP}{an apparent Xu_bd_problem}

\item{basic_or_wrapped_or_comb_str}{a character string}

\item{starting_dir}{a character string}
}
\value{
an apparent Xu_bd_problem
}
\description{
Create directories for this apparent problem.
}
