% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_new_res_sel_replicate_subtree}
\alias{create_new_res_sel_replicate_subtree}
\title{Create a directory subtree to hold output from a run of a reserve selector}
\usage{
create_new_res_sel_replicate_subtree(bdpg_dir_names, res_sel_name = "marxan",
  base_outdir = ".", create_dirs = TRUE)
}
\arguments{
\item{bdpg_dir_names}{nested list of paths to directories of bdpg output}

\item{res_sel_name}{character string giving name of reserve selector to add a subtree for, e.g., "marxan"}

\item{base_outdir}{character string giving path to base directory of bdpg output tree for this run}

\item{create_dirs}{boolean flag indicating whether to create the directories or, e.g, for testing, just create the list structure but not the directories}
}
\value{
the input bdpg_dir_names list with the new reserve selection run subtree added
}
\description{
This function is meant to deal with the fact that for a given
 reserve selection problem (whether correct or apparent, wrapped
 or simple, etc.), we may want to run a given reserve selector
 multiple times, e.g., with a different set of parameters each time.
}
\details{
A good example of when to use this function would be running marxan on the
 same problem with different numbers of iterations allowed, to get some idea
 of run time or problem difficulty.  For example, you might want to run it
 the first time only allowing 1000 iterations to get an idea of how long that
 takes and how good is the solution. You might also want to run it allowing 1
 million iterations to get a better solution when you have more run time
 available. Both of these runs might be done inside the same tzar run or they
 might be done in separate tzar runs.

In the case of doing both marxan runs inside the same tzar run, you'd like
 to have the directory structure under res_sel/marxan allow for marxan.1 and
 marxan.2, etc.  One way to do this is to have the code check for the largest
 marxan directory number in the current tzar run whenever it calls the marxan
 code and then create a new subdirectory tree with a higher number for the
 new run.

(Note that you don't have to be running tzar here.  The phrase "tzar run" is
 just a convenient term for the whole big run as opposed to just one call to
 marxan inside the whole big run.)

Creates a dir tree for this reserve selector parameterization and run. For
 example, if the reserve selector is marxan and this is the first
 parameterization and run of marxan for this problem, create
 res_sel/marxan/marxan.1 and its 2 subdirs, input and output. If this is the
 second parameterization and run for marxan, then create the same structure
 using marxan.2 instead of marxan.1.  Similarly, if you're using a different
 reserve selector, e.g., simpleRichness instead of marxan, this routine will
 create an analogous subtree res_sel/simpleRichness/simpleRichness.1, etc.
}

