% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSnet.R
\name{create_RSnet}
\alias{create_RSnet}
\title{Create an RSnet}
\usage{
create_RSnet(prob_UUID, parameters, cor_or_app_str,
  basic_or_wrapped_or_comb_str, net_method_name)
}
\arguments{
\item{prob_UUID}{UUID for the biodiversity problem the network metrics are
run over}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{cor_or_app_str}{character string}

\item{basic_or_wrapped_or_comb_str}{character string}
}
\value{
Returns an RSnet object
}
\description{
Create a run of a reserve selector
}
