% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_rs_analysis_and_output.R
\name{create_RSrun}
\alias{create_RSrun}
\title{Create an RSrun}
\usage{
create_RSrun(prob_UUID, targets, parameters, cor_or_app_str,
  basic_or_wrapped_or_comb_str, rs_method_name)
}
\arguments{
\item{prob_UUID}{UUID for the biodiversity problem the reserve selector is
run over}

\item{targets}{numeric vector}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{cor_or_app_str}{character string}

\item{basic_or_wrapped_or_comb_str}{character string}

\item{rs_method_name}{character string for reserve selection method, e.g.,
"marxan_sa"}
}
\value{
Returns an RSrun object
}
\description{
Create a run of a reserve selector
}
