% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{create_wrapping_PU_spp_table}
\alias{create_wrapping_PU_spp_table}
\title{Create wrapping PU/species table}
\usage{
create_wrapping_PU_spp_table(extra_abund, dep_set, eligible_set,
  use_testing_only_rand_seed = FALSE, testing_only_rand_seed = 17)
}
\arguments{
\item{extra_abund}{vector}

\item{dep_set}{vector}

\item{eligible_set}{vector}

\item{use_testing_only_rand_seed}{boolean}

\item{testing_only_rand_seed}{integer}
}
\value{
PU_spp_table
}
\description{
Create wrapping PU/species table
}
