% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_9a_create_Xu_graph.R
\name{create_Xu_graph}
\alias{create_Xu_graph}
\title{Create Xu graph}
\usage{
create_Xu_graph(num_nodes_per_group, n__num_groups, nodes,
  max_possible_tot_num_links, target_num_links_between_2_groups_per_round,
  num_rounds_of_linking_between_groups, duplicate_links_allowed = FALSE,
  bdpg_error_codes)
}
\arguments{
\item{num_nodes_per_group}{integer number of planning units per group in the
Xu problem}

\item{n__num_groups}{integer number of groups in the Xu problem, one of the
4 base input parameters for specifying the Xu problem}

\item{nodes}{data frame containing integer node_ID, integer group_ID, and boolean dependent_set_member}

\item{max_possible_tot_num_links}{integer maximum possible number of links
in the Xu problem given all the other parameter settings}

\item{target_num_links_between_2_groups_per_round}{integer target number of
links between any two groups per round of linking in constructing the
Xu problem}

\item{num_rounds_of_linking_between_groups}{integer number of rounds of
linking to be done between groups in constructing the Xu problems}

\item{duplicate_links_allowed}{boolean indicating whether more than one link
is allowed between the same two planning units, i.e., whether more than
one species can occupy exactly the same pair of planning units; TRUE
implies more than one allowed, FALSE implies only one allowed}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns an edge list, which is a two column integer matrix of
    node IDs with one row for each edge and columns for the 2 ends of
    the edge
}
\description{
Given a set of controlling parameters derived from the initial 4 input
parameters, this function the creates an edge list for a Xu graph using the
algorithm described by Xu et al.  [Need citations here.]
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{bdpg_error_codes}{
\preformatted{
bdpg_error_codes : List of 6
 $ ERROR_STATUS_num_inside_or_within_group_links_less_than_one: num 1001
 $ ERROR_STATUS_optimal_solution_is_not_optimal               : num 1002
 $ ERROR_STATUS_num_nodes_per_group_must_be_at_least_2        : num 1003
 $ ERROR_STATUS_duplicate_spp_in_Xu_input_file                : num 1004
 $ ERROR_STATUS_unknown_spp_occ_FP_error_type                 : num 1005
 $ ERROR_STATUS_unknown_spp_occ_FN_error_type                 : num 1006
}}
\subsection{DEBUG_LEVEL}{
\preformatted{
DEBUG_LEVEL :  num 0
}}
\subsection{duplicate_links_allowed}{
\preformatted{
duplicate_links_allowed :  logi FALSE
}}
\subsection{edge_list}{
\preformatted{
edge_list :  int [1:814, 1:2] 1 3 5 7 9 11 13 15 17 19 ...
}}
\subsection{edge_list_and_cur_row}{
\preformatted{
edge_list_and_cur_row : List of 2
 $ edge_list: int [1:1003, 1:2] 1 3 5 7 9 11 13 15 17 19 ...
 $ cur_row  : num 62
}}
\subsection{first_row_of_intergroup_links}{
\preformatted{
first_row_of_intergroup_links :  num 62
}}
\subsection{max_possible_tot_num_links}{
\preformatted{
max_possible_tot_num_links :  num 1003
}}
\subsection{n__num_groups}{
\preformatted{
n__num_groups :  num 61
}}
\subsection{nodes}{
\preformatted{
nodes : 'data.frame':	122 obs. of  3 variables:
 $ node_ID             : int  1 2 3 4 5 6 7 8 9 10 ...
 $ group_ID            : num  1 1 2 2 3 3 4 4 5 5 ...
 $ dependent_set_member: logi  FALSE TRUE FALSE TRUE FALSE TRUE ...
}}
\subsection{num_nodes_per_group}{
\preformatted{
num_nodes_per_group :  num 2
}}
\subsection{num_non_unique_edge_list}{
\preformatted{
num_non_unique_edge_list :  int 1003
}}
\subsection{num_rounds_of_linking_between_groups}{
\preformatted{
num_rounds_of_linking_between_groups :  num 942
}}
\subsection{num_unique_edge_list}{
\preformatted{
num_unique_edge_list :  int 814
}}
\subsection{target_num_links_between_2_groups_per_round}{
\preformatted{
target_num_links_between_2_groups_per_round :  num 1
}}
}

