% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob.R
\name{create_Xu_problem_from_scratch_not_using_4_Xu_metaparams}
\alias{create_Xu_problem_from_scratch_not_using_4_Xu_metaparams}
\title{create_Xu_problem_from_scratch_not_using_4_Xu_metaparams}
\usage{
create_Xu_problem_from_scratch_not_using_4_Xu_metaparams(
  max_allowed_num_spp,
  duplicate_links_allowed,
  parameters,
  integerize
)
}
\arguments{
\item{max_allowed_num_spp}{integer}

\item{duplicate_links_allowed}{boolean (defaults to FALSE)}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}
}
\value{
Returns a PU_spp_pair_info_class object
}
\description{
Generate a Xu biodiversity problem NOT based on 4 input control parameters
rather than reading it from a file.
}
\seealso{
Other interfaces to creation of Xu problems: 
\code{\link{create_Xu_problem_from_scratch_given_4_Xu_metaparams}()},
\code{\link{create_Xu_problem_from_scratch_given_params}()},
\code{\link{create_Xu_problem_from_scratch}()}
}
\concept{interfaces to creation of Xu problems}
