% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{decode_block_start}
\alias{decode_block_start}
\title{Start new function block and clean up previous block if necessary}
\usage{
decode_block_start(prev_state, visual_separator = "\\n\\n\\n\\n")
}
\arguments{
\item{prev_state}{character string indicating the previous state, e.g.,
"state__block_start"}

\item{visual_separator}{character string to write between the end of the
previous block and the new block to make it easier to pick out where
one ends and the other begins}
}
\value{
character string indicating the next state to transition to, e.g.,
    "state__block_start"
}
\description{
At start of new block of variable descriptions for a new function,
write some output to visually separate the new output from the previous
block.  Also clean up any variable subsection that's still open from
the previous block, i.e., write closing brackets for it.
}
\seealso{
\code{\link{close_open_subsection_if_necessary}}
}
