% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_5_derive_control_parameters.R
\name{derive_Xu_control_parameters}
\alias{derive_Xu_control_parameters}
\title{Derive full Xu control params from 4 base params}
\usage{
derive_Xu_control_parameters(parameters, bdpg_error_codes, integerize)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{bdpg_error_codes}{list of error names and corresponding error codes}

\item{integerize}{function to use in converting floats to ints}
}
\value{
a Xu_parameters object
}
\description{
Derive full Xu control params from 4 base params
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{alpha__}{
\preformatted{
alpha__ :  num 0.169
}}
\subsection{at_least_1_for_target_num_links_between_2_groups_per_round}{
\preformatted{
at_least_1_for_target_num_links_between_2_groups_per_round :  logi FALSE
}}
\subsection{base_for_target_num_links_between_2_groups_per_round}{
\preformatted{
base_for_target_num_links_between_2_groups_per_round :  chr "num_nodes_per_group"
}}
\subsection{bdpg_error_codes}{
\preformatted{
bdpg_error_codes : List of 6
 $ ERROR_STATUS_num_inside_or_within_group_links_less_than_one: num 1001
 $ ERROR_STATUS_optimal_solution_is_not_optimal               : num 1002
 $ ERROR_STATUS_num_nodes_per_group_must_be_at_least_2        : num 1003
 $ ERROR_STATUS_duplicate_spp_in_Xu_input_file                : num 1004
 $ ERROR_STATUS_unknown_spp_occ_FP_error_type                 : num 1005
 $ ERROR_STATUS_unknown_spp_occ_FN_error_type                 : num 1006
}}
\subsection{desired_num_PUs}{
\preformatted{
desired_num_PUs :  num 122
}}
\subsection{integerize}{
\preformatted{
integerize : function (x, digits = 0)
}}
\subsection{max_possible_num_links_between_groups}{
\preformatted{
max_possible_num_links_between_groups :  num 942
}}
\subsection{max_possible_tot_num_links}{
\preformatted{
max_possible_tot_num_links :  num 1003
}}
\subsection{max_possible_tot_num_node_link_pairs}{
\preformatted{
max_possible_tot_num_node_link_pairs :  num 2006
}}
\subsection{n__num_groups}{
\preformatted{
n__num_groups :  num 61
}}
\subsection{num_dependent_set_nodes}{
\preformatted{
num_dependent_set_nodes :  num 61
}}
\subsection{num_independent_nodes_per_group}{
\preformatted{
num_independent_nodes_per_group :  num 1
}}
\subsection{num_independent_set_nodes}{
\preformatted{
num_independent_set_nodes :  num 61
}}
\subsection{num_links_within_one_group}{
\preformatted{
num_links_within_one_group :  num 1
}}
\subsection{num_nodes_per_group}{
\preformatted{
num_nodes_per_group :  num 2
}}
\subsection{num_rounds_of_linking_between_groups}{
\preformatted{
num_rounds_of_linking_between_groups :  num 942
}}
\subsection{opt_solution_as_frac_of_tot_num_nodes}{
\preformatted{
opt_solution_as_frac_of_tot_num_nodes :  num 0.5
}}
\subsection{p__prop_of_links_between_groups}{
\preformatted{
p__prop_of_links_between_groups :  num 0.462
}}
\subsection{parameters}{
\preformatted{
parameters : List of 66
 $ summary_without_run_id_filename                           : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/prob_diff_results_with_0_ru"| __truncated__
 ...
 $ fullOutputDir_NO_slash                                    : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{r__density}{
\preformatted{
r__density :  num 3.76
}}
\subsection{target_num_links_between_2_groups_per_round}{
\preformatted{
target_num_links_between_2_groups_per_round :  num 1
}}
\subsection{tot_num_links_inside_groups}{
\preformatted{
tot_num_links_inside_groups :  num 61
}}
\subsection{tot_num_nodes}{
\preformatted{
tot_num_nodes :  num 122
}}
\subsection{Xu_base_parameters}{
\preformatted{
Xu_base_parameters : Formal class 'Xu_base_params' [package "bdpg"] with 4 slots
}}
\subsection{Xu_bdpg_extended_parameters}{
\preformatted{
Xu_bdpg_extended_parameters : Formal class 'Xu_bdpg_extended_params' [package "bdpg"] with 18 slots
}}
\subsection{Xu_derived_parameters}{
\preformatted{
Xu_derived_parameters : Formal class 'Xu_derived_params' [package "bdpg"] with 13 slots
}}
\subsection{Xu_parameters}{
\preformatted{
Xu_parameters : Formal class 'Xu_params' [package "bdpg"] with 3 slots
}}
}

