% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_marxan_analysis_and_output.R
\name{do_APP_marxan_analysis_and_output}
\alias{do_APP_marxan_analysis_and_output}
\title{Run marxan on APP problem and write output from all analysis}
\usage{
do_APP_marxan_analysis_and_output(APP_bd_prob, COR_bd_prob, parameters,
  src_rds_file_dir = NULL, targets = rep(1, COR_bd_prob@num_spp))
}
\arguments{
\item{APP_bd_prob}{a Xu_bd_problem}

\item{COR_bd_prob}{a Xu_bd_problem}

\item{parameters}{list}

\item{src_rds_file_dir}{character string}

\item{targets}{numeric vector}
}
\value{
Returns nothing
}
\description{
Run marxan on APP problem and write output from all analysis
}
