% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_graph_and_marxan_analysis.R
\name{do_COR_marxan_analysis_and_output}
\alias{do_COR_marxan_analysis_and_output}
\title{Run marxan on COR problem and write output from all analysis}
\usage{
do_COR_marxan_analysis_and_output(COR_bd_prob, parameters)
}
\arguments{
\item{parameters}{}
}
\description{
Run marxan on COR problem and write output from all analysis
}

