% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_simple_richness_control_file_and_run_SR.R
\name{do_COR_simple_richness_analysis_and_output}
\alias{do_COR_simple_richness_analysis_and_output}
\title{Run simple_richness on COR problem and write output from all analysis}
\usage{
do_COR_simple_richness_analysis_and_output(
  COR_bd_prob,
  parameters,
  rs_method_name,
  forward = TRUE,
  src_rds_file_dir = NULL,
  spp_rep_targets = rep(1, COR_bd_prob@num_spp)
)
}
\arguments{
\item{COR_bd_prob}{a correct Xu_bd_problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{rs_method_name}{character string for reserve selection method, e.g.,
"marxan_sa"}

\item{forward}{boolean indicating whether to do the greedy search in the
forward direction (adding PUs to the solution set) or the backward
direction (removing PUs from the solution set); TRUE implies forward}

\item{src_rds_file_dir}{string containing full path of directory where
source rds file is or was found}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}
}
\value{
Returns nothing
}
\description{
Run simple_richness on COR problem and write output from all analysis
}
