% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_Greedy_analysis_and_output.R, R/rs.R
\name{do_Greedy_analysis_and_output}
\alias{do_Greedy_analysis_and_output}
\alias{do_Greedy_analysis_and_output}
\title{Run greedy reserve selector on bd problem and write output from all analysis}
\usage{
do_Greedy_analysis_and_output(APP_bd_prob, COR_bd_prob, parameters,
  rs_method_name, resSel_func, RS_specific_params, src_rds_file_dir = NULL,
  spp_rep_targets = rep(1, COR_bd_prob@num_spp))

do_Greedy_analysis_and_output(APP_bd_prob, COR_bd_prob, parameters,
  rs_method_name, resSel_func, RS_specific_params, src_rds_file_dir = NULL,
  spp_rep_targets = rep(1, COR_bd_prob@num_spp))
}
\arguments{
\item{APP_bd_prob}{an apparent Xu_bd_problem}

\item{COR_bd_prob}{a correct Xu_bd_problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{rs_method_name}{character string for reserve selection method, e.g.,
"marxan_sa"}

\item{resSel_func}{greedy reserve selection function, e.g.,
simpleRichness}

\item{RS_specific_params}{list of parameters specific to the given
reserve selector}

\item{src_rds_file_dir}{string containing full path of directory where
source rds file is or was found}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}
}
\value{
Returns nothing

Returns nothing
}
\description{
Note that the COR and APP arguments are the same if running on a correct
problem.

Note that the COR and APP arguments are the same if running on a correct
problem.
}
