% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_marxan_analysis_and_output.R
\name{do_marxan_analysis_and_output}
\alias{do_marxan_analysis_and_output}
\title{Run marxan on bd problem and write output from all analysis}
\usage{
do_marxan_analysis_and_output(
  APP_bd_prob,
  COR_bd_prob,
  parameters,
  starting_dir,
  rs_method_name,
  src_rds_file_dir = NULL,
  spp_rep_targets = rep(1, COR_bd_prob@num_spp)
)
}
\arguments{
\item{APP_bd_prob}{an apparent Xu_bd_problem}

\item{COR_bd_prob}{a correct Xu_bd_problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{starting_dir}{string containing full path of top directory where
problems and/or runs are to be created}

\item{rs_method_name}{character string for reserve selection method, e.g.,
"marxan_sa"}

\item{src_rds_file_dir}{string containing full path of directory where
source rds file is or was found}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}
}
\value{
Returns nothing
}
\description{
Note that the COR and APP arguments are the same if running on a correct
problem.
}
