% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_rs_analysis_and_output.R
\name{do_rs_analysis_and_output}
\alias{do_rs_analysis_and_output}
\title{Run reserve selector(s) on bdproblem and write output from all analyses}
\usage{
do_rs_analysis_and_output(APP_bd_prob, COR_bd_prob, parameters,
  src_rds_file_dir = NULL, spp_rep_targets = rep(1,
  COR_bd_prob@num_spp))
}
\arguments{
\item{APP_bd_prob}{an apparent Xu_bd_problem}

\item{COR_bd_prob}{a correct Xu_bd_problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{src_rds_file_dir}{string containing full path of directory where
source rds file is or was found}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}
}
\value{
Returns nothing
}
\description{
Note that the COR and APP arguments are the same if running on a correct
problem.
}
