% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{doc_vars_in_this_func}
\alias{doc_vars_in_this_func}
\title{Document variables and their types that are visibile inside current function}
\usage{
doc_vars_in_this_func()
}
\value{
Nothing
}
\description{
Document variables and their types that are visibile inside current function
}
