% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_vars_in_this_func.R
\name{doc_vars_in_this_func_once}
\alias{doc_vars_in_this_func_once}
\title{Document function's variables a limited number of times}
\usage{
doc_vars_in_this_func_once(max_ct = 1,
  ctr_name_prefix = "TEMPCTR___RM_THIS_AT_END__")
}
\arguments{
\item{max_ct}{integer maximum number of times to write the documentation
for a function, no matter how many times it's called}

\item{ctr_name_prefix}{character string to prepend to function name when
building a counter name that is supposed to be unique}
}
\value{
Returns nothing
}
\description{
This is a wrapper function for \code{\link{doc_vars_in_this_func}}.
It makes sure that \code{doc_vars_in_this_func} is only run once
in a given invocation of a program, no matter how many times the function
containing the \code{doc_vars_in_this_func} call is called.  This is
useful in avoiding getting a bunch of duplicate outputs for the same
function when it is called more than once in a program.
}
\seealso{
\code{\link{doc_vars_in_this_func}} for the main action of this
    routine

\code{\link{doc_vars_in_this_func_once}} for explanation about
    ctr_name_prefix, which you will probably never touch
}
