% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lognormal_overlay.R
\name{EF}
\alias{EF}
\title{Evaluation Function to be passed to optim()}
\usage{
EF(seed_value, num_spp_to_generate, min_num_spp_on_2_PUs, max_max_abundance,
  target_max_abundance_ct, add_one_to_abundances, outfile, mean_sd_pair)
}
\arguments{
\item{seed_value}{integer}

\item{num_spp_to_generate}{integer}

\item{min_num_spp_on_2_PUs}{integer}

\item{max_max_abundance}{integer}

\item{target_max_abundance_ct}{integer}

\item{add_one_to_abundances}{boolean}

\item{outfile}{character string}

\item{mean_sd_pair}{2 element numeric vector containing the mean and
standard deviation to use for the lognormal}
}
\value{
Returns numeric score from evaluating the given lognormal; the
    closer to 0 the better
}
\description{
Note that the "Local Variable Structures and examples" section below was
not labelled as belonging to a function "EF".  It was labelled as belonging
to a function "fn".  I'm assuming that the variables actually belong to EF
because "fn" is the argument name that EF gets assigned to when passed to
optim().
}
