% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14a_compute_marxan_solution_scores.R
\name{find_best_marxan_solutions}
\alias{find_best_marxan_solutions}
\title{Title}
\usage{
find_best_marxan_solutions(marxan_output_dir_path, num_spp, cor_PU_costs,
  cor_bpm, app_bpm, marxan_best_df_sorted_as_vector, plot_output_dir,
  largest_PU_ID, largest_spp_ID, targets)
}
\arguments{
\item{targets}{}
}
\value{
nothing?
}
\description{
Title
}
