% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lognormal_overlay.R
\name{find_lognormal_to_wrap_around_Xu}
\alias{find_lognormal_to_wrap_around_Xu}
\title{Find lognormal to wrap around Xu problem}
\usage{
find_lognormal_to_wrap_around_Xu(Xu_bdprob, parameters,
  desired_Xu_spp_frac_of_all_spp, solution_frac_of_landscape,
  desired_max_abundance_frac, seed_value_for_search, max_search_iterations,
  add_one_to_lognormal_abundances, search_outfile_name)
}
\arguments{
\item{Xu_bdprob}{a Xu_bd_problem object}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{desired_Xu_spp_frac_of_all_spp}{numeric}

\item{solution_frac_of_landscape}{numeric}

\item{desired_max_abundance_frac}{numeric}

\item{seed_value_for_search}{integer}

\item{max_search_iterations}{integer}

\item{add_one_to_lognormal_abundances}{boolean}

\item{search_outfile_name}{character string}
}
\value{
Returns numeric vector of abundances
}
\description{
Uses optimizer to search for parameters of a lognormal distribution that
will have at least as many species occurring on exactly 2 patches as the
given Xu problem and meet the other curve shape criteria given in the
argument list.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{add_one_to_lognormal_abundances}{
\preformatted{
add_one_to_lognormal_abundances :  logi FALSE
}}
\subsection{desired_max_abundance_frac}{
\preformatted{
desired_max_abundance_frac :  num 0.7
}}
\subsection{desired_Xu_spp_frac_of_all_spp}{
\preformatted{
desired_Xu_spp_frac_of_all_spp :  num 0.5
}}
\subsection{initial_meanlog_for_search}{
\preformatted{
initial_meanlog_for_search :  num 1.04
}}
\subsection{initial_sdlog_for_search}{
\preformatted{
initial_sdlog_for_search :  num 0.1
}}
\subsection{lognormal_search_results}{
\preformatted{
lognormal_search_results : List of 4
 $ meanlog       : num 0.676
 $ sdlog         : num 0.347
 $ score         : num 0.951
 $ abundance_data:List of 3
}}
\subsection{max_abundance_frac}{
\preformatted{
max_abundance_frac :  num 0.7
}}
\subsection{max_max_abundance_ct}{
\preformatted{
max_max_abundance_ct :  num 203
}}
\subsection{max_search_iterations}{
\preformatted{
max_search_iterations :  num 500
}}
\subsection{min_num_spp_on_2_PUs}{
\preformatted{
min_num_spp_on_2_PUs :  int 814
}}
\subsection{mu}{
\preformatted{
mu :  num 1.04
}}
\subsection{num_PUs_per_spp_ie_rarity}{
\preformatted{
num_PUs_per_spp_ie_rarity :  num 2
}}
\subsection{num_spp_to_generate}{
\preformatted{
num_spp_to_generate :  num 1628
}}
\subsection{num_spp_with_given_num_PUs}{
\preformatted{
num_spp_with_given_num_PUs :  int 814
}}
\subsection{num_Xu_dep_set_nodes}{
\preformatted{
num_Xu_dep_set_nodes :  int 61
}}
\subsection{parameters}{
\preformatted{
parameters : List of 66
 $ summary_without_run_id_filename                           : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/prob_diff_results_with_0_ru"| __truncated__
 ...
 $ fullOutputDir_NO_slash                                    : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{plusOrMinus1}{
\preformatted{
plusOrMinus1 :  num 1
}}
\subsection{rounded_abundances}{
\preformatted{
rounded_abundances :  num [1:1628] 2 2 1 1 3 1 3 2 2 1 ...
}}
\subsection{search_outfile_name}{
\preformatted{
search_outfile_name :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/wrap_search_outfile.csv"
}}
\subsection{seed_value_for_search}{
\preformatted{
seed_value_for_search :  num 11
}}
\subsection{sigma}{
\preformatted{
sigma :  num 0.1
}}
\subsection{solution_frac_of_landscape}{
\preformatted{
solution_frac_of_landscape :  num 0.3
}}
\subsection{target_max_abundance_ct}{
\preformatted{
target_max_abundance_ct :  num 142
}}
\subsection{tot_num_PUs_in_landscape}{
\preformatted{
tot_num_PUs_in_landscape :  num 203
}}
\subsection{Xu_bdprob}{
\preformatted{
Xu_bdprob : Formal class 'Xu_bd_problem' [package "bdpg"] with 35 slots
}}
\subsection{Xu_dep_set}{
\preformatted{
Xu_dep_set :  int [1:61] 2 4 6 8 10 12 14 16 18 20 ...
}}
\subsection{Xu_nodes}{
\preformatted{
Xu_nodes : 'data.frame':	122 obs. of  3 variables:
 $ node_ID             : int  1 2 3 4 5 6 7 8 9 10 ...
 $ group_ID            : num  1 1 2 2 3 3 4 4 5 5 ...
 $ dependent_set_member: logi  FALSE TRUE FALSE TRUE FALSE TRUE ...
}}
\subsection{Xu_tot_num_PUs}{
\preformatted{
Xu_tot_num_PUs :  int 122
}}
\subsection{Xu_tot_num_spp}{
\preformatted{
Xu_tot_num_spp :  int 814
}}
}

