% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{finish_up}
\alias{finish_up}
\title{Clean up any subsection left open when end of file is found}
\usage{
finish_up(prev_state)
}
\arguments{
\item{prev_state}{character string indicating the previous state, e.g.,
"state__block_start"}
}
\value{
Returns nothing
}
\description{
Clean up any subsection left open when end of file is found
}
\seealso{
\code{\link{close_open_subsection_if_necessary}}
}
