% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{fix_0_length_list_elements}
\alias{fix_0_length_list_elements}
\title{Return NA for any given value whose length is 0}
\usage{
fix_0_length_list_elements(value)
}
\arguments{
\item{a_list}{a list}
}
\value{
NA or the input argument
}
\description{
This is originally built to be used on each element of a list inside an
lapply where you're trying to get rid of NULLs, numeric(0)s, etc. in a
list to be converted to a data frame.
}
