% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_20_basic_variants.R
\name{gen_20_basic_variants_including_cost_error}
\alias{gen_20_basic_variants_including_cost_error}
\title{Generate 20 basic variants of a problem}
\usage{
gen_20_basic_variants_including_cost_error(parameters, integerize,
  err_amt = 0.5)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}

\item{err_amt}{numeric amount of error to add in creating apparent problems}
}
\value{
Returns nothing.
}
\description{
Create 1 correct problem and 1 wrap of that correct problem, then from each
of those, create 4 apparent variants using the following errors: FP only,
FN only, FP and FN, and finally, FP and FN with matching counts.  Based on
each of these 10 variants (2 correct and 4 apparent for each of the 2), then
create another apparent version of each problem that includes cost error too.
This gives a total of 20 problems.
}
