% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_4_basic_variants.R
\name{gen_4_basic_variants}
\alias{gen_4_basic_variants}
\title{Create 1 of each of the 4 types of problems and run reserve selector on each}
\usage{
gen_4_basic_variants(parameters, bdpg_error_codes, integerize)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{bdpg_error_codes}{list of error names and corresponding error codes}

\item{integerize}{function to use in converting floats to ints}
}
\value{
Returns nothing.
}
\description{
This is what used to be the mainline for bdpgxupaper.  It creates a correct
biodiversity problem, then from that correct, it creates a wrapped problem.
It also creates an apparent problem for the original correct problem
and another for the wrapped problem.  It also runs marxan on each of these
four problems.
}
