% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_20_basic_variants.R
\name{gen_4_simple_random_variants_mixing_FP_FN_and_cost_error}
\alias{gen_4_simple_random_variants_mixing_FP_FN_and_cost_error}
\title{Generate 4 simple random variants of a problem}
\usage{
gen_4_simple_random_variants_mixing_FP_FN_and_cost_error(
  parameters,
  integerize
)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}

\item{err_amt}{numeric amount of error to add in creating apparent problems}
}
\value{
Returns nothing.
}
\description{
Create 1 correct problem and 1 wrap of that correct problem, then from the
Wrap problem, create 1 apparent variant using FP and FN error but no cost
error.  Then create another apparent using FP and FN, this time WITH cost
error.  This gives a total of 4 problems.
}
