% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_6_basic_variants.R
\name{gen_6_basic_variants}
\alias{gen_6_basic_variants}
\title{Generate 6 basic variants of a problem}
\usage{
gen_6_basic_variants(parameters, integerize, err_amt = NA)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}

\item{err_amt}{numeric amount of error to add in creating apparent problems}
}
\value{
Returns nothing.
}
\description{
Create 1 correct problem and 1 wrap of that correct problem, then from the
wrap problem, create 4 apparent variants using the following errors: FP only,
FN only, FP and FN, and finally, FP and FN with matching counts.
}
