% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bdprob.R
\name{gen_bdprob}
\alias{gen_bdprob}
\title{Generate a biodiversity problem}
\usage{
gen_bdprob(parameters, integerize, base_bdprob = NULL)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}

\item{base_bdprob}{NULL or an existing bdprob to use in creating multi_bdproblem}
}
\value{
biodiversity problems
}
\description{
Generate a biodiversity problem.
}
