% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{gen_new_seed}
\alias{gen_new_seed}
\title{Get and set a new seed for the random number generator}
\usage{
gen_new_seed()
}
\arguments{
\item{location_string}{a string to print indicating where this function is
called}
}
\value{
Returns an integer
}
\description{
Get and set a new seed for the random number generator
}
