% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lognormal_overlay.R
\name{gen_rounded_abundances}
\alias{gen_rounded_abundances}
\title{Generate rounded abundances}
\usage{
gen_rounded_abundances(
  num_spp_to_generate,
  meanlog,
  sdlog,
  add_one_to_abundances,
  plot_rounded_abundances = FALSE
)
}
\arguments{
\item{num_spp_to_generate}{integer}

\item{meanlog}{numeric}

\item{sdlog}{numeric}

\item{add_one_to_abundances}{boolean}

\item{plot_rounded_abundances}{boolean}
}
\value{
Returns a vector with one count for each species
}
\description{
Function to generate a sampled lognormal distribution whose values are
 rounded to produce a vector of counts of the number of PUs that each
 species will occupy.
}
\details{
No zero counts are allowed, so any species that was assigned a 0 count in
 the lognormal generation is removed from the returned vector.
 This means that the number of species returned may be less than the value
 of num_spp_to_generate specified in the inputs.
}
