% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSetCoverProblem.R
\name{gen_set_cover_problem}
\alias{gen_set_cover_problem}
\title{Base function for generating set cover problems}
\usage{
gen_set_cover_problem(parameters)
}
\arguments{
\item{parameters}{List of parameters controlling the current run (usually
decoded from project.yaml by tzar)}
}
\value{
returns nothing
}
\description{
The main function that is to be called when generating a biodiversity
problem generator.
}
\examples{
\dontrun{
gen_set_cover_problem (parameters)
}
}

