% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob.R
\name{gen_single_bdprob}
\alias{gen_single_bdprob}
\title{Generate a single biodiversity problem}
\usage{
gen_single_bdprob(read_Xu_problem_from_Xu_file, infile_name,
  given_correct_solution_cost, max_allowed_num_spp, bdpg_error_codes,
  integerize)
}
\arguments{
\item{read_Xu_problem_from_Xu_file}{boolean TRUE if problem is to be
read from a file created by Xu; FALSE if problem is to be created from
scratch.}

\item{infile_name}{DESCRIPTION.}

\item{given_correct_solution_cost}{DESCRIPTION.}

\item{max_allowed_num_spp}{DESCRIPTION.}

\item{bdpg_error_codes}{DESCRIPTION.}

\item{integerize}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Read a Xu problem from files of ones already created by Xu or
create one from scratch.
}
\examples{
# ADD_EXAMPLES_HERE
}

