% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob_APP.R
\name{gen_single_bdprob_APP}
\alias{gen_single_bdprob_APP}
\title{Generate a single biodiversity problem with error added to it}
\usage{
gen_single_bdprob_APP(Xu_bdprob_COR, parameters, bdpg_error_codes)
}
\arguments{
\item{Xu_bdprob_COR}{correct Xu problem that is to have error added to it}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns apparent version of either a Xu_bd_problem or a
    Xu_wrapped_bd_problem
}
\description{
All values that are altered for the apparent problem should be assumed to
 be altered in the APP_prob_info_class object in the problem's APP_prob_info
 slot.
 Values in the main part of the Xu_bdprob_APP object should be the same
 as the values for the underlying COR problem that the APP problem is
 adding error to, with the exception of a few object-specific things
 like the UUID of the APP problem itself.
}
