% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob.R
\name{gen_single_bdprob_COR}
\alias{gen_single_bdprob_COR}
\title{Generate a single correct biodiversity problem from scratch or from a file}
\usage{
gen_single_bdprob_COR(parameters, integerize,
  base_prob_name_stem = "base_prob", cor_dir_name_stem = "cor")
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}

\item{base_prob_name_stem}{a character string}

\item{cor_dir_name_stem}{a character string}
}
\value{
Returns a correct Xu biodiversity problem
}
\description{
Generate a single correct biodiversity problem from scratch or from a file
}
