% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{gen_single_bdprob_WRAP}
\alias{gen_single_bdprob_WRAP}
\title{Generate a single wrapped biodiversity problem from a given base problem}
\usage{
gen_single_bdprob_WRAP(bdprob_to_wrap, parameters)
}
\arguments{
\item{bdprob_to_wrap}{a correct Xu_bd_problem whose correct solution vector is
known}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}
}
\value{
Returns a correct wrapped biodiversity problem
}
\description{
Generate a single wrapped biodiversity problem from a given base problem
}
