% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{gen_wrapped_bdprob_COR}
\alias{gen_wrapped_bdprob_COR}
\title{Generate a correct wrapped Xu biodiversity problem from a correct base problem}
\usage{
gen_wrapped_bdprob_COR(starting_dir, compute_network_metrics_for_this_prob,
  parameters, base_bdprob, bdpg_error_codes)
}
\arguments{
\item{starting_dir}{character string}

\item{compute_network_metrics_for_this_prob}{boolean indicating whether to
compute network metrics for the current problem; TRUE indicates to
compute one or more network metrics specified elsewhere and FALSE
indicate not compute any network metrics for the current problem.}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{base_bdprob}{a correct Xu_bd_problem whose correct solution vector is
known}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns a correct wrapped biodiversity problem
}
\description{
Wrap a given distribution around a given Xu problem's distribution.
That is, add more planning units and species to the flat Xu distribution's
set of PUs and species so that the Xu distribution is a proper subset of
the larger distribution.
}
\details{
This is intended as a way of embedding the Xu problem's unrealistic
species distribution (i.e., every species occurs on exactly 2 patches)
inside a more realistic distribution, but one that has exactly the same
correct solution set as the base Xu problem.
}
\section{Restrictions on wrapping distribution}{

At the moment, the only wrapping distribution that there is code for
generating is the lognormal distribution.  However, the basic idea allows
for ANY distribution where the base set of Xu species and planning units
is a proper subset of the final distribution.

To enhance the capabilities of this routine to allow other distributions,
you would just have to

\itemize{
   \item{provide some kind of option(s) to choose what kind
of wrapping distribution to use}
   \item{provide a function to generate that wrapping distribution}
   \item{replace the call to find_lognormal_to_wrap_around_Xu() with
   the new function}
}

One caveat is that all species that occur on one and only one planning unit
are removed from the distribution.  This is because those species would
automatically require their planning unit to be included in the final
solution and therefore, make the problem simpler for the optimizer.
}

\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{add_one_to_lognormal_abundances}{
\preformatted{
add_one_to_lognormal_abundances :  logi FALSE
}}
\subsection{base_bdprob}{
\preformatted{
base_bdprob : Formal class 'Xu_bd_problem' [package "bdpg"] with 35 slots
}}
\subsection{bdpg_error_codes}{
\preformatted{
bdpg_error_codes : List of 6
 $ ERROR_STATUS_num_inside_or_within_group_links_less_than_one: num 1001
 $ ERROR_STATUS_optimal_solution_is_not_optimal               : num 1002
 $ ERROR_STATUS_num_nodes_per_group_must_be_at_least_2        : num 1003
 $ ERROR_STATUS_duplicate_spp_in_Xu_input_file                : num 1004
 $ ERROR_STATUS_unknown_spp_occ_FP_error_type                 : num 1005
 $ ERROR_STATUS_unknown_spp_occ_FN_error_type                 : num 1006
}}
\subsection{dep_set_PUs_eligible}{
\preformatted{
dep_set_PUs_eligible :  logi FALSE
}}
\subsection{desired_max_abundance_frac}{
\preformatted{
desired_max_abundance_frac :  num 0.7
}}
\subsection{desired_Xu_spp_frac_of_all_spp}{
\preformatted{
desired_Xu_spp_frac_of_all_spp :  num 0.5
}}
\subsection{max_search_iterations}{
\preformatted{
max_search_iterations :  num 500
}}
\subsection{parameters}{
\preformatted{
parameters : List of 66
 $ summary_without_run_id_filename                           : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/prob_diff_results_with_0_ru"| __truncated__
 ...
 $ fullOutputDir_NO_slash                                    : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{rounded_abundances}{
\preformatted{
rounded_abundances :  num [1:1628] 2 2 1 1 3 1 3 2 2 1 ...
}}
\subsection{search_outfile_name}{
\preformatted{
search_outfile_name :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/wrap_search_outfile.csv"
}}
\subsection{seed_value_for_search}{
\preformatted{
seed_value_for_search :  num 11
}}
\subsection{solution_frac_of_landscape}{
\preformatted{
solution_frac_of_landscape :  num 0.3
}}
\subsection{starting_dir}{
\preformatted{
starting_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{tot_num_PUs_in_landscape}{
\preformatted{
tot_num_PUs_in_landscape :  num 407
}}
\subsection{wrapped_bdprob_COR}{
\preformatted{
wrapped_bdprob_COR : Formal class 'Xu_wrapped_bd_problem' [package "bdpg"] with 36 slots
}}
}

