% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{generate_func_var_roxygen_comments}
\alias{generate_func_var_roxygen_comments}
\title{Title}
\usage{
generate_func_var_roxygen_comments(infile, sinkFilePath)
}
\description{
Title
}
