% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{get_and_set_new_rand_seed}
\alias{get_and_set_new_rand_seed}
\title{Get and set a new seed for the random number generator}
\usage{
get_and_set_new_rand_seed(location_string, forced_seed = NULL)
}
\arguments{
\item{location_string}{a string to print indicating where this function is
called}

\item{forced_seed}{an integer value to be used as random number seed when
a specific seed is desired in creating this object (e.g., for
reproducing a previous execution of the code).}
}
\value{
Returns an integer
}
\description{
Get and set a new seed for the random number generator
}
