% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_action_using_tzar_reps.R
\name{get_base_cor_bdprob_for_given_app_bdprob}
\alias{get_base_cor_bdprob_for_given_app_bdprob}
\title{Get the COR bdprob whose UUID is the base UUID for the given APP bdprob}
\usage{
get_base_cor_bdprob_for_given_app_bdprob(app_bdprob, parameters)
}
\arguments{
\item{app_bdprob}{an apparent Xu_bd_problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}
}
\value{
Returns the COR Xu_bd_problem that the apparent problem added error
to
}
\description{
Get the COR bdprob whose UUID is the base UUID for the given APP bdprob
}
