% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_initialization.R
\name{get_bdpg_error_code}
\alias{get_bdpg_error_code}
\title{Look up a specific bdpg error code given its name}
\usage{
get_bdpg_error_code(bdpg_error_name)
}
\arguments{
\item{bdpg_error_name}{string containing name of error code to look up in list of error codes}
}
\value{
integer error code for the given error name
}
\description{
Look up a specific bdpg error code given its name
}
\examples{
get_bdpg_error_code ("ERROR_STATUS_unknown_spp_occ_FP_error_type")

}

