% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_initialization.R
\name{get_bdpg_error_codes}
\alias{get_bdpg_error_codes}
\title{Builds the error codes returned by errors in bdpg package}
\usage{
get_bdpg_error_codes()
}
\value{
list of integer error codes
}
\description{
Builds the error codes returned by errors in bdpg package
}
\examples{
get_bdpg_error_codes ()
}

