% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_action_using_tzar_reps.R
\name{get_bdprob_from_rds_file}
\alias{get_bdprob_from_rds_file}
\title{Load a bdprob from an R rds output file, possibly specified in an array}
\usage{
get_bdprob_from_rds_file(prob_src, cur_input_prob_idx, rds_file_set_path,
  rds_file_set_yaml_array, rds_file_path, bdpg_error_codes)
}
\arguments{
\item{prob_src}{character string}

\item{cur_input_prob_idx}{integer}

\item{rds_file_set_path}{character string}

\item{rds_file_set_yaml_array}{vector of character strings}

\item{rds_file_path}{character string}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns a list containing named elements related to the source
        problem, i.e. src_Xu_bd_problem and src_rds_file_dir
}
\description{
Takes a file name for an R rds output file that contains a saved Xu_bd_problem
object and loads the bdprob from the file.  The location of the file comes
from the parameters list and can be given in several ways.  The source
can be an rds file path or a text file containing multiple rds file paths or
it can be a yaml array containing multiple rds file paths.  If the source is
either of the multiple path sources, then the parameters list will contain
a variable called cur_input_prob_idx, giving the location of the path in
the text file or yaml array.
}
