% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_initialization.R
\name{get_current_os}
\alias{get_current_os}
\title{Get name and version number of current operating system as a string}
\usage{
get_current_os()
}
\value{
Returns string containing name and version of current operating system
}
\description{
This function is only here as a convenience because I can never remember
where the current operating system is stored.
}
\examples{
get_current_os ()
}
