% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{get_first_token_type_of_line}
\alias{get_first_token_type_of_line}
\title{Get type of first token on the line}
\usage{
get_first_token_type_of_line(data_line)
}
\arguments{
\item{data_line}{character string containing one full line from the input
file}
}
\value{
character string containing the first token type of the line,
    e.g., "whitespace" or "symbol"
}
\description{
Tokenizes the whole line and then returns the type associated with the
first token on the line.  This is useful because the first token is what
flags whether the line is the start of a chunk, variable, or function or
whether it's a continuation line of a function or a variable.
}
\details{
This doesn't really need to tokenize the whole line since it only uses the
first token.  Not sure if there's a way to ask for just one token.
}
