% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_seed_setting_utilities.R
\name{get_forced_seed_value_if_necessary}
\alias{get_forced_seed_value_if_necessary}
\title{Get forced seed value if necessary}
\usage{
get_forced_seed_value_if_necessary(
  is_rsrun,
  is_rsprob,
  parameters,
  cor_or_app_str,
  basic_or_wrapped_or_comb_str
)
}
\arguments{
\item{is_rsrun}{boolean indicating that the seed creation test is being
done just before the creation of an rs RUN object}

\item{is_rsprob}{boolean indicating that the seed creation test is being
done just before the creation of an rs PROBLEM object}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{cor_or_app_str}{character string containing "COR" or "APP"}

\item{basic_or_wrapped_or_comb_str}{character string containing "BASE" or
"WRAP" or "COMB"}
}
\value{
Returns an integer seed value or NA
}
\description{
Determines whether a seed value is given or needs to be set or not
generated at all based on input options in the parameters list.
}
\details{
2017 12 06 - BTL
 NOTE:  basic_or_wrapped_or_comb_str = "COMB" does not exist at the moment,
 but it's mentioned in BDProb.R, so it's covered here for
 future compatibility.
}
