% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_initialization.R
\name{get_integerize_function}
\alias{get_integerize_function}
\title{Get function to use when converting numeric values to integers}
\usage{
get_integerize_function(integerize_string)
}
\arguments{
\item{integerize_string}{string containing name of the function to use to
convert floats to integers}
}
\value{
Returns function to use to convert floats to integers
}
\description{
Many computations return floating point numbers that need to be passed
to other routines as integers.  The method used to convert them to integers
may affect the outcome of the downstream operation, so this function
looks up which function to use based on the user's specification in the
parameter inputs.
}
\examples{
get_integerize_function ("ceiling")

}
