% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{get_next_numbered_dir_name}
\alias{get_next_numbered_dir_name}
\title{Build a new numbered file name whose extension is 1 higher than the largest
extension already in the given directory.}
\usage{
get_next_numbered_dir_name(dir_to_search, name_stem)
}
\arguments{
\item{dir_to_search}{character string giving the path to the directory to
search for numbered directory names}

\item{name_stem}{character string giving just the stem of the subdirectory
names, e.g,. "app" or "marxan", etc.}
}
\value{
character string giving the name of the next highest number
    directory, e.g., if the dir_to_search was "tzarout/exp1/base_prob/app"
    and the name_stem was "app" and there were directories
    "tzarout/exp1/base_prob/app/app.1" and "tzarout/exp1/base_prob/app/app.2"
    then "app.3" would be returned.
}
\description{
If there are no numbered directories (e.g., app.1, app.2, etc.) in the
directory to be searched, then this routine will return the stem with ".1"
appended, e.g., "marxan.1".
}
\details{
This routine is a little bit loose in that it doesn't check for errors
such as whether there are any non integer file extensions.
}

