% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_initialization.R
\name{init_for_bdpg}
\alias{init_for_bdpg}
\title{Initialize for biodiversity problem generation}
\usage{
init_for_bdpg(parameters)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}
}
\value{
Returns two element list containing parameters list and bdpg error codes
}
\description{
Central point for doing a few odds and ends of initializing.
}
