% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_initialization.R
\name{initialize_and_derive_parameters}
\alias{initialize_and_derive_parameters}
\title{Title}
\usage{
initialize_and_derive_parameters(parameters)
}
\arguments{
\item{parameters}{list of parameters (normally read from project.yaml)}
}
\value{
derived_bdpg_parameters
}
\description{
Title
}

