% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_8_link_nodes_within_groups.R
\name{link_nodes_within_groups}
\alias{link_nodes_within_groups}
\title{Link nodes within groups}
\usage{
link_nodes_within_groups(num_nodes_per_group, n__num_groups, nodes, edge_list,
  bdpg_error_codes)
}
\arguments{
\item{num_nodes_per_group}{integer number of planning units per group in the
Xu problem}

\item{n__num_groups}{integer number of groups in the Xu problem, one of the
4 base input parameters for specifying the Xu problem}

\item{nodes}{data frame containing integer node_ID, integer group_ID, and boolean dependent_set_member}

\item{edge_list}{two column integer matrix of node IDs with one row for
each edge and columns for the 2 ends of the edge}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns list containing edge_list and cur_row
}
\description{
Create species in a Xu problem by creating links between nodes in different
groups.  Each link will represent a species that appears on the planning
units at each end of the link and those planning units are in separate
groups.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{bdpg_error_codes}{
\preformatted{
bdpg_error_codes : List of 6
 $ ERROR_STATUS_num_inside_or_within_group_links_less_than_one: num 1001
 $ ERROR_STATUS_optimal_solution_is_not_optimal               : num 1002
 $ ERROR_STATUS_num_nodes_per_group_must_be_at_least_2        : num 1003
 $ ERROR_STATUS_duplicate_spp_in_Xu_input_file                : num 1004
 $ ERROR_STATUS_unknown_spp_occ_FP_error_type                 : num 1005
 $ ERROR_STATUS_unknown_spp_occ_FN_error_type                 : num 1006
}}
\subsection{cur_group_ID}{
\preformatted{
cur_group_ID :  int 61
}}
\subsection{cur_group_nodes_sorted}{
\preformatted{
cur_group_nodes_sorted :  int [1:2] 121 122
}}
\subsection{cur_idx}{
\preformatted{
cur_idx :  int 1
}}
\subsection{cur_row}{
\preformatted{
cur_row :  num 62
}}
\subsection{edge_list}{
\preformatted{
edge_list :  int [1:1003, 1:2] 1 3 5 7 9 11 13 15 17 19 ...
}}
\subsection{n__num_groups}{
\preformatted{
n__num_groups :  num 61
}}
\subsection{nodes}{
\preformatted{
nodes : 'data.frame':	122 obs. of  3 variables:
 $ node_ID             : int  1 2 3 4 5 6 7 8 9 10 ...
 $ group_ID            : num  1 1 2 2 3 3 4 4 5 5 ...
 $ dependent_set_member: logi  FALSE TRUE FALSE TRUE FALSE TRUE ...
}}
\subsection{num_nodes_per_group}{
\preformatted{
num_nodes_per_group :  num 2
}}
\subsection{num_nodes_per_group_minus_1}{
\preformatted{
num_nodes_per_group_minus_1 :  num 1
}}
\subsection{other_node_idx}{
\preformatted{
other_node_idx :  int 2
}}
}

