% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{list_as_data_frame_with_0_length_vals_replaced_by_NA}
\alias{list_as_data_frame_with_0_length_vals_replaced_by_NA}
\title{Convert list to data frame with 0 length values replaced by NA}
\usage{
list_as_data_frame_with_0_length_vals_replaced_by_NA(a_list)
}
\arguments{
\item{a_list}{a list}
}
\value{
a data frame whose columns correspond to elements of the list and
any 0 length data value in the input list has been replaced with NA
}
\description{
Convert list to data frame with 0 length values replaced by NA
}
