% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{list_with_0_length_vals_replaced_by_NA}
\alias{list_with_0_length_vals_replaced_by_NA}
\title{Convert list to a new list with 0 length values replaced by NA}
\usage{
list_with_0_length_vals_replaced_by_NA(a_list)
}
\arguments{
\item{a_list}{a list}
}
\value{
the input list with its 0 length values replaced by NA values and
non-0 length values left as they were
}
\description{
Convert list to a new list with 0 length values replaced by NA
}
