% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14_read_marxan_output_files.R
\name{load_marxan_best_df_from_file_and_sort_and_add_missing_PUs}
\alias{load_marxan_best_df_from_file_and_sort_and_add_missing_PUs}
\title{Load marxan best data frame from file, sort, add missing PUs}
\usage{

  load_marxan_best_df_from_file_and_sort_and_add_missing_PUs(marxan_output_dir_path,
  all_correct_node_IDs)
}
\arguments{
\item{marxan_output_dir_path}{character string}

\item{all_correct_node_IDs}{integer vector}
}
\value{
Returns marxan_best_df_sorted data frame
}
\description{
Load marxan best data from file and sort in increasing order by planning
unit ID.  Also add any missing planning units.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{all_correct_node_IDs}{
\preformatted{
all_correct_node_IDs :  int [1:407] 1 2 3 4 5 6 7 8 9 10 ...
}}
\subsection{marxan_best_df_sorted}{
\preformatted{
marxan_best_df_sorted : 'data.frame':	407 obs. of  2 variables:
 $ PUID    : int  1 2 3 4 5 6 7 8 9 10 ...
 $ SOLUTION: int  0 1 0 1 0 1 0 1 0 1 ...
}}
\subsection{marxan_best_df_unsorted}{
\preformatted{
marxan_best_df_unsorted : 'data.frame':	407 obs. of  2 variables:
 $ PUID    : int  1 2 3 4 5 6 7 8 9 10 ...
 $ SOLUTION: int  0 1 0 1 0 1 0 1 0 1 ...
}}
\subsection{marxan_best_df_unsorted_without_missing_rows}{
\preformatted{
marxan_best_df_unsorted_without_missing_rows : 'data.frame':	407 obs. of  2 variables:
 $ PUID    : int  1 2 3 4 5 6 7 8 9 10 ...
 $ SOLUTION: int  0 1 0 1 0 1 0 1 0 1 ...
}}
\subsection{marxan_output_best_file_name}{
\preformatted{
marxan_output_best_file_name :  chr "output_best.csv"
}}
\subsection{marxan_output_dir_path}{
\preformatted{
marxan_output_dir_path :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSrun_-COR-Wrap-Marxan_SA.9"| __truncated__
}}
}

