% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14_read_marxan_output_files.R
\name{load_marxan_solutionsmatrix_and_add_missing_PUs}
\alias{load_marxan_solutionsmatrix_and_add_missing_PUs}
\title{Load marxan solutions matrix from file, sort, add missing PUs}
\usage{
load_marxan_solutionsmatrix_and_add_missing_PUs(
  marxan_output_dir_path,
  largest_PU_ID
)
}
\arguments{
\item{marxan_output_dir_path}{character string}

\item{largest_PU_ID}{integer}
}
\value{
Returns marxan_solutions_matrix_and_num_solutions
}
\description{
Load marxan solutions matrix from file, sort, add missing PUs
}
