% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14_read_marxan_output_files.R
\name{load_marxan_solutionsmatrix_from_file_and_sort_and_add_missing_PUs}
\alias{load_marxan_solutionsmatrix_from_file_and_sort_and_add_missing_PUs}
\title{Load marxan solutions matrix from file, sort, add missing PUs}
\usage{

  load_marxan_solutionsmatrix_from_file_and_sort_and_add_missing_PUs(marxan_output_dir_path,
  largest_PU_ID)
}
\arguments{
\item{marxan_output_dir_path}{character string}

\item{largest_PU_ID}{integer}
}
\value{
Returns marxan_solutions_matrix_and_num_solutions
}
\description{
Load marxan solutions matrix from file, sort, add missing PUs
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{cur_idx}{
\preformatted{
cur_idx :  num 407
}}
\subsection{cur_PU_ID}{
\preformatted{
cur_PU_ID :  num 407
}}
\subsection{cur_PU_name}{
\preformatted{
cur_PU_name :  chr "P407"
}}
\subsection{largest_PU_ID}{
\preformatted{
largest_PU_ID :  num 407
}}
\subsection{marxan_output_dir_path}{
\preformatted{
marxan_output_dir_path :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSrun_-COR-Wrap-Marxan_SA.9"| __truncated__
}}
\subsection{marxan_output_solutionsmatrix_df_unsorted_without_missing_rows}{
\preformatted{
marxan_output_solutionsmatrix_df_unsorted_without_missing_rows : 'data.frame':	4 obs. of  408 variables:
 $ SolutionsMatrix: chr  "S1" "S2" "S3" "S4"
 $ P1             : int  0 0 0 0
 $ P2             : int  1 1 1 1
 ...
 $ P98            : int  1 1 1 1
  [list output truncated]
}}
\subsection{marxan_solutions_matrix}{
\preformatted{
marxan_solutions_matrix :  num [1:4, 1:407] 0 0 0 0 1 1 1 1 0 0 ...
}}
\subsection{marxan_solutions_matrix_and_num_solutions}{
\preformatted{
marxan_solutions_matrix_and_num_solutions : List of 2
 $ marxan_solutions_matrix: num [1:4, 1:407] 0 0 0 0 1 1 1 1 0 0 ...
 $ num_marxan_solutions   : int 4
}}
\subsection{num_marxan_solutions}{
\preformatted{
num_marxan_solutions :  int 4
}}
\subsection{num_PU_IDs_in_solutionsmatrix}{
\preformatted{
num_PU_IDs_in_solutionsmatrix :  int 407
}}
\subsection{output_solutionsmatrix_file_name}{
\preformatted{
output_solutionsmatrix_file_name :  chr "output_solutionsmatrix.csv"
}}
\subsection{PU_IDs_in_solutions}{
\preformatted{
PU_IDs_in_solutions :  num [1:407] 1 2 3 4 5 6 7 8 9 10 ...
}}
\subsection{PU_names}{
\preformatted{
PU_names :  chr [1:407] "P1" "P2" "P3" "P4" "P5" "P6" "P7" "P8" "P9" "P10" ...
}}
}

