% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_action_using_tzar_reps.R
\name{look_up_object_by_UUID}
\alias{look_up_object_by_UUID}
\title{Search the set of files for a file name containing the UUID and load that file}
\usage{
look_up_object_by_UUID(UUID_to_find, rds_file_set)
}
\arguments{
\item{UUID_to_find}{a UUID string to look for in file names}

\item{rds_file_set}{vector of file names (possibly including paths)}
}
\value{
Returns the object loaded from the rds file if found, NULL if no
file name containing given UUID was found
}
\description{
Given a vector of file names (possibly including their paths), find the
file name that contains the given UUID and load that file
}
