% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_action_using_tzar_reps.R
\name{make_sure_that_cor_bdprob_is_base_of_app_bdprob}
\alias{make_sure_that_cor_bdprob_is_base_of_app_bdprob}
\title{Make sure that the correct bdprob is the base for the given apparent bdprob}
\usage{
make_sure_that_cor_bdprob_is_base_of_app_bdprob(
  app_bdprob,
  cor_bdprob,
  parameters
)
}
\arguments{
\item{app_bdprob}{an apparent bdprob}

\item{cor_bdprob}{a correct bdprob}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}
}
\value{
Returns nothing.
}
\description{
If the given correct Xu_bd_problem is not the problem that the apparent has
added error to, this is a fatal error, so give an error message and quit.
}
