% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_11_summarize_and_plot_graph_structure_information.R
\name{plot_degree_and_abundance_dists_for_node_graph}
\alias{plot_degree_and_abundance_dists_for_node_graph}
\title{Plot degree and abundance distributions for node graph}
\usage{
plot_degree_and_abundance_dists_for_node_graph(
  final_link_counts_for_each_node,
  final_rank_abundance_dist,
  PU_col_name,
  plot_output_dir,
  cor_or_app_label,
  spp_col_name
)
}
\arguments{
\item{final_link_counts_for_each_node}{integer vector of counts of number
of links for each node, i.e., number of species for each planning unit}

\item{final_rank_abundance_dist}{integer vector of counts of number
of nodes for each link, i.e., number of planning units for each species
(same as final_node_counts_for_each_link)}

\item{PU_col_name}{character string giving planning unit column name in data frames}

\item{plot_output_dir}{character string giving pat to directory where plot
should be written}

\item{cor_or_app_label}{character string for labelling as Correct or Apparent}

\item{spp_col_name}{character string giving species column name in data frames}
}
\value{
Doesn't return anything
}
\description{
Plot the final degree distribution for the node graph, which is essentially
the distribution of number of species per planning unit.  Similarly,
plot number of planning units for each species, sorted in decreasing order
of number of planning units.
}
