% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_11_summarize_and_plot_graph_structure_information.R
\name{plot_final_degree_dist_for_node_graph}
\alias{plot_final_degree_dist_for_node_graph}
\title{Plot final degree distribution for node graph}
\usage{
plot_final_degree_dist_for_node_graph(final_link_counts_for_each_node,
  PU_col_name, plot_output_dir, cor_or_app_label)
}
\arguments{
\item{final_link_counts_for_each_node}{integer vector of counts of number
of links for each node, i.e., number of species for each planning unit}

\item{PU_col_name}{character string giving planning unit column name in data frames}

\item{plot_output_dir}{character string giving pat to directory where plot
should be written}

\item{cor_or_app_label}{character string for labelling as Correct or Apparent}
}
\value{
Doesn't return anything
}
\description{
Plot the final degree distribution for the node graph, which is essentially
the distribution of number of species per planning unit.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{cor_or_app_label}{
\preformatted{
cor_or_app_label :  chr "COR"
}}
\subsection{final_degree_dist}{
\preformatted{
final_degree_dist : 'data.frame':	122 obs. of  2 variables:
 $ PU_ID: int  1 2 3 4 5 6 7 8 9 10 ...
 $ freq : int  33 33 33 31 31 31 30 29 29 29 ...
}}
\subsection{final_link_counts_for_each_node}{
\preformatted{
final_link_counts_for_each_node : 'data.frame':	122 obs. of  2 variables:
 $ PU_ID: int  1 2 3 4 5 6 7 8 9 10 ...
 $ freq : int  1 25 1 27 1 29 1 28 1 24 ...
}}
\subsection{plot_output_dir}{
\preformatted{
plot_output_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSprob-COR-Base.d0729e1c-ea"| __truncated__
}}
\subsection{PU_col_name}{
\preformatted{
PU_col_name :  chr "PU_ID"
}}
}

