% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_14c_plot_incremental_marxan_summed_solution_reps.R
\name{plot_incremental_marxan_summed_solution_representations}
\alias{plot_incremental_marxan_summed_solution_representations}
\title{Plot incremental marxan summed solution representations}
\usage{
plot_incremental_marxan_summed_solution_representations(marxan_ssoln_df,
  cor_PU_costs, optimum_cost, bpm, cor_app_prefix_string, num_spp,
  plot_output_dir)
}
\arguments{
\item{marxan_ssoln_df}{data frame}

\item{cor_PU_costs}{numeric vector}

\item{optimum_cost}{numeric}

\item{bpm}{matrix}

\item{cor_app_prefix_string}{character string}

\item{num_spp}{integer}

\item{plot_output_dir}{character string}
}
\value{
Returns nothing
}
\description{
For each step in order by Marxan summed solution PU ID:
 Want the fraction of all species who have met or exceeded their target
 when all PUs with the same number of votes or more are included in the
 solution.
}
