% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14_read_marxan_output_files.R
\name{read_APP_marxan_output_files}
\alias{read_APP_marxan_output_files}
\title{Read APP marxan output files}
\usage{
read_APP_marxan_output_files(rsrun, APP_bd_prob, COR_bd_prob, parameters)
}
\arguments{
\item{rsrun}{an RSrun object}

\item{APP_bd_prob}{a Xu_bd_problem object}

\item{COR_bd_prob}{a Xu_bd_problem object}

\item{parameters}{list}
}
\value{
Returns marxan_output_values
}
\description{
Read APPARENT marxan output files.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.
\subsection{APP_bd_prob}{
\preformatted{
APP_bd_prob : Formal class 'Xu_bd_problem' [package "bdpg"] with 35 slots
}}
\subsection{COR_bd_prob}{
\preformatted{
COR_bd_prob : Formal class 'Xu_wrapped_bd_problem' [package "bdpg"] with 36 slots
}}
\subsection{marxan_output_dir}{
\preformatted{
marxan_output_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSrun_-APP-Wrap-Marxan_SA.3"| __truncated__
}}
\subsection{marxan_output_values}{
\preformatted{
marxan_output_values : List of 3
 $ marxan_best_df_sorted:'data.frame':	407 obs. of  2 variables:
 $ marxan_ssoln_df      :'data.frame':	407 obs. of  2 variables:
 $ marxan_mvbest_df     :'data.frame':	1277 obs. of  10 variables:
}}
\subsection{parameters}{
\preformatted{
parameters : List of 66
 $ summary_without_run_id_filename                           : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/prob_diff_results_with_0_ru"| __truncated__
 ...
 $ fullOutputDir_NO_slash                                    : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{rsrun}{
\preformatted{
rsrun : Formal class 'RSrun' [package "bdpg"] with 12 slots
}}
\subsection{top_dir}{
\preformatted{
top_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
}

