% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob.R
\name{read_Xu_problem_from_Xu_file}
\alias{read_Xu_problem_from_Xu_file}
\title{FUNCTION_TITLE}
\usage{
read_Xu_problem_from_Xu_file(infile_name, given_correct_solution_cost)
}
\arguments{
\item{infile_name}{DESCRIPTION.}

\item{given_correct_solution_cost}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

