% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{remove_base_spp_abundances_from_wrapping_distribution}
\alias{remove_base_spp_abundances_from_wrapping_distribution}
\title{FUNCTION_TITLE}
\usage{
remove_base_spp_abundances_from_wrapping_distribution(Xu_PU_spp_table,
  trimmed_rounded_abund_per_spp, spp_col_name)
}
\arguments{
\item{Xu_PU_spp_table}{DESCRIPTION.}

\item{trimmed_rounded_abund_per_spp}{DESCRIPTION.}

\item{spp_col_name}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

