% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSrun.R
\name{repro_RSrun_gui}
\alias{repro_RSrun_gui}
\title{Create an RSrun}
\usage{
repro_RSrun_gui(repro_RDS_file_loc = NULL, fullOutputDir_NO_slash = NULL)
}
\arguments{
\item{repro_RDS_file_loc}{path to RDS file to be loaded}

\item{fullOutputDir_NO_slash}{full path to directory where output should go
(with no slash on the end of the path)}
}
\value{
Returns an RSrun object
}
\description{
Create a run of a reserve selector
}
