% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_marxan_control_file_and_run_marxan.R
\name{run_marxan}
\alias{run_marxan}
\title{Run marxan}
\usage{
run_marxan(marxan_dir, marxan_executable_name)
}
\arguments{
\item{marxan_dir}{character string}

\item{marxan_executable_name}{character string}
}
\value{
Returns system return code from running marxan
}
\description{
Run marxan program
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{marxan_dir}{
\preformatted{
marxan_dir :  chr "/Users/bill/tzar/modelcode/marxanLibrary_1360018653/"
}}
\subsection{marxan_executable_name}{
\preformatted{
marxan_executable_name :  chr "MarOpt_v243_Mac64"
}}
\subsection{original_dir}{
\preformatted{
original_dir :  chr "/Users/bill/D/Projects/ProblemDifficulty/pkgs/bdpgxupaper"
}}
\subsection{retval}{
\preformatted{
retval :  int 0
}}
\subsection{system.command.run.marxan}{
\preformatted{
system.command.run.marxan :  chr "./MarOpt_v243_Mac64 -s"
}}
}

