% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{safe_sample}
\alias{safe_sample}
\title{Safe version of R sample function}
\usage{
safe_sample(x, ...)
}
\arguments{
\item{x}{Either a vector of one or more elements from which to choose, or a positive integer.}

\item{...}{Arguments that can be passed on to base::sample()}
}
\value{
Returns a value that is probably an integer but is definitely not null
}
\description{
2015 04 08 - BTL
 I just got bitten very badly by the incredibly annoying behavior of R's
 sample() function, so here is a replacement function that I need to
 use everywhere now.
 When I called sample with a vector that sometimes had length n=1,
 it sampled from 1:n instead of returning the single value.
 This majorly screwed all kinds of things in a very subtle, very hard
 to find way.
}
