% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{save_bdprob}
\alias{save_bdprob}
\title{Save bd problem to disk}
\usage{
save_bdprob(bdprob_type, app_vs_cor, uuid, base_outdir, bdprob)
}
\arguments{
\item{bdprob_type}{character string indicating problem type, either "BASIC" or "WRAPPED" (could eventually also be "combined")}

\item{app_vs_cor}{character string indicating whether it's a correct or apparent problem, either "COR" or "APP"}

\item{uuid}{character string indicating the UUID of the problem being saved}

\item{base_outdir}{character sting giving the full path to the directory where the problem will be saved}

\item{bdprob}{a Xu_bd_problem to write to disk (subclasses of Xu_bd_problem allowed)}
}
\value{
character string giving the full path (including file name) to file where problem is saved on disk
}
\description{
After a problem is generated, its R representation is saved to disk so
that it can be archived and re-used in future experiments without having
to regenerate the problem.  Saving it is also useful for reproducibility
in that it allows re-creation of exactly the problem used in an experiment.
}
\details{
Writes to a file whose name contains:
\itemize{
 \item{the UUID of the problem}
 \item{whether it's a basic problem or a wrapped problem}
 \item{whether it's a correct or an apparent problem}
}
For example:  "saved_bdprob.f470f75b-116b-4ff9-9db0-4e9448bcb2ef.BASIC.COR.rds"

Note that the BASIC/WRAPPED and COR/APP arguments to save_bdprob() don't have
to be capitalized or even use the suggested strings.  They can be anything
that you want to have built into those spots in the file name.  Capitalizing
and using those strings has just proven easy to spot or search for in
directory listings so they seem like a good convention to use.
}
\examples{
\dontrun{
Xu_bdprob_cor@full_saved_bdprob_path =
    save_bdprob ("BASIC", "COR", Xu_bdprob_cor@UUID, Xu_bdprob_cor@base_outdir,
                 Xu_bdprob_cor)
Xu_bdprob_app@full_saved_bdprob_path =
    save_bdprob ("BASIC", "APP", Xu_bdprob_app@UUID, Xu_bdprob_app@base_outdir,
                 Xu_bdprob_app)

Xu_bdprob_cor@full_saved_bdprob_path =
    save_bdprob ("WRAPPED", "COR", Xu_bdprob_cor@UUID, Xu_bdprob_cor@base_outdir,
                 Xu_bdprob_cor)
Xu_bdprob_app@full_saved_bdprob_path =
    save_bdprob ("WRAPPED", "APP", Xu_bdprob_app@UUID, Xu_bdprob_app@base_outdir,
                 Xu_bdprob_app)

#  To reload the problem as a Xu_bd_problem, call readRDS () with the
#  full path to the saved file
reloaded_Xu_prob =
    readRDS ("saved_bdprob.f470f75b-116b-4ff9-9db0-4e9448bcb2ef.BASIC.COR.rds")
}
}

