% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_checksums.R
\name{save_obj_with_checksum}
\alias{save_obj_with_checksum}
\title{Save object with checksum}
\usage{
save_obj_with_checksum(obj_with_UUID_and_checksum, base_outdir_for_checksum)
}
\arguments{
\item{obj_with_UUID_and_checksum}{object whose checksum is to be computed}

\item{base_outdir_for_checksum}{directory where temporary file for use in checksum
computation will be written and then deleted}
}
\value{
Returns input object with its checksum slot filled
}
\description{
Save the object to disk with its checksum slot set.
}
\details{
Note that there may be a way to compress these files when they're saved,
but I don't quite understand from the saveRDS() help file whether you
have to do the compression by calling gzfile() or some argument to saveRDS().
It may be worth looking into more if these objects get large because of
the adjacency matrix, since the examples in the help file for readRDS()
look like readRDS() can directly restore an R object that's been saved
using saveRDS() and then had gzip() applied to it.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{base_outdir_for_checksum}{
\preformatted{
base_outdir_for_checksum :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSprob-COR-Base.d0729e1c-ea"| __truncated__
}}
\subsection{full_saved_obj_path}{
\preformatted{
full_saved_obj_path :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSprob-COR-Base.d0729e1c-ea"| __truncated__
}}
\subsection{obj_with_UUID_and_checksum}{
\preformatted{
obj_with_UUID_and_checksum : Formal class 'Xu_bd_problem' [package "bdpg"] with 35 slots
}}
\subsection{saved_obj_with_checksum_filename}{
\preformatted{
saved_obj_with_checksum_filename :  chr "saved.RSprob-COR-Base.d0729e1c-eadc-4899-a382-8cb7ac2c08d7.rds"
}}
}

