% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_error_to_spp_occupancy_data.R
\name{set_const_FP_and_FN_rates}
\alias{set_const_FP_and_FN_rates}
\title{Set constant FP and FN rates}
\usage{
set_const_FP_and_FN_rates(parameters, bdpg_error_codes)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns FP_and_FN_const_rates list
}
\description{
Set the False Positive and False Negative error rate to either a
given constant value or a constant value chosen from a uniform random
distribution whose upper and lower bounds are given.
}
\details{
As a result, the False
Positive error rate for every PU/spp pair in the problem will be identical.
The False Negative error rate for every PU/spp pair will also be identical
but not necessarily the same value as the False Positive rate.
}
