% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_13_write_marxan_control_file_and_run_marxan.R
\name{set_marxan_controls_and_run_marxan}
\alias{set_marxan_controls_and_run_marxan}
\title{Title}
\usage{
set_marxan_controls_and_run_marxan(marxan_input_dir, marxan_output_dir,
  parameters, marxan_IO_dir)
}
\arguments{
\item{marxan_IO_dir}{}
}
\description{
Title
}

