% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_seed_setting_utilities.R
\name{set_seed_if_necessary_helper}
\alias{set_seed_if_necessary_helper}
\title{Helper function for setting a seed when necessary}
\usage{
set_seed_if_necessary_helper(
  set_rand_seed_at_creation_of_all_new_major_objects,
  location_string,
  forced_seed = NULL
)
}
\arguments{
\item{set_rand_seed_at_creation_of_all_new_major_objects}{boolean indicating
whether a new seed should be set every time an rsrun or rsproblem object
is created}

\item{location_string}{string to indicate where this function was called,
e.g., at the start of the creation of an rsproblem}

\item{forced_seed}{integer to use as seed in set.seed() call or
NULL to indicate that no seed has been provided}
}
\value{
Returns a 2 element named list with element new_seed containing
the new integer seed value or NA and the element R_internal_seed_array
containing the array value .Random.seed at the end of this function (i.e.,
the internal state of R's random number generator)
}
\description{
Helper function for setting a seed when necessary
}
