% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_gurobi_control_file_and_run_gurobi.R
\name{set_up_for_and_run_gurobi_APP}
\alias{set_up_for_and_run_gurobi_APP}
\title{Set up for and run gurobi for APP problem}
\usage{
set_up_for_and_run_gurobi_APP(APP_bd_prob, COR_bd_prob, gurobi_run, parameters)
}
\arguments{
\item{APP_bd_prob}{an apparent Xu_bd_problem (or subclass)}

\item{COR_bd_prob}{the correct Xu_bd_problem (or subclass) that the apparent problem is derived from}

\item{gurobi_run}{an RSrun object (or subclass)}

\item{parameters}{parameters list for the run, usually derived from project.yaml}
}
\value{
Returns named list of gurobi control and return values
}
\description{
Convenience function to call set_up_for_and_run_gurobi with proper
arguments for an apparent Xu_bd_problem
}
