% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_gurobi_control_file_and_run_gurobi.R
\name{set_up_for_and_run_gurobi_COR}
\alias{set_up_for_and_run_gurobi_COR}
\title{Set up for and run gurobi for COR problem}
\usage{
set_up_for_and_run_gurobi_COR(COR_bd_prob, rsrun, parameters, starting_dir,
  marxan_elapsed_time = NA)
}
\arguments{
\item{COR_bd_prob}{a correct Xu_bd_problem}

\item{rsrun}{an RSrun object (or subclass)}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{starting_dir}{string containing full path of top directory where
problems and/or runs are to be created}

\item{marxan_elapsed_time}{numeric value containing elapsed time for running
marxan (in seconds)}
}
\value{
Returns named list of gurobi control and return values
}
\description{
Convenience function to call set_up_for_and_run_gurobi with proper
arguments for a correct Xu_bd_problem.
}
