% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_marxan_control_file_and_run_marxan.R
\name{set_up_for_and_run_marxan_APP}
\alias{set_up_for_and_run_marxan_APP}
\title{Set up for and run marxan for APP problem}
\usage{
set_up_for_and_run_marxan_APP(
  APP_bd_prob,
  COR_bd_prob,
  marxan_run,
  parameters,
  starting_dir
)
}
\arguments{
\item{APP_bd_prob}{an apparent Xu_bd_problem (or subclass)}

\item{COR_bd_prob}{the correct Xu_bd_problem (or subclass) that the apparent problem is derived from}

\item{marxan_run}{an RSrun object (or subclass)}

\item{parameters}{parameters list for the run, usually derived from project.yaml}

\item{starting_dir}{string containing full path of top directory where
problems and/or runs are to be created}
}
\value{
list containing marxan_control_values and updated APP_bd_prob
}
\description{
Convenience function to call set_up_for_and_run_marxan with proper
arguments for an apparent Xu_bd_problem
}
