% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_simple_richness_control_file_and_run_SR.R
\name{set_up_for_and_run_simple_richness}
\alias{set_up_for_and_run_simple_richness}
\title{Set up for and run simple_richness}
\usage{
set_up_for_and_run_simple_richness(
  num_spp,
  num_PUs,
  bpm,
  PU_costs,
  spp_rep_targets,
  forward,
  rsrun,
  top_dir,
  save_inputs,
  save_outputs
)
}
\arguments{
\item{num_spp}{integer number of species in the problem}

\item{num_PUs}{integer number of planning units}

\item{bpm}{bipartite matrix; integer matrix with one row for each species
and one column for each planning unit.  Each matrix entry specifies
whether that species occupies that planning unit; 1 indicates the
species does occupy the planning unit and 0 indicates it does not.
Same as occ_matrix.}

\item{PU_costs}{numeric vector of planning unit costs}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}

\item{forward}{boolean indicating whether to do the greedy search in the
forward direction (adding PUs to the solution set) or the backward
direction (removing PUs from the solution set); TRUE implies forward}

\item{rsrun}{an RSrun (reserve selection run) object, e.g., a marxan run}

\item{top_dir}{character string giving the path to the top directory of the
tree where the object is written out, usually something like the tzar
output directory of the experiment followed by some object-specific
information and the UUID of the object, e.g.,
~/tzarout/RSprob-COR-Base.299140ae-3be2-4331-bb72-b24570902587}

\item{save_inputs}{boolean indicating whether input model and input params
should be saved to disk}

\item{save_outputs}{boolean indicating whether results and
simple_richness_controls_and_results params should be saved to disk}
}
\value{
Returns named list of simple_richness control and return values
}
\description{
Set up for and run simple_richness for either COR or APP
}
