% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_simple_richness_control_file_and_run_SR.R
\name{set_up_for_and_run_simple_richness_COR}
\alias{set_up_for_and_run_simple_richness_COR}
\title{Set up for and run simple_richness for COR problem}
\usage{
set_up_for_and_run_simple_richness_COR(
  COR_bd_prob,
  rsrun,
  parameters,
  starting_dir,
  forward = TRUE
)
}
\arguments{
\item{COR_bd_prob}{a correct Xu_bd_problem}

\item{rsrun}{an RSrun (reserve selection run) object, e.g., a marxan run}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{starting_dir}{string containing full path of top directory where
problems and/or runs are to be created}

\item{forward}{boolean indicating whether to do the greedy search in the
forward direction (adding PUs to the solution set) or the backward
direction (removing PUs from the solution set); TRUE implies forward}
}
\value{
Returns named list of simple_richness control and return values
}
\description{
Convenience function to call set_up_for_and_run_simple_richness with proper
arguments for a correct Xu_bd_problem.
}
