% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_9a_create_Xu_graph.R
\name{sort_within_rows}
\alias{sort_within_rows}
\title{Sort a 2 column matrix within rows}
\usage{
sort_within_rows(a_2_col_matrix, decreasing = FALSE)
}
\arguments{
\item{a_2_col_matrix}{a 2 column numeric matrix}

\item{decreasing}{boolean flag indicating whether to sort in increasing or
decreasing order; TRUE implies sort in decreasing order, FALSE implies
sort in increasing order}
}
\value{
Returns sorted 2 column matrix
}
\description{
Sort a 2 column matrix within rows
}
