% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function_documentation.R
\name{stop_due_to_unexpected_token}
\alias{stop_due_to_unexpected_token}
\title{Quit since an unexpected token was found}
\usage{
stop_due_to_unexpected_token(token, data_line_num, data_line)
}
\arguments{
\item{token}{character string indicating the token type, e.g.,
"whitespace" or "symbol"}

\item{data_line_num}{integer line number of data_line in original input file}

\item{data_line}{character string containing one full line from the input
file}
}
\value{
Returns nothing
}
\description{
Emits an error message containint the unexpected token, the input line
number where it occurred, and the line itself; then it quits.
}
