% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_pu_and_spp_histograms_from_marxan_pusvpr_file.R
\name{summarize_pu_and_spp_dist_data_from_marxan_puvspr_input_file}
\alias{summarize_pu_and_spp_dist_data_from_marxan_puvspr_input_file}
\title{Summarize PU and spp dist data from marxan puvspr input file}
\usage{
summarize_pu_and_spp_dist_data_from_marxan_puvspr_input_file(infile_path)
}
\arguments{
\item{infile_path}{character string giving path to the marxan puvspr.dat
file to be analyzed (including the name of the file itself), e.g.,
"~/marxan/input/puvspr.dat"}
}
\value{
Returns pu_and_spp_dist_data
}
\description{
Summarize planning unit and species distribution data from marxan puvspr
input file.
}
\examples{
\dontrun{
too_long_path_part1 = "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/"
too_long_path_part2 = "1834_marxan_simulated_annealing.completedTzarEmulation/"
too_long_path_part3 = "RSrun_-COR-Wrap-Marxan_SA."
too_long_path_part4 = "408782d5-2bb6-4c09-8b18-f05ad3089b15/input/puvspr.dat"
infile_path <- paste0 (too_long_path_part1, too_long_path_part2,
                       too_long_path_part3, too_long_path_part4)
pu_and_spp_dist_data <-
  summarize_pu_and_spp_dist_data_from_marxan_puvspr_input_file (infile_path)
}
}
