% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{trim_abundances}
\alias{trim_abundances}
\title{FUNCTION_TITLE}
\usage{
trim_abundances(rounded_abundances, min_abund = 2,
  max_abund = .Machine$double.xmax)
}
\arguments{
\item{rounded_abundances}{DESCRIPTION.}

\item{min_abund}{DESCRIPTION.}

\item{max_abund}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

