% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{validate_wrap}
\alias{validate_wrap}
\title{Validate the wrapping of a bdproblem}
\usage{
validate_wrap(extra_abund, dep_set_PUs_eligible, PU_spp_table, dep_set)
}
\arguments{
\item{extra_abund}{blah}

\item{dep_set_PUs_eligible}{blah}

\item{PU_spp_table}{blah}

\item{dep_set}{blah}
}
\value{
TRUE if wrap is valid, otherwise quits before end of function
export
}
\description{
Check a wrapping distribution to see that it will generate a legal wrapped
problem.
}
