% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{value_or_1_if_null}
\alias{value_or_1_if_null}
\title{Get parameter value that should be integer and return 1 if value is null}
\usage{
value_or_1_if_null(value)
}
\arguments{
\item{value}{a value that is probably a number, but could be anything}
}
\value{
Returns a value that is probably an integer but is definitely not null
}
\description{
Get parameter value that should be integer and return 1 if value is null
}
