% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{vb}
\alias{vb}
\title{Validate input as boolean and replace with default on empty input if desired}
\usage{
vb(var_value, def_on_empty = FALSE, def = FALSE,
  treat_NULL_as_empty = TRUE, treat_NA_as_empty = FALSE,
  allow_num = FALSE)
}
\arguments{
\item{var_value}{the value to be checked to see if it's a boolean}

\item{def_on_empty}{boolean flag indicating whether to return a default
value instead of the input value when the input value is empty (where
empty is defined by other flags below)}

\item{def}{a TRUE or FALSE default value to return instead of the input value
when a default is requested}

\item{treat_NULL_as_empty}{a boolean flag set to TRUE if a NULL input is to
be treated as an empty input; FALSE otherwise}

\item{treat_NA_as_empty}{a boolean flag set to TRUE if an NA input is to
be treated as an empty input; FALSE otherwise}

\item{allow_num}{a boolean flag set to TRUE if a numeric input is to be
allowed and returned as FALSE if 0 and TRUE if non-zero}
}
\value{
Returns the input value if it is boolean or, returns a specified
    boolean if other arguments force a valid converted or default boolean
    to return; otherwise, throws an error
}
\description{
This function is intended to be a more flexible replacement for calling
is.logical().  In particular, it makes it possible to replace empty values
with a default value (e.g., FALSE) and to define whether NULL and/or NA
are treated as empty values.  It also allows control over whether numeric
values other than 0 and 1 are considered boolean by this validation function.
This is sometimes useful when validating an input option where a non-0/1
value would be suggestive of a mistake in the input file.
}
