% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{wrap_abundance_dist_around_Xu_problem}
\alias{wrap_abundance_dist_around_Xu_problem}
\title{FUNCTION_TITLE}
\usage{
wrap_abundance_dist_around_Xu_problem(starting_dir, rounded_abundances,
  Xu_bdprob, dep_set_PUs_eligible, tot_num_PUs_in_landscape, bdpg_error_codes,
  wrap_prob_name_stem = "wrap_prob", cor_dir_name_stem = "cor")
}
\arguments{
\item{starting_dir}{DESCRIPTION.}

\item{rounded_abundances}{DESCRIPTION.}

\item{Xu_bdprob}{DESCRIPTION.}

\item{dep_set_PUs_eligible}{DESCRIPTION.}

\item{tot_num_PUs_in_landscape}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
