% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{wrap_abundances_around_eligible_set}
\alias{wrap_abundances_around_eligible_set}
\title{FUNCTION_TITLE}
\usage{
wrap_abundances_around_eligible_set(dep_set, eligible_set,
  rounded_abund_per_spp, num_base_spp, Xu_PU_spp_table,
  min_allowed_abundance = 2, PU_col_name = "PU_ID",
  spp_col_name = "spp_ID")
}
\arguments{
\item{dep_set}{DESCRIPTION.}

\item{eligible_set}{DESCRIPTION.}

\item{rounded_abund_per_spp}{DESCRIPTION.}

\item{num_base_spp}{DESCRIPTION.}

\item{Xu_PU_spp_table}{DESCRIPTION.}

\item{min_allowed_abundance}{DESCRIPTION.}

\item{PU_col_name}{DESCRIPTION.}

\item{spp_col_name}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

