% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_marxan_input_files.R
\name{write_marxan_pu.dat_input_file}
\alias{write_marxan_pu.dat_input_file}
\title{Write marxan planning units input file (pu.dat) using constant values for cost and status.}
\usage{
write_marxan_pu.dat_input_file(PU_IDs, cost_const = 1,
  status_const = 0)
}
\arguments{
\item{PU_IDs}{A vector of planning unit IDs}

\item{cost_const}{A constant cost to be applied to all species}

\item{status_const}{A constant status to be applied to all species}
}
\value{
nothing.
}
\description{
Write marxan planning units input file (pu.dat) using constant values for cost and status.
}
\examples{
\dontrun{
PU_IDs = c(3,5,6,7,21,32)
write_marxan_pu.dat_input_file (PU_IDs)

PU_IDs = c(3,5,6,7,21,32)
write_marxan_pu.dat_input_file (PU_IDs, 3, 0)
         }
}
