% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_marxan_input_files.R
\name{write_marxan_puvspr.dat_input_file}
\alias{write_marxan_puvspr.dat_input_file}
\title{Write planning unit vs. species input file (puvspr.dat).}
\usage{
write_marxan_puvspr.dat_input_file(spp_PU_amount_table)
}
\arguments{
\item{spp_PU_amount_table}{A data frame of species IDs vs planning unit IDs}
}
\value{
nothing.
}
\description{
Write planning unit vs. species input file (puvspr.dat).
}
\details{
The spp_PU_amount_table is expected to have 3 columns called (in left to
right order) "species",
"pu", and "amount".  Marxan requires this table to be sorted in increasing
order on the planning unit column.  I don't think the order within planning
unit matters on the
other columns.  See source code for the test function
\code{gen_random_spp_PU_amount_table} for an example of creating
this table.
}
\examples{
\dontrun{
num_PUs = 100
num_spp = 3
spp_PU_amount_table = gen_random_spp_PU_amount_table (num_PUs, num_spp)
write_marxan_puvspr.dat_input_file (spp_PU_amount_table)
         }
}

