% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_marxan_input_files.R
\name{write_marxan_spec.dat_input_file}
\alias{write_marxan_spec.dat_input_file}
\title{Write species input file (spec.dat) using constant values for species penalties and targets.}
\usage{
write_marxan_spec.dat_input_file(spp_IDs, spf_const = 1, target_const = 1)
}
\arguments{
\item{spp_IDs}{A vector of species IDs}

\item{spf_const}{A constant species penalty factor to be applied to all species}

\item{target_const}{A constant target representation level to be applied to all species}
}
\value{
nothing.
}
\description{
Write species input file (spec.dat) using constant values for species penalties and targets.
}
\examples{
\dontrun{
spp_IDs = 1:10
write_marxan_spec.dat_input_file (spp_IDs)
#'
spp_IDs = 1:10
write_marxan_spec.dat_input_file (spp_IDs, 10, 1)
         }
}
