% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_marxan_input_files.R
\name{write_marxan_spec.dat_input_file_from_vectors}
\alias{write_marxan_spec.dat_input_file_from_vectors}
\title{Write species input file (spec.dat) given vectors of target and spf values.}
\usage{
write_marxan_spec.dat_input_file_from_vectors(spp_IDs, spf_values,
  target_values)
}
\arguments{
\item{spp_IDs}{A vector of species IDs}

\item{spf_values}{A vector of species penalty factors}

\item{target_values}{A vector of target values}
}
\value{
nothing.
}
\description{
Write species input file (spec.dat) given vectors of target and spf values.
}
\examples{
\dontrun{
spp_IDs = 1:5
spf_values = c(2,3,8,1,4)
target_values = c(20,20,10,100,5)
write_marxan_spec.dat_input_file (spp_IDs, spf_values, target_values)
         }
}

