% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{write_results_to_files}
\alias{write_results_to_files}
\title{Write results to files}
\usage{
write_results_to_files(csv_outfile_name, results_df, run_ID, out_dir,
  tzar_run_id_field_name)
}
\arguments{
\item{csv_outfile_name}{Name of the csv file to write results to, not
including csv extension but not including path}

\item{results_df}{data frame containing the results}

\item{run_ID}{integer ID of a run}

\item{out_dir}{character string telling what directory to put the results
file in}

\item{tzar_run_id_field_name}{character string containing name of tzar run ID
field in output file}
}
\value{
Returns nothing.
}
\description{
This function is used two different ways.
It's called when the program quits because there are too many species
or it's called when the program runs successfully.
}
