% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_seed_setting_utilities.R
\name{always_set_new_or_forced_rand_seed}
\alias{always_set_new_or_forced_rand_seed}
\title{Always set a seed, either by generating a new one or reusing a given one}
\usage{
always_set_new_or_forced_rand_seed(location_string, forced_seed = NULL)
}
\arguments{
\item{location_string}{string to indicate where this function was called,
e.g., at the start of the creation of an rsproblem}

\item{forced_seed}{integer to use as seed in set.seed() call or
NULL to indicate that no seed has been provided}
}
\value{
Returns an integer seed value or NA
}
\description{
Generate a new seed based on the current time if the forced_seed argument
is NULL.  If that argument is not NULL, then use it as a seed.  Regardless
of which way the seed is derived, call set.seed() with it as the argument.
}
