% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_error_to_spp_occupancy_data.R
\name{apply_const_error_to_spp_occupancy_data}
\alias{apply_const_error_to_spp_occupancy_data}
\title{Apply constant error to spp occ data}
\usage{
apply_const_error_to_spp_occupancy_data(bpm, FP_rates, FN_rates, num_PUs,
  num_spp, random_values, bdpg_error_codes)
}
\arguments{
\item{bpm}{bipartite matrix; integer matrix with one row for each species
and one column for each planning unit.  Each matrix entry specifies
whether that species occupies that planning unit; 1 indicates the
species does occupy the planning unit and 0 indicates it does not.
Same as occ_matrix.}

\item{FP_rates}{numeric vector}

\item{FN_rates}{numeric vector}

\item{num_PUs}{integer number of planning units}

\item{num_spp}{integer number of species in the problem}

\item{random_values}{numeric vector}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns bpm matrix
}
\description{
Walk through the occupancy matrix (PU vs spp) and randomly
 choose to flip some of the TPs to FNs and TNs to FPs based on the
 given FP and FN error rates.
 Update the occupancy matrix (bpm) as you go along, but don't update
 the PU_spp_pair_indices until you know all locations that have
 flipped.  Update PU_spp_pair_indices at the end so that you don't
 have to constantly resize this big array as you go along.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{bdpg_error_codes}{
\preformatted{
bdpg_error_codes : List of 6
 $ ERROR_STATUS_num_inside_or_within_group_links_less_than_one: num 1001
 $ ERROR_STATUS_optimal_solution_is_not_optimal               : num 1002
 $ ERROR_STATUS_num_nodes_per_group_must_be_at_least_2        : num 1003
 $ ERROR_STATUS_duplicate_spp_in_Xu_input_file                : num 1004
 $ ERROR_STATUS_unknown_spp_occ_FP_error_type                 : num 1005
 $ ERROR_STATUS_unknown_spp_occ_FN_error_type                 : num 1006
}}
\subsection{bpm}{
\preformatted{
bpm :  num [1:1277, 1:407] 1 0 0 0 0 0 0 0 0 0 ...
}}
\subsection{cur_PU_col}{
\preformatted{
cur_PU_col :  int 407
}}
\subsection{cur_spp_row}{
\preformatted{
cur_spp_row :  int 1277
}}
\subsection{FN_rates}{
\preformatted{
FN_rates :  num [1:1277, 1:407] 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 ...
}}
\subsection{FP_rates}{
\preformatted{
FP_rates :  num [1:1277, 1:407] 0.000588 0.000588 0.000588 0.000588 0.000588 ...
}}
\subsection{num_PUs}{
\preformatted{
num_PUs :  num 407
}}
\subsection{num_spp}{
\preformatted{
num_spp :  int 1277
}}
\subsection{random_values}{
\preformatted{
random_values :  num [1:1277, 1:407] 0.133 0.307 0.519 0.987 0.753 ...
}}
}

