% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_error_to_spp_occupancy_data.R
\name{apply_error_to_spp_occupancy_data}
\alias{apply_error_to_spp_occupancy_data}
\title{Apply error to species occupancy data.}
\usage{
apply_error_to_spp_occupancy_data(bpm, FP_rates, FN_rates, num_PUs, num_spp,
  random_values, bdpg_error_codes)
}
\arguments{
\item{bpm}{bipartite matrix; integer matrix with one row for each species
and one column for each planning unit.  Each matrix entry specifies
whether that species occupies that planning unit; 1 indicates the
species does occupy the planning unit and 0 indicates it does not.
Same as occ_matrix.}

\item{FP_rates}{numeric vector}

\item{FN_rates}{numeric vector}

\item{num_PUs}{integer number of planning units}

\item{num_spp}{integer number of species in the problem}

\item{random_values}{numeric vector}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns bpm matrix
}
\description{
Walk through the occupancy matrix (PU vs spp) and randomly
 choose to flip some of the TPs to FNs and TNs to FPs based on the
 given FP and FN error rates.
 Update the occupancy matrix (bpm) as you go along, but don't update
 the PU_spp_pair_indices until you know all locations that have
 flipped.  Update PU_spp_pair_indices at the end so that you don't
 have to constantly resize this big array as you go along.
}
