% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_9a_create_Xu_graph.R
\name{assert_edge_list_does_not_violate_assumptions}
\alias{assert_edge_list_does_not_violate_assumptions}
\title{Assert edge list does not violate assumptions}
\usage{
assert_edge_list_does_not_violate_assumptions(
  edge_list,
  first_row_of_intergroup_links,
  nodes
)
}
\arguments{
\item{edge_list}{two column integer matrix of node IDs with one row for
each edge and columns for the 2 ends of the edge}

\item{first_row_of_intergroup_links}{integer row number of first row
containing intergroup links}

\item{nodes}{data frame containing integer node_ID, integer group_ID, and boolean dependent_set_member}
}
\value{
Returns a boolean
}
\description{
Assert edge list does not violate assumptions
}
