% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{build_PU_spp_pair_indices_from_occ_matrix}
\alias{build_PU_spp_pair_indices_from_occ_matrix}
\title{Build PU spp pair indices from occ matrix}
\usage{
build_PU_spp_pair_indices_from_occ_matrix(occ_matrix, num_PUs, num_spp)
}
\arguments{
\item{occ_matrix}{matrix}

\item{num_PUs}{integer}

\item{num_spp}{integer}
}
\value{
Returns PU_spp_pair_indices data frame
}
\description{
Build planning unit vs. species indices from occupancy matrix.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{cur_PU_col}{
\preformatted{
cur_PU_col :  int 407
}}
\subsection{cur_PU_spp_row_idx}{
\preformatted{
cur_PU_spp_row_idx :  num 3022
}}
\subsection{cur_spp_row}{
\preformatted{
cur_spp_row :  int 1277
}}
\subsection{num_PU_spp_pairs}{
\preformatted{
num_PU_spp_pairs :  num 3022
}}
\subsection{num_PUs}{
\preformatted{
num_PUs :  num 407
}}
\subsection{num_spp}{
\preformatted{
num_spp :  int 1277
}}
\subsection{occ_matrix}{
\preformatted{
occ_matrix :  num [1:1277, 1:407] 1 0 0 0 0 0 0 0 0 0 ...
}}
\subsection{PU_spp_pair_indices}{
\preformatted{
PU_spp_pair_indices : 'data.frame':	3022 obs. of  2 variables:
 $ PU_ID : int  1 3 5 6 7 8 9 10 209 335 ...
 $ spp_ID: int  1 2 3 3 4 4 5 5 5 5 ...
}}
}

