% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{build_topdir_name}
\alias{build_topdir_name}
\title{Build top directory name}
\usage{
build_topdir_name(obj)
}
\arguments{
\item{obj}{an object such as a Xu_bd_problem or an RSrun}
}
\value{
Returns a directory name (not a path)
}
\description{
Builds a directory name for the highest level directory for a bdpg-related
object such as a reserve selection problem.
}
\details{
#'@section Local Variable Structures and examples:
Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{obj}{
\preformatted{
obj : Formal class 'Xu_bd_problem' [package "bdpg"] with 35 slots
}}
\subsection{topdir_name}{
\preformatted{
topdir_name :  chr "RSprob-COR-Base.d0729e1c-eadc-4899-a382-8cb7ac2c08d7"
}}
}
