% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_11b_network_measures_using_igraph_package.R
\name{ccTopDot}
\alias{ccTopDot}
\title{ccTopDot}
\usage{
ccTopDot(bg)
}
\arguments{
\item{bg}{the graph}
}
\description{
The last clustering measures relies on the functions ccBip
written by Gabor Csardi, while ccLowDot and ccTopDot are
essentially the same function with only a minor change
}
