% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{combine_2_bdprobs}
\alias{combine_2_bdprobs}
\title{FUNCTION_TITLE}
\usage{
combine_2_bdprobs(bdprob_1, bdprob_2, parameters, bdpg_error_codes, integerize)
}
\arguments{
\item{bdprob_1}{a BD_Prob to be combined}

\item{bdprob_2}{a BD_Prob to be combined}

\item{parameters}{list of control parameters, generally from project.yaml}

\item{bdpg_error_codes}{list of error names and corresponding error codes}

\item{integerize}{function to use in converting floats to ints}
}
\description{
FUNCTION_DESCRIPTION
}
