% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{combine_2_bdprobs}
\alias{combine_2_bdprobs}
\title{Combine two problems}
\usage{
combine_2_bdprobs(bdprob_1, bdprob_2, parameters, integerize)
}
\arguments{
\item{bdprob_1}{a BD_Prob to be combined}

\item{bdprob_2}{a BD_Prob to be combined}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}
}
\description{
Combine two problems
}
