% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14a_compute_marxan_solution_scores.R
\name{compute_marxan_solution_scores}
\alias{compute_marxan_solution_scores}
\title{Compute marxan solution scores}
\usage{
compute_marxan_solution_scores(
  spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU,
  marxan_solution_PU_IDs,
  targets,
  num_spp,
  marxan_solutions_matrix,
  cur_solution_num,
  marxan_solution_scores,
  cor_PU_costs,
  total_landscape_cost
)
}
\arguments{
\item{spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU}{matrix}

\item{marxan_solution_PU_IDs}{numeric vector}

\item{targets}{numeric vector}

\item{num_spp}{integer}

\item{marxan_solutions_matrix}{matrix}

\item{cur_solution_num}{integer}

\item{marxan_solution_scores}{data frame}

\item{cor_PU_costs}{numeric vector}

\item{total_landscape_cost}{numeric}
}
\value{
Returns marxan_solution_scores data frame
}
\description{
Compute marxan solution scores
}
