% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_10b_compute_solution_rep_levels_and_costs.R
\name{compute_rep_fraction}
\alias{compute_rep_fraction}
\title{Compute species representation fractions achieved}
\usage{
compute_rep_fraction(spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU,
  PU_set_to_test, spp_rep_targets = 1)
}
\arguments{
\item{spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU}{integer matrix of
abundance of each species in each planning unit, with a row for each
species ID and a column for each planning unit ID}

\item{PU_set_to_test}{integer vector of planning unit IDs where each ID
specifies a planning unit to include in the set to be tested}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}
}
\value{
Returns spp_rep_fracs numeric vector of fractions of species
    representation targets achieved by a given set of planning units
}
\description{
Compute fractions of species representation targets achieved by a given
set of planning units.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{PU_set_to_test}{
\preformatted{
PU_set_to_test :  int [1:61] 2 4 6 8 10 12 14 16 18 20 ...
}}
\subsection{selected_PUs_matrix_of_spp_cts_per_PU}{
\preformatted{
selected_PUs_matrix_of_spp_cts_per_PU :  num [1:814, 1:61] 1 0 0 0 0 0 0 0 0 0 ...
}}
\subsection{spp_rep_cts}{
\preformatted{
spp_rep_cts :  num [1:814] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{spp_rep_fracs}{
\preformatted{
spp_rep_fracs :  num [1:814] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{spp_rep_targets}{
\preformatted{
spp_rep_targets :  num [1:814] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU}{
\preformatted{
spp_rows_by_PU_cols_matrix_of_spp_cts_per_PU :  num [1:814, 1:122] 1 0 0 0 0 0 0 0 0 0 ...
}}
}

