% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_10b_compute_solution_rep_levels_and_costs.R
\name{compute_solution_cost}
\alias{compute_solution_cost}
\title{Compute cost of given solution vector of PUs to include}
\usage{
compute_solution_cost(PU_set_to_test, PU_costs)
}
\arguments{
\item{PU_set_to_test}{integer vector of planning unit IDs where each ID
specifies a planning unit to include in the set to be tested}

\item{PU_costs}{numeric vector of planning unit costs}
}
\value{
Returns numeric cost of the given solution
}
\description{
Given a set of planning units to include in a solution, compute the total
cost for that set.
}
