% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob_APP.R
\name{create_and_init_APP_bdprob}
\alias{create_and_init_APP_bdprob}
\title{Create and initialize apparent bdproblem from a correct problem.}
\usage{
create_and_init_APP_bdprob(Xu_bdprob_COR, parameters, compound_err_name)
}
\arguments{
\item{Xu_bdprob_COR}{a Xu_bd_problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{compound_err_name}{string containing name of compound error condition}
}
\value{
an apparent Xu_bd_problem
}
\description{
All values that are altered for the apparent problem should be assumed to
 be altered in the APP_prob_info_class object in the problem's APP_prob_info
 slot.
 Values in the main part of the Xu_bdprob_APP object should be the same
 as the values for the underlying COR problem that the APP problem is
 adding error to, with the exception of a few object-specific things
 like the UUID of the APP problem itself.
}
