% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob_APP.R
\name{create_and_init_APP_bdprob}
\alias{create_and_init_APP_bdprob}
\title{Create and initialize apparent bdproblem from a correct problem.}
\usage{
create_and_init_APP_bdprob(Xu_bdprob_COR)
}
\description{
Create and initialize apparent bdproblem from a correct problem.
}
