% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{create_eligible_PU_set}
\alias{create_eligible_PU_set}
\title{Create set of planning units eligible for adding species into}
\usage{
create_eligible_PU_set(Xu_dep_set, extra_PUs, dep_set_PUs_eligible)
}
\arguments{
\item{Xu_dep_set}{integer vector of planning unit IDs for the solution set
of the original Xu problem being wrapped around}

\item{extra_PUs}{integer vector of plannning unit IDs being added to the
original Xu problem to make the wrapped problem}

\item{dep_set_PUs_eligible}{boolean indicating whether new species instances
can be added to the original Xu problem's solution set in addition to
the adding them to the newly added planning units; TRUE implies new
species instances can be added to both sets; FALSE implies new species
instances can only be added to the newly added planning units}
}
\value{
set of planning unit IDs where new species can be placed
}
\description{
When wrapping a species distribution around a Xu biodiversity problem,
the first instance of any new species is added to the solution set of the
original Xu problem that is being wrapped.  However, if more than one
instance of the new species is being added in the wrapped problem, the
extra instances could go in a) just the new planning units outside the
original Xu problem added for the wrapper, or b) in both the original
solution set and the newly added planning units.  This function builds
that target set of planning units based on the value of the
dep_set_PUs_eligible flag passed to it.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{dep_set_PUs_eligible}{
\preformatted{
dep_set_PUs_eligible :  logi FALSE
}}
\subsection{eligible_PUs}{
\preformatted{
eligible_PUs :  int [1:285] 123 124 125 126 127 128 129 130 131 132 ...
}}
\subsection{extra_PUs}{
\preformatted{
extra_PUs :  int [1:285] 123 124 125 126 127 128 129 130 131 132 ...
}}
\subsection{Xu_dep_set}{
\preformatted{
Xu_dep_set :  int [1:61] 2 4 6 8 10 12 14 16 18 20 ...
}}
}

