% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{create_eligible_PU_set}
\alias{create_eligible_PU_set}
\title{FUNCTION_TITLE}
\usage{
create_eligible_PU_set(Xu_dep_set, extra_PUs, dep_set_PUs_eligible)
}
\arguments{
\item{Xu_dep_set}{DESCRIPTION.}

\item{extra_PUs}{DESCRIPTION.}

\item{dep_set_PUs_eligible}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

