% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_15_create_master_output_structure.R
\name{create_master_output_structure}
\alias{create_master_output_structure}
\title{Create a master output structure for all problem attributes and reserve
selection results.}
\usage{
create_master_output_structure(COR_bd_prob, APP_bd_prob, apply_error,
  parameters, marxan_output_values, marxan_control_values)
}
\description{
Create a master output structure for all problem attributes and reserve
selection results.
}

