% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_6_create_data_structures.R
\name{create_nodes_data_structure}
\alias{create_nodes_data_structure}
\title{Create nodes data structure}
\usage{
create_nodes_data_structure(
  tot_num_nodes,
  num_nodes_per_group,
  n__num_groups,
  num_independent_nodes_per_group
)
}
\arguments{
\item{tot_num_nodes}{integer}

\item{num_nodes_per_group}{integer}

\item{n__num_groups}{integer}

\item{num_independent_nodes_per_group}{integer}
}
\value{
data frame containing node_ID, group_ID, dependent_set_member
}
\description{
Create nodes data structure
}
