% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_RSprob_dir_and_subdirs}
\alias{create_RSprob_dir_and_subdirs}
\title{Create directories for an RS problem}
\usage{
create_RSprob_dir_and_subdirs(top_dir, rsprob)
}
\arguments{
\item{top_dir}{character string giving the path to the top directory of the
tree where the problem is written out, usually something like the tzar
output directory of the experiment followed by some object-specific
information and the UUID of the object, e.g.,
~/tzarout/RSprob-COR-Base.299140ae-3be2-4331-bb72-b24570902587}

\item{rsprob}{a reserve selection problem object, e.g., a Xu_bd_problem}
}
\value{
Doesn't return anything.
}
\description{
Create directory and subdirectory structure for an RS problem inside an
experiment.  Creates the highest level directory for the problem and the
plots and networks directories under that directory.
}
