% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSrun.R
\name{create_RSrun}
\alias{create_RSrun}
\title{Create an RSrun}
\usage{
create_RSrun(
  prob_UUID,
  spp_rep_targets,
  parameters,
  starting_dir,
  cor_or_app_str,
  basic_or_wrapped_or_comb_str,
  rs_method_name
)
}
\arguments{
\item{prob_UUID}{UUID for the biodiversity problem the reserve selector is
run over}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{starting_dir}{string containing full path of top directory where
problems and/or runs are to be created}

\item{cor_or_app_str}{character string}

\item{basic_or_wrapped_or_comb_str}{character string}

\item{rs_method_name}{character string for reserve selection method, e.g.,
"marxan_sa"}
}
\value{
Returns an RSrun object
}
\description{
Create a run of a reserve selector
}
