% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_marxan_analysis_and_output.R
\name{create_RSrun}
\alias{create_RSrun}
\title{Create an RSrun}
\usage{
create_RSrun(prob_UUID, targets, parameters, cor_or_app_str,
  basic_or_wrapped_or_comb_str, method_name)
}
\arguments{
\item{prob_UUID}{UUID for the biodiversity problem the reserve selector is
run over}

\item{targets}{numeric vector}

\item{cor_or_app_str}{character string}

\item{basic_or_wrapped_or_comb_str}{character string}

\item{method_name}{character string for reserve selection method, e.g.,
"marxan_sa"}

\item{starting_dir}{character string}
}
\value{
Returns an RSrun object
}
\description{
Create a run of a reserve selector
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{basic_or_wrapped_or_comb_str}{
\preformatted{
basic_or_wrapped_or_comb_str :  chr "Wrap"
}}
\subsection{cor_or_app_str}{
\preformatted{
cor_or_app_str :  chr "COR"
}}
\subsection{method_name}{
\preformatted{
method_name :  chr "Marxan_SA"
}}
\subsection{prob_UUID}{
\preformatted{
prob_UUID :  chr "9b4b5b43-8be8-4e9c-bd7d-d1938d9921bd"
}}
\subsection{rsrun}{
\preformatted{
rsrun : Formal class 'RSrun' [package "bdpg"] with 12 slots
}}
\subsection{starting_dir}{
\preformatted{
starting_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{targets}{
\preformatted{
targets :  num [1:1277] 1 1 1 1 1 1 1 1 1 1 ...
}}
}

