% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSrun.R
\name{create_RSrun_dir_and_subdirs}
\alias{create_RSrun_dir_and_subdirs}
\title{Create directories for an RS run}
\usage{
create_RSrun_dir_and_subdirs(rsrun, exp_root_dir)
}
\arguments{
\item{rsrun}{an RSrun (reserve selection run) object, e.g., a marxan run}

\item{exp_root_dir}{character string path to root directory of an
experiment, usually parameters$fullOutputDir_NO_slash}
}
\value{
Doesn't return anything.
}
\description{
Create directory and subdirectory structure for an RS run inside an
experiment.  Creates the highest level directory for the run and the
input, output, and plots directories under that directory.
}
