% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_9a_create_Xu_graph.R
\name{create_Xu_graph}
\alias{create_Xu_graph}
\title{Create Xu graph}
\usage{
create_Xu_graph(
  num_nodes_per_group,
  n__num_groups,
  nodes,
  max_possible_tot_num_links,
  target_num_links_between_2_groups_per_round,
  num_rounds_of_linking_between_groups,
  duplicate_links_allowed = FALSE
)
}
\arguments{
\item{num_nodes_per_group}{integer number of planning units per group in the
Xu problem}

\item{n__num_groups}{integer number of groups in the Xu problem, one of the
4 base input parameters for specifying the Xu problem}

\item{nodes}{data frame containing integer node_ID, integer group_ID, and boolean dependent_set_member}

\item{max_possible_tot_num_links}{integer maximum possible number of links
in the Xu problem given all the other parameter settings}

\item{target_num_links_between_2_groups_per_round}{integer target number of
links between any two groups per round of linking in constructing the
Xu problem}

\item{num_rounds_of_linking_between_groups}{integer number of rounds of
linking to be done between groups in constructing the Xu problems}

\item{duplicate_links_allowed}{boolean indicating whether more than one link
is allowed between the same two planning units, i.e., whether more than
one species can occupy exactly the same pair of planning units; TRUE
implies more than one allowed, FALSE implies only one allowed}
}
\value{
Returns edge list, a two column integer matrix of
    node IDs with one row for each edge and columns for the 2 ends of
    the edge
}
\description{
Given a set of controlling parameters derived from the initial 4 input
parameters, this function the creates an edge list for a Xu graph using the
algorithm described by Xu et al.  [Need citations here.]
}
