% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob.R
\name{create_Xu_problem_from_scratch}
\alias{create_Xu_problem_from_scratch}
\title{FUNCTION_TITLE}
\usage{
create_Xu_problem_from_scratch(max_allowed_num_spp, parameters, emulatingTzar,
  bdpg_error_codes, integerize, DEBUG_LEVEL)
}
\arguments{
\item{max_allowed_num_spp}{DESCRIPTION.}

\item{parameters}{DESCRIPTION.}

\item{emulatingTzar}{DESCRIPTION.}

\item{bdpg_error_codes}{DESCRIPTION.}

\item{integerize}{DESCRIPTION.}

\item{DEBUG_LEVEL}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

