% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_5_derive_control_parameters.R
\name{derive_Xu_control_parameters}
\alias{derive_Xu_control_parameters}
\title{Derive full Xu control params from 4 base params}
\usage{
derive_Xu_control_parameters(parameters, bdpg_error_codes, integerize)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{bdpg_error_codes}{list of error names and corresponding error codes}

\item{integerize}{function to use in converting floats to ints}
}
\value{
a Xu_parameters object
}
\description{
Derive full Xu control params from 4 base params
}
