% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_vars_in_this_func.R
\name{doc_vars_in_this_func}
\alias{doc_vars_in_this_func}
\title{Document variables and their types that are visibile inside current function}
\usage{
doc_vars_in_this_func(sys.call_ht = -1, sys.frame_ht = -1)
}
\arguments{
\item{sys.call_ht}{negative integer to pass to sys.call() to specify how
high to look on the \code{\link{sys.call}} stack.}

\item{sys.frame_ht}{negative integer to pass to sys.frame() to specify how
    high to look on the \code{\link{sys.frame}} stack.

Generally shouldn't need to set either of the arguments since the value
either defaults to the correct value or is set correctly inside
\code{\link{doc_vars_in_this_func_once}}.}
}
\value{
Returns nothing
}
\description{
Call the str() function for all variables visible inside the current
function.  Before doing that, write out a flag line to indicate the start
of a section of output for a particular function and write the call made
to that function, i.e., function name followed by its argument list.
After writing out the structures of all the variables write out a
closing line to flag the end of str() output for the current function.
}
\seealso{
\code{\link{doc_vars_in_this_func_once}}
}
