% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lognormal_overlay.R
\name{EF}
\alias{EF}
\title{Evaluation Function to be passed to optim()}
\usage{
EF(seed_value, num_spp_to_generate, min_num_spp_on_2_PUs, max_max_abundance,
  target_max_abundance_ct, add_one_to_abundances, outfile, mean_sd_pair)
}
\arguments{
\item{seed_value}{integer}

\item{num_spp_to_generate}{integer}

\item{min_num_spp_on_2_PUs}{integer}

\item{max_max_abundance}{integer}

\item{target_max_abundance_ct}{integer}

\item{add_one_to_abundances}{boolean}

\item{outfile}{character string}

\item{mean_sd_pair}{2 element numeric vector containing the mean and
standard deviation to use for the lognormal}
}
\value{
Returns numeric score from evaluating the given lognormal; the
    closer to 0 the better
}
\description{
Note that the "Local Variable Structures and examples" section below was
not labelled as belonging to a function "EF".  It was labelled as belonging
to a function "fn".  I'm assuming that the variables actually belong to EF
because "fn" is the argument name that EF gets assigned to when passed to
optim().
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{abundance_data}{
\preformatted{
abundance_data : List of 3
 $ rounded_abundances          : num [1:1628] 3 3 2 2 3 3 3 3 3 3 ...
 $ num_spp_on_exactly_2_patches: int 173
 $ max_abundance_ct            : num 4
}}
\subsection{add_one_to_abundances}{
\preformatted{
add_one_to_abundances :  logi FALSE
}}
\subsection{double_diff}{
\preformatted{
double_diff :  num 1.07e+14
}}
\subsection{double_diff_penalty}{
\preformatted{
double_diff_penalty :  num 1.07e+14
}}
\subsection{double_diff_penalty_exponent}{
\preformatted{
double_diff_penalty_exponent :  num 5
}}
\subsection{double_frac}{
\preformatted{
double_frac :  num 1.32e+11
}}
\subsection{EF_num}{
\preformatted{
EF_num :  num 1
}}
\subsection{max_abundance_ct}{
\preformatted{
max_abundance_ct :  num 4
}}
\subsection{max_abundance_diff}{
\preformatted{
max_abundance_diff :  num 138
}}
\subsection{max_abundance_frac}{
\preformatted{
max_abundance_frac :  num 0.972
}}
\subsection{max_abundance_penalty_exponent}{
\preformatted{
max_abundance_penalty_exponent :  num 2
}}
\subsection{max_max_abundance}{
\preformatted{
max_max_abundance :  num 203
}}
\subsection{mean_sd_pair}{
\preformatted{
mean_sd_pair :  num [1:2] 1.04 0.1
}}
\subsection{min_num_spp_on_2_PUs}{
\preformatted{
min_num_spp_on_2_PUs :  int 814
}}
\subsection{num_spp_on_exactly_2_patches}{
\preformatted{
num_spp_on_exactly_2_patches :  int 173
}}
\subsection{num_spp_to_generate}{
\preformatted{
num_spp_to_generate :  num 1628
}}
\subsection{out_record}{
\preformatted{
out_record :  chr "1,131909314986.928,1.03988718083831,0.1,131909314986.928,0.971830985915493,814,142,203,11,1628,FALSE"
}}
\subsection{outfile}{
\preformatted{
outfile :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/wrap_search_outfile.csv"
}}
\subsection{score}{
\preformatted{
score :  num 1.32e+11
}}
\subsection{seed_value}{
\preformatted{
seed_value :  num 11
}}
\subsection{target_max_abundance_ct}{
\preformatted{
target_max_abundance_ct :  num 142
}}
\subsection{warn_on_constraint_violation}{
\preformatted{
warn_on_constraint_violation :  logi TRUE
}}
}

