% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14a_compute_marxan_solution_scores.R
\name{find_best_marxan_solutions}
\alias{find_best_marxan_solutions}
\title{Find best marxan solutions}
\usage{
find_best_marxan_solutions(marxan_output_dir_path, num_spp, cor_PU_costs,
  cor_bpm, app_bpm, marxan_best_df_sorted_as_vector, plot_output_dir,
  largest_PU_ID, largest_spp_ID, targets, marxan_top_dir)
}
\arguments{
\item{marxan_output_dir_path}{character string}

\item{num_spp}{integer}

\item{cor_PU_costs}{numeric vector}

\item{cor_bpm}{matrix}

\item{app_bpm}{matrix}

\item{marxan_best_df_sorted_as_vector}{data frame}

\item{plot_output_dir}{character string}

\item{largest_PU_ID}{integer}

\item{largest_spp_ID}{integer}

\item{targets}{numeric vector}

\item{marxan_top_dir}{character string}
}
\value{
Returns nothing
}
\description{
Find best marxan soltuions
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{app_bpm}{
\preformatted{
app_bpm :  num [1:1277, 1:407] 1 0 0 0 0 0 0 0 0 0 ...
}}
\subsection{app_marxan_solution_scores}{
\preformatted{
app_marxan_solution_scores : 'data.frame':	4 obs. of  3 variables:
 $ solution_num  : int  1 2 3 4
 $ representation: num  1 1 1 1
 $ cost          : num  0.15 0.15 0.15 0.15
}}
\subsection{best_solution_ID_according_to_marxan}{
\preformatted{
best_solution_ID_according_to_marxan :  int 1
}}
\subsection{cor_bpm}{
\preformatted{
cor_bpm :  num [1:1277, 1:407] 1 0 0 0 0 0 0 0 0 0 ...
}}
\subsection{cor_marxan_solution_scores}{
\preformatted{
cor_marxan_solution_scores : 'data.frame':	4 obs. of  3 variables:
 $ solution_num  : int  1 2 3 4
 $ representation: num  1 1 1 1
 $ cost          : num  0.15 0.15 0.15 0.15
}}
\subsection{cor_PU_costs}{
\preformatted{
cor_PU_costs :  num [1:407] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{cur_col}{
\preformatted{
cur_col :  int 4
}}
\subsection{cur_dist_from_marxan_best_df_sorted_as_vector}{
\preformatted{
cur_dist_from_marxan_best_df_sorted_as_vector :  num 0
}}
\subsection{cur_marxan_solution_PU_IDs}{
\preformatted{
cur_marxan_solution_PU_IDs :  int [1:61] 2 4 6 8 10 12 14 16 18 20 ...
}}
\subsection{cur_row}{
\preformatted{
cur_row :  int 4
}}
\subsection{cur_solution_num}{
\preformatted{
cur_solution_num :  int 4
}}
\subsection{distances_between_marxan_solutions}{
\preformatted{
distances_between_marxan_solutions :  num [1:4, 1:4] 0 0 0 0 0 0 0 0 0 0 ...
}}
\subsection{IDs_of_vectors_matching_marxan_best_solution_choice}{
\preformatted{
IDs_of_vectors_matching_marxan_best_solution_choice :  int [1:4] 1 2 3 4
}}
\subsection{largest_PU_ID}{
\preformatted{
largest_PU_ID :  num 407
}}
\subsection{largest_spp_ID}{
\preformatted{
largest_spp_ID :  int 1277
}}
\subsection{marxan_best_df_sorted_as_vector}{
\preformatted{
marxan_best_df_sorted_as_vector :  int [1:407] 0 1 0 1 0 1 0 1 0 1 ...
}}
\subsection{marxan_output_dir_path}{
\preformatted{
marxan_output_dir_path :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSrun_-COR-Wrap-Marxan_SA.9"| __truncated__
}}
\subsection{marxan_solutions_matrix}{
\preformatted{
marxan_solutions_matrix :  num [1:4, 1:407] 0 0 0 0 1 1 1 1 0 0 ...
}}
\subsection{marxan_solutions_matrix_and_num_solutions}{
\preformatted{
marxan_solutions_matrix_and_num_solutions : List of 2
 $ marxan_solutions_matrix: num [1:4, 1:407] 0 0 0 0 1 1 1 1 0 0 ...
 $ num_marxan_solutions   : int 4
}}
\subsection{num_marxan_solutions}{
\preformatted{
num_marxan_solutions :  int 4
}}
\subsection{num_spp}{
\preformatted{
num_spp :  int 1277
}}
\subsection{plot_output_dir}{
\preformatted{
plot_output_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSrun_-COR-Wrap-Marxan_SA.9"| __truncated__
}}
\subsection{short_range}{
\preformatted{
short_range :  num 4
}}
\subsection{targets}{
\preformatted{
targets :  num [1:1277] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{total_landscape_cost}{
\preformatted{
total_landscape_cost :  num 407
}}
}

