% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lognormal_overlay.R
\name{find_lognormal_to_wrap_around_Xu}
\alias{find_lognormal_to_wrap_around_Xu}
\title{Find lognormal to wrap around Xu problem}
\usage{
find_lognormal_to_wrap_around_Xu(
  Xu_bdprob,
  parameters,
  desired_Xu_spp_frac_of_all_spp,
  solution_frac_of_landscape,
  desired_max_abundance_frac,
  seed_value_for_search,
  max_search_iterations,
  add_one_to_lognormal_abundances,
  search_outfile_name
)
}
\arguments{
\item{Xu_bdprob}{a Xu_bd_problem object}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{desired_Xu_spp_frac_of_all_spp}{numeric}

\item{solution_frac_of_landscape}{numeric}

\item{desired_max_abundance_frac}{numeric}

\item{seed_value_for_search}{integer}

\item{max_search_iterations}{integer}

\item{add_one_to_lognormal_abundances}{boolean}

\item{search_outfile_name}{character string}
}
\value{
Returns numeric vector of abundances
}
\description{
Uses optimizer to search for parameters of a lognormal distribution that
will have at least as many species occurring on exactly 2 patches as the
given Xu problem and meet the other curve shape criteria given in the
argument list.
}
