% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bdprob.R
\name{gen_bdprob}
\alias{gen_bdprob}
\title{Generate a biodiversity problem}
\usage{
gen_bdprob(parameters, bdpg_error_codes, integerize)
}
\arguments{
\item{parameters}{parameter list, generally from project.yaml}

\item{bdpg_error_codes}{list of error codes and corresponding error numbers}

\item{integerize}{function to use in converting floats to integers}
}
\value{
biodiversity problems
}
\description{
Generate a biodiversity problem
}

