% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bdprob.R
\name{gen_bdprob}
\alias{gen_bdprob}
\title{FUNCTION_TITLE}
\usage{
gen_bdprob(parameters, bdpg_error_codes, integerize, DEBUG_LEVEL)
}
\arguments{
\item{parameters}{DESCRIPTION.}

\item{bdpg_error_codes}{DESCRIPTION.}

\item{integerize}{DESCRIPTION.}

\item{DEBUG_LEVEL}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

