% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bdprob.R
\name{gen_bdprob}
\alias{gen_bdprob}
\title{Generate a biodiversity problem}
\usage{
gen_bdprob(parameters, compute_network_metrics_for_this_prob, bdpg_error_codes,
  integerize, base_bdprob = NULL)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{compute_network_metrics_for_this_prob}{boolean indicating whether to
compute network metrics for the current problem; TRUE indicates to
compute one or more network metrics specified elsewhere and FALSE
indicate not compute any network metrics for the current problem.}

\item{bdpg_error_codes}{list of error names and corresponding error codes}

\item{integerize}{function to use in converting floats to ints}

\item{base_bdprob}{NULL or an existing bdprob to use in creating multi_bdproblem}
}
\value{
biodiversity problems
}
\description{
Generate a biodiversity problem.
}
