% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bdprob.R
\name{gen_multi_bdprob}
\alias{gen_multi_bdprob}
\title{Generate a new bd problem by modifying or combining existing problem(s)}
\usage{
gen_multi_bdprob(parameters, integerize, bdprob_1 = NULL)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{integerize}{function to use in converting floats to ints}

\item{bdprob_1}{a Xu biodiversity problem object}
}
\value{
Returns a multi-BD_Prob
}
\description{
Sometimes it may be useful to combine 2 or more bd problems into
 one bigger problem since it may have the potential to produce a
 more difficult compound problem.
 If nothing else, it bears some resemblance to having a larger
 landscape with subregion characteristics.
}
\details{
This function is intended to allow either wrapping a distribution around
 an existing problem or combining two problems.  However, at the moment,
 it only allows wrapping.  Combining two problems is pretty straightforward
 but has not been implemented yet and probably won't be implemented unless
 there is a demand for it.

Dummy code that was in here to demonstrate high-level parts of that
 combination (i.e., combine_2_bdprobs()) have been removed but can be
 found in github versions of the code up until around commit 1c0fbba6
 on Feb 4, 2017.
}
