% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{gen_multi_bdprob}
\alias{gen_multi_bdprob}
\title{FUNCTION_TITLE}
\usage{
gen_multi_bdprob(parameters, wrap_lognormal_dist_around_Xu,
  read_Xu_problem_from_Xu_file, infile_name, given_correct_solution_cost,
  max_allowed_num_spp, bdpg_error_codes, integerize)
}
\arguments{
\item{wrap_lognormal_dist_around_Xu}{DESCRIPTION.}

\item{read_Xu_problem_from_Xu_file}{DESCRIPTION.}

\item{infile_name}{DESCRIPTION.}

\item{given_correct_solution_cost}{DESCRIPTION.}

\item{max_allowed_num_spp}{DESCRIPTION.}

\item{bdpg_error_codes}{DESCRIPTION.}

\item{integerize}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}

