% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lognormal_overlay.R
\name{gen_rounded_abundances}
\alias{gen_rounded_abundances}
\title{Generate rounded abundances}
\usage{
gen_rounded_abundances(num_spp_to_generate, meanlog, sdlog,
  add_one_to_abundances, plot_rounded_abundances = FALSE)
}
\arguments{
\item{num_spp_to_generate}{integer}

\item{meanlog}{numeric}

\item{sdlog}{numeric}

\item{add_one_to_abundances}{boolean}

\item{plot_rounded_abundances}{boolean}
}
\value{
Returns a vector with one count for each species
}
\description{
Function to generate a sampled lognormal distribution whose values are
 rounded to produce a vector of counts of the number of PUs that each
 species will occupy.
}
\details{
No zero counts are allowed, so any species that was assigned a 0 count in
 the lognormal generation is removed from the returned vector.
 This means that the number of species returned may be less than the value
 of num_spp_to_generate specified in the inputs.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{abundance_data}{
\preformatted{
abundance_data : List of 3
 $ rounded_abundances          : num [1:1628] 3 3 2 2 3 3 3 3 3 3 ...
 $ num_spp_on_exactly_2_patches: int 173
 $ max_abundance_ct            : num 4
}}
\subsection{abundances}{
\preformatted{
abundances :  num [1:1628] 2.67 2.84 2.43 2.47 3.18 ...
}}
\subsection{add_one_to_abundances}{
\preformatted{
add_one_to_abundances :  logi FALSE
}}
\subsection{max_abundance_ct}{
\preformatted{
max_abundance_ct :  num 4
}}
\subsection{meanlog}{
\preformatted{
meanlog :  num 1.04
}}
\subsection{num_spp_on_exactly_2_patches}{
\preformatted{
num_spp_on_exactly_2_patches :  int 173
}}
\subsection{num_spp_to_generate}{
\preformatted{
num_spp_to_generate :  num 1628
}}
\subsection{plot_rounded_abundances}{
\preformatted{
plot_rounded_abundances :  logi FALSE
}}
\subsection{rounded_abundances}{
\preformatted{
rounded_abundances :  num [1:1628] 3 3 2 2 3 3 3 3 3 3 ...
}}
\subsection{sdlog}{
\preformatted{
sdlog :  num 0.1
}}
}

