% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob_APP.R
\name{gen_single_bdprob_APP}
\alias{gen_single_bdprob_APP}
\title{Generate a single biodiversity problem with error added to it}
\usage{
gen_single_bdprob_APP(Xu_bdprob_COR, parameters, starting_dir,
  gen_cost_errors, gen_FP_FN_errors, compound_err_name = NULL,
  ret_vals_from_build_const_err = NULL,
  ret_vals_from_apply_cost_errors = NULL)
}
\arguments{
\item{Xu_bdprob_COR}{correct Xu problem that is to have error added to it}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{starting_dir}{string containing full path of top directory where
problems and/or runs are to be created}

\item{gen_cost_errors}{boolean flag indicating whether to add cost error;
TRUE implies add cost error and FALSE implies don't add cost error}

\item{gen_FP_FN_errors}{boolean flag indicating whether to add false positives
and/or false negatives;
TRUE implies add FP/FN error and FALSE implies don't add FP/FN error}

\item{compound_err_name}{string containing name of compound error condition}

\item{ret_vals_from_build_const_err}{list of return values from building
constant error values if already done, NULL if not already done}

\item{ret_vals_from_apply_cost_errors}{list of return values from building
cost error values if already done, NULL if not already done}
}
\value{
Returns apparent version of either a Xu_bd_problem or a
    Xu_wrapped_bd_problem
}
\description{
All values that are altered for the apparent problem should be assumed to
 be altered in the APP_prob_info_class object in the problem's APP_prob_info
 slot.
 Values in the main part of the Xu_bdprob_APP object should be the same
 as the values for the underlying COR problem that the APP problem is
 adding error to, with the exception of a few object-specific things
 like the UUID of the APP problem itself.
}
