% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_error_to_spp_occupancy_data.R
\name{gen_single_bdprob_APP}
\alias{gen_single_bdprob_APP}
\title{Generate a single biodiversity problem with error added to it}
\usage{
gen_single_bdprob_APP(Xu_bdprob_COR, starting_dir, parameters, bdpg_error_codes,
  integerize)
}
\arguments{
\item{Xu_bdprob_COR}{correct Xu problem that is to have error added to it}

\item{parameters}{named list of all parameters, generally from project.yaml file}

\item{bdpg_error_codes}{DESCRIPTION.}

\item{integerize}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Create an apparent problem from a correct problem.
}

