% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob.R
\name{gen_single_bdprob_COR}
\alias{gen_single_bdprob_COR}
\title{Generate a single biodiversity problem from scratch of from a file}
\usage{
gen_single_bdprob_COR(COR_prob_src = NULL, R_obj_dump_file_path = NULL,
  exp_root_dir = NULL, compute_network_metrics_for_this_prob = NULL,
  parameters = NULL, read_Xu_problem_from_Xu_file = NULL,
  Xu_bench_infile_name = NULL, given_correct_solution_cost = NULL,
  max_allowed_num_spp = NULL, bdpg_error_codes = NULL, integerize = NULL,
  base_prob_name_stem = "base_prob", cor_dir_name_stem = "cor")
}
\arguments{
\item{COR_prob_src}{string indicating whether this COR problem is
generated from scratch ("generator"), read from a saved R object
("R_obj_dump_file"), or read from a Xu benchmark file ("Xu_bench_file")}

\item{R_obj_dump_file_path}{string containing full path (including
file name) to the file containing an R object that was saved}

\item{exp_root_dir}{character string path to root directory of an
experiment, usually parameters$fullOutputDir_NO_slash}

\item{compute_network_metrics_for_this_prob}{boolean indicating whether to
compute network metrics for the current problem; TRUE indicates to
compute one or more network metrics specified elsewhere and FALSE
indicate not compute any network metrics for the current problem.}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{read_Xu_problem_from_Xu_file}{boolean TRUE if problem is to be
read from a benchmark problem description file created by Xu;
FALSE if problem is to be created from scratch.}

\item{Xu_bench_infile_name}{filename of Xu benchmark problem file}

\item{given_correct_solution_cost}{boolean indicating whether the correct
optimal solution cost is given for a reserve selection problem}

\item{max_allowed_num_spp}{integer maximum number of species allowed in the
problem (to keep from generating problems that are too large when trying
to keep the run-time down)}

\item{bdpg_error_codes}{list of error names and corresponding error codes}

\item{integerize}{function to use in converting floats to ints}
}
\value{
Returns a COR Xu_bd_problem
}
\description{
Generate a single biodiversity problem from scratch of from a file
}
