% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_single_bdprob.R
\name{gen_single_bdprob_COR_from_scratch_or_Xu_bench_file}
\alias{gen_single_bdprob_COR_from_scratch_or_Xu_bench_file}
\title{Generate a single biodiversity problem}
\usage{
gen_single_bdprob_COR_from_scratch_or_Xu_bench_file(
  exp_root_dir,
  parameters,
  read_Xu_problem_from_Xu_file,
  Xu_bench_infile_name,
  given_correct_solution_cost,
  max_allowed_num_spp,
  integerize,
  base_prob_name_stem = "base_prob",
  cor_dir_name_stem = "cor"
)
}
\arguments{
\item{exp_root_dir}{character string path to root directory of an
experiment, usually parameters$fullOutputDir_NO_slash}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{read_Xu_problem_from_Xu_file}{boolean TRUE if problem is to be
read from a benchmark problem description file created by Xu;
FALSE if problem is to be created from scratch.}

\item{Xu_bench_infile_name}{string containing the name of the Xu benchmark file to
read a problem from if reading from a Xu benchmark file}

\item{given_correct_solution_cost}{boolean indicating whether the correct
optimal solution cost is given for a reserve selection problem}

\item{max_allowed_num_spp}{integer maximum number of species allowed in the
problem (to keep from generating problems that are too large when trying
to keep the run-time down)}

\item{integerize}{function to use in converting floats to ints}

\item{base_prob_name_stem}{character string}

\item{cor_dir_name_stem}{character string}
}
\value{
Returns a new Xu_bd_problem
}
\description{
Read a Xu problem from files of ones already created by Xu or
create one from scratch.
}
\details{
**NOTE that if other kinds of problems are created, this routine will have
to be replaced or cloned into something appropriate for the new problem
type.**
}
