% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_probs.R
\name{gen_single_bdprob_WRAP}
\alias{gen_single_bdprob_WRAP}
\title{Generate a single wrapped Xu biodiversity problem}
\usage{
gen_single_bdprob_WRAP(bdprob_to_wrap, parameters, bdpg_error_codes)
}
\arguments{
\item{bdprob_to_wrap}{a bdproblem to wrap a distribution around}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns a wrapped Xu biodiversity problem
}
\description{
Generate a single wrapped Xu biodiversity problem
}
