% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{gen_wrapped_bdprob_COR}
\alias{gen_wrapped_bdprob_COR}
\title{Generate COR wrapped bd problem}
\usage{
gen_wrapped_bdprob_COR(starting_dir, compute_network_metrics_for_this_prob,
  parameters, base_bdprob, bdpg_error_codes)
}
\arguments{
\item{starting_dir}{character string}

\item{compute_network_metrics_for_this_prob}{boolean indicating whether to
compute network metrics for the current problem; TRUE indicates to
compute one or more network metrics specified elsewhere and FALSE
indicate not compute any network metrics for the current problem.}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{base_bdprob}{Xu_bd_problem}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns wrapped_bdprob_COR
}
\description{
Generate COR wrapped bd problem
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{add_one_to_lognormal_abundances}{
\preformatted{
add_one_to_lognormal_abundances :  logi FALSE
}}
\subsection{base_bdprob}{
\preformatted{
base_bdprob : Formal class 'Xu_bd_problem' [package "bdpg"] with 35 slots
}}
\subsection{bdpg_error_codes}{
\preformatted{
bdpg_error_codes : List of 6
 $ ERROR_STATUS_num_inside_or_within_group_links_less_than_one: num 1001
 $ ERROR_STATUS_optimal_solution_is_not_optimal               : num 1002
 $ ERROR_STATUS_num_nodes_per_group_must_be_at_least_2        : num 1003
 $ ERROR_STATUS_duplicate_spp_in_Xu_input_file                : num 1004
 $ ERROR_STATUS_unknown_spp_occ_FP_error_type                 : num 1005
 $ ERROR_STATUS_unknown_spp_occ_FN_error_type                 : num 1006
}}
\subsection{dep_set_PUs_eligible}{
\preformatted{
dep_set_PUs_eligible :  logi FALSE
}}
\subsection{desired_max_abundance_frac}{
\preformatted{
desired_max_abundance_frac :  num 0.7
}}
\subsection{desired_Xu_spp_frac_of_all_spp}{
\preformatted{
desired_Xu_spp_frac_of_all_spp :  num 0.5
}}
\subsection{max_search_iterations}{
\preformatted{
max_search_iterations :  num 500
}}
\subsection{parameters}{
\preformatted{
parameters : List of 66
 $ summary_without_run_id_filename                           : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/prob_diff_results_with_0_ru"| __truncated__
 ...
 $ fullOutputDir_NO_slash                                    : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{rounded_abundances}{
\preformatted{
rounded_abundances :  num [1:1628] 2 2 1 1 3 1 3 2 2 1 ...
}}
\subsection{search_outfile_name}{
\preformatted{
search_outfile_name :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/wrap_search_outfile.csv"
}}
\subsection{seed_value_for_search}{
\preformatted{
seed_value_for_search :  num 11
}}
\subsection{solution_frac_of_landscape}{
\preformatted{
solution_frac_of_landscape :  num 0.3
}}
\subsection{starting_dir}{
\preformatted{
starting_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{tot_num_PUs_in_landscape}{
\preformatted{
tot_num_PUs_in_landscape :  num 407
}}
\subsection{wrapped_bdprob_COR}{
\preformatted{
wrapped_bdprob_COR : Formal class 'Xu_wrapped_bd_problem' [package "bdpg"] with 36 slots
}}
}

