% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_action_using_tzar_reps.R
\name{get_base_cor_bdprob_for_given_app_bdprob}
\alias{get_base_cor_bdprob_for_given_app_bdprob}
\title{Get the COR bdprob whose UUID is the base UUID for the given APP bdprob}
\usage{
get_base_cor_bdprob_for_given_app_bdprob(app_bdprob, parameters)
}
\arguments{
\item{app_bdprob}{an apparent Xu_bd_problem}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}
}
\value{
Returns the COR Xu_bd_problem that the apparent problem added error
to
Get COR problem that given APP problem is based on

 The parameters list needs to show locations for both the APP problems
 and the COR problems that they're based on, so the project.yaml file
 needs to have a section like this:

   run_rs_on_APP_prob: TRUE

   RS_app_input_prob_src: "rds_file_set_from_file"
   RS_app_input_rds_file_set_path: "/Users/bill/D/Projects/ProblemDifficulty/pkgs/bdpgxupaper/inst/extdata/input_files/Tzar_input_files/APP_rds_input_file_paths.txt"

   RS_cor_input_prob_src: "rds_file_set_from_file"
   RS_cor_input_rds_file_set_path: "/Users/bill/D/Projects/ProblemDifficulty/pkgs/bdpgxupaper/inst/extdata/input_files/Tzar_input_files/COR_rds_input_file_paths.txt"
}
\description{
Get the COR bdprob whose UUID is the base UUID for the given APP bdprob
}
