% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_initialization.R
\name{get_bdpg_error_codes}
\alias{get_bdpg_error_codes}
\title{Builds the error codes returned by errors in bdpg package}
\usage{
get_bdpg_error_codes()
}
\value{
Returns list of integer error codes
}
\description{
This function is used as a central way to get at the error codes and make
sure they're always the same, but may want to move them to the
Global_Constants structure instead...
}
\examples{
get_bdpg_error_codes ()
}
