% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14d_see_if_marxan_best_was_actually_best.R
\name{get_marxan_solution_choice_string}
\alias{get_marxan_solution_choice_string}
\title{Get marxan solution choice string}
\usage{
get_marxan_solution_choice_string(
  marxan_best_cost,
  marxan_best_rep,
  sorted_best_cost,
  sorted_best_rep
)
}
\arguments{
\item{marxan_best_cost}{numeric}

\item{marxan_best_rep}{numeric}

\item{sorted_best_cost}{numeric}

\item{sorted_best_rep}{numeric}
}
\value{
Returns character string
}
\description{
Get marxan solution choice string
}
