% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_action_using_tzar_reps.R
\name{get_UUID_from_file_path}
\alias{get_UUID_from_file_path}
\title{Get the probable UUID of an object from its file name}
\usage{
get_UUID_from_file_path(f)
}
\arguments{
\item{f}{rds file name (possibly full path) containing a UUID}
}
\value{
Returns UUID section of the given file name
}
\description{
Problems and reserve selections saved from bdpg in rds files are given a
file name with a specific syntax, i.e., "saved", object type, UUID, "rds".
For example,
    saved.RSprob-COR-Base.489a016d-14fd-40af-89bb-3e1edf38f14f.rds
This function decodes that file name and returns the UUID.
}
\examples{
 get_UUID_from_file_path ("saved.RSprob-COR-Base.489a016d-14fd-40af-89bb-3e1edf38f14f.rds")
 get_UUID_from_file_path ("/Users/bill/tzar/outputdata/bdpgxupaper_single_action_COR_prob/default_runset/1_marxan_simulated_annealing/RSprob-COR-Base.489a016d-14fd-40af-89bb-3e1edf38f14f/saved.RSprob-COR-Base.489a016d-14fd-40af-89bb-3e1edf38f14f.rds")
}
