% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_9_link_nodes_between_groups.R
\name{link_nodes_between_groups}
\alias{link_nodes_between_groups}
\title{Link nodes between groups}
\usage{
link_nodes_between_groups(
  target_num_links_between_2_groups_per_round,
  num_rounds_of_linking_between_groups,
  n__num_groups,
  nodes,
  edge_list,
  cur_row
)
}
\arguments{
\item{target_num_links_between_2_groups_per_round}{integer}

\item{num_rounds_of_linking_between_groups}{integer}

\item{n__num_groups}{integer}

\item{nodes}{data frame containing integer node_ID, integer group_ID, and boolean dependent_set_member}

\item{edge_list}{two column integer matrix of node IDs with one row for
each edge and columns for the 2 ends of the edge}

\item{cur_row}{integer}
}
\value{
Returns edge_list as two column integer matrix of node IDs with
    one row for each edge and columns for the 2 ends of the edge
}
\description{
After all groups and their within group links have been built,
 ready to start doing rounds of intergroup linking.
}
