% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_8_link_nodes_within_groups.R
\name{link_nodes_within_groups}
\alias{link_nodes_within_groups}
\title{Link nodes within groups}
\usage{
link_nodes_within_groups(num_nodes_per_group, n__num_groups, nodes, edge_list)
}
\arguments{
\item{num_nodes_per_group}{integer number of planning units per group in the
Xu problem}

\item{n__num_groups}{integer number of groups in the Xu problem, one of the
4 base input parameters for specifying the Xu problem}

\item{nodes}{data frame containing integer node_ID, integer group_ID, and boolean dependent_set_member}

\item{edge_list}{two column integer matrix of node IDs with one row for
each edge and columns for the 2 ends of the edge}
}
\value{
Returns list containing edge_list and cur_row
}
\description{
Create species in a Xu problem by creating links between nodes within SAME
group.  Each link will represent a species that appears on the planning
units at each end of the link and those planning units are in separate
groups.
}
