% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14_read_marxan_output_files.R
\name{load_marxan_best_df_from_file_and_sort_and_add_missing_PUs}
\alias{load_marxan_best_df_from_file_and_sort_and_add_missing_PUs}
\title{Load marxan best data frame from file, sort, add missing PUs}
\usage{

  load_marxan_best_df_from_file_and_sort_and_add_missing_PUs(marxan_output_dir_path,
  all_correct_node_IDs)
}
\arguments{
\item{marxan_output_dir_path}{character string}

\item{all_correct_node_IDs}{integer vector}
}
\value{
Returns marxan_best_df_sorted data frame
}
\description{
Load marxan best data from file and sort in increasing order by planning
unit ID.  Also add any missing planning units.
}
