% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14_read_marxan_output_files.R
\name{load_marxan_mvbest_df_from_file_and_sort_by_CF}
\alias{load_marxan_mvbest_df_from_file_and_sort_by_CF}
\title{Load marxan mvbest df from file and sort by CF}
\usage{
load_marxan_mvbest_df_from_file_and_sort_by_CF(marxan_output_dir_path)
}
\arguments{
\item{marxan_output_dir_path}{character string}
}
\value{
Returns marxan mvbest data frame
}
\description{
Load marxan's best solution data frame from a file and sort it by
conservation features (i.e., species).
}
