% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_error_to_spp_occupancy_data.R
\name{match_FP_and_FN_counts_to_smaller_of_the_two}
\alias{match_FP_and_FN_counts_to_smaller_of_the_two}
\title{Match FP and FN counts to smaller of the two}
\usage{
match_FP_and_FN_counts_to_smaller_of_the_two(num_TPs, num_TNs, FP_const_rate,
  FN_const_rate)
}
\arguments{
\item{num_TPs}{integer}

\item{num_TNs}{integer}

\item{FP_const_rate}{numeric}

\item{FN_const_rate}{numeric}
}
\value{
Returns FP_FN_const_rate_pair list
}
\description{
Usually, the number of TNs and TPs will be unbalanced.
 Therefore, there will be far more opportunities to inject
 FPs than FNs or vice versa.
 Consequently, even if the FP and FN rates are set to the
 same value, there are likely to be far more FPs than FNs
 or vice versa in the apparent matrix.
 If you want to keep the opportunities for each of them
 to be more balanced, then you can multiply the dominant
 one by the lesser one's fraction of occurrence and
 reset the rate for the dominant so that both end up
 with the same counts.
}
\details{
Example: if there are
      - 100 entries total
      - 70 TNs
      - 30 TPs
 and you want 0.1 probability of FN, then you should get
 approximately 3 FNs.  If you want the FPs to match the FNs,
 then x * 70 FPs must equal 3 FPs too.
 So, the multiplier x = 3 / 70 ~ 0.0429
 i.e., the adjusted_P(FP) = num_FNs / num_TNs
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{approx_num_FNs}{
\preformatted{
approx_num_FNs :  num 304
}}
\subsection{approx_num_FPs}{
\preformatted{
approx_num_FPs :  num 304
}}
\subsection{FN_const_rate}{
\preformatted{
FN_const_rate :  num 0.1
}}
\subsection{FP_const_rate}{
\preformatted{
FP_const_rate :  num 0.000588
}}
\subsection{FP_FN_const_rate_pair}{
\preformatted{
FP_FN_const_rate_pair : List of 2
 $ FP_const_rate: num 0.000588
 $ FN_const_rate: num 0.1
}}
\subsection{num_TNs}{
\preformatted{
num_TNs :  num 516702
}}
\subsection{num_TPs}{
\preformatted{
num_TPs :  num 3037
}}
}

