% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{options_are_legal_for_single_bdprob_WRAP}
\alias{options_are_legal_for_single_bdprob_WRAP}
\title{Check whether options for wrapping have legal values}
\usage{
options_are_legal_for_single_bdprob_WRAP(bdprob_to_wrap, parameters,
  bdpg_error_codes)
}
\arguments{
\item{bdprob_to_wrap}{a bdproblem to wrap a distribution around}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
boolean with TRUE if options are legal and FALSE otherwise
}
\description{
Check whether options for wrapping have legal values
}
