% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_11_summarize_and_plot_graph_structure_information.R
\name{plot_final_degree_dist_for_node_graph}
\alias{plot_final_degree_dist_for_node_graph}
\title{Plot final degree distribution for node graph}
\usage{
plot_final_degree_dist_for_node_graph(final_link_counts_for_each_node,
  PU_col_name, plot_output_dir, cor_or_app_label)
}
\arguments{
\item{final_link_counts_for_each_node}{integer vector of counts of number
of links for each node, i.e., number of species for each planning unit}

\item{PU_col_name}{character string giving planning unit column name in data frames}

\item{plot_output_dir}{character string giving pat to directory where plot
should be written}

\item{cor_or_app_label}{character string for labelling as Correct or Apparent}
}
\value{
Doesn't return anything
}
\description{
Plot the final degree distribution for the node graph, which is essentially
the distribution of number of species per planning unit.
}
