% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_14c_plot_incremental_marxan_summed_solution_reps.R
\name{plot_incremental_marxan_summed_solution_representations}
\alias{plot_incremental_marxan_summed_solution_representations}
\title{Plot incremental marxan summed solution representations}
\usage{
plot_incremental_marxan_summed_solution_representations(marxan_ssoln_df,
  cor_PU_costs, optimum_cost, bpm, cor_app_prefix_string, num_spp,
  plot_output_dir)
}
\arguments{
\item{marxan_ssoln_df}{data frame}

\item{cor_PU_costs}{numeric vector}

\item{optimum_cost}{numeric}

\item{bpm}{matrix}

\item{cor_app_prefix_string}{character string}

\item{num_spp}{integer}

\item{plot_output_dir}{character string}
}
\value{
Returns nothing
}
\description{
For each step in order by Marxan summed solution PU ID:
 Want the fraction of all species who have met or exceeded their target
 when all PUs with the same number of votes or more are included in the
 solution.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{bpm}{
\preformatted{
bpm :  num [1:1277, 1:407] 1 0 0 0 0 0 0 0 0 0 ...
}}
\subsection{cor_app_prefix_string}{
\preformatted{
cor_app_prefix_string :  chr "cor"
}}
\subsection{cor_PU_costs}{
\preformatted{
cor_PU_costs :  num [1:407] 1 1 1 1 1 1 1 1 1 1 ...
}}
\subsection{correct_optimum_landscape_frac_cost}{
\preformatted{
correct_optimum_landscape_frac_cost :  num 0.15
}}
\subsection{cost}{
\preformatted{
cost :  num [1:2] 61 407
}}
\subsection{cost_thresh_for_all_spp_meeting_targets}{
\preformatted{
cost_thresh_for_all_spp_meeting_targets :  num 61
}}
\subsection{cur_cost}{
\preformatted{
cur_cost :  num 407
}}
\subsection{cur_frac_of_all_spp_meeting_their_target}{
\preformatted{
cur_frac_of_all_spp_meeting_their_target :  num 1
}}
\subsection{cur_frac_rep_met_over_optimal_frac_cost}{
\preformatted{
cur_frac_rep_met_over_optimal_frac_cost :  num 0.15
}}
\subsection{cur_landscape_frac_cost}{
\preformatted{
cur_landscape_frac_cost :  num 1
}}
\subsection{cur_num_spp_meeting_their_target}{
\preformatted{
cur_num_spp_meeting_their_target :  int 1277
}}
\subsection{cur_optimal_frac_cost}{
\preformatted{
cur_optimal_frac_cost :  num 6.67
}}
\subsection{cur_rep_fractions}{
\preformatted{
cur_rep_fractions :  num [1:1277] 2 2 2 2 2 2 2 2 2 2 ...
}}
\subsection{cur_run_end_idx_in_PU_IDs}{
\preformatted{
cur_run_end_idx_in_PU_IDs :  num 407
}}
\subsection{cur_run_index}{
\preformatted{
cur_run_index :  num 2
}}
\subsection{cur_run_indices}{
\preformatted{
cur_run_indices :  int [1:346] 62 63 64 65 66 67 68 69 70 71 ...
}}
\subsection{cur_run_length}{
\preformatted{
cur_run_length :  int 346
}}
\subsection{cur_run_start_idx}{
\preformatted{
cur_run_start_idx :  num 408
}}
\subsection{cur_solution_PUs}{
\preformatted{
cur_solution_PUs :  int [1:407] 2 4 6 8 10 12 14 16 18 20 ...
}}
\subsection{frac_of_all_spp_meeting_their_target}{
\preformatted{
frac_of_all_spp_meeting_their_target :  num [1:2] 1 1
}}
\subsection{frac_rep_met_over_optimal_frac_cost}{
\preformatted{
frac_rep_met_over_optimal_frac_cost :  num [1:2] 1 0.15
}}
\subsection{landscape_frac_cost}{
\preformatted{
landscape_frac_cost :  num [1:2] 0.15 1
}}
\subsection{landscape_frac_cost_thresh_for_all_spp_meeting_targets}{
\preformatted{
landscape_frac_cost_thresh_for_all_spp_meeting_targets :  num 0.15
}}
\subsection{marxan_ssoln_df}{
\preformatted{
marxan_ssoln_df : 'data.frame':	407 obs. of  2 variables:
 $ planning_unit: int  1 2 3 4 5 6 7 8 9 10 ...
 $ number       : int  0 4 0 4 0 4 0 4 0 4 ...
}}
\subsection{marxan_ssoln_PUs_ranked_by_votes_df}{
\preformatted{
marxan_ssoln_PUs_ranked_by_votes_df : 'data.frame':	407 obs. of  2 variables:
 $ planning_unit: int  2 4 6 8 10 12 14 16 18 20 ...
 $ number       : int  4 4 4 4 4 4 4 4 4 4 ...
}}
\subsection{num_runs}{
\preformatted{
num_runs :  int 2
}}
\subsection{num_spp}{
\preformatted{
num_spp :  int 1277
}}
\subsection{optimal_frac_cost}{
\preformatted{
optimal_frac_cost :  num [1:2] 1 6.67
}}
\subsection{optimal_frac_cost_thresh_for_all_spp_meeting_targets}{
\preformatted{
optimal_frac_cost_thresh_for_all_spp_meeting_targets :  num 1
}}
\subsection{optimum_cost}{
\preformatted{
optimum_cost :  num 61
}}
\subsection{plot_output_dir}{
\preformatted{
plot_output_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSrun_-COR-Wrap-Marxan_SA.9"| __truncated__
}}
\subsection{rle_lengths_and_values}{
\preformatted{
rle_lengths_and_values : List of 2
 $ lengths: int [1:2] 61 346
 $ values : int [1:2] 4 0
}}
\subsection{thresh_found}{
\preformatted{
thresh_found :  logi TRUE
}}
\subsection{total_landscape_cost}{
\preformatted{
total_landscape_cost :  num 407
}}
}

