% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14a_compute_marxan_solution_scores.R
\name{plot_incremental_marxan_summed_solution_representations}
\alias{plot_incremental_marxan_summed_solution_representations}
\title{plot incrementatl marxan summed solution representations}
\usage{
plot_incremental_marxan_summed_solution_representations(marxan_ssoln_df,
  cor_PU_costs, optimum_cost, bpm, cor_app_prefix_string, num_spp,
  plot_output_dir)
}
\description{
For each step in order by Marxan summed solution PU ID:
 Want the fraction of all species who have met or exceeded their target
 when all PUs with the same number of votes or more are included in the
 solution.
}

