% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_14c_plot_incremental_marxan_summed_solution_reps.R
\name{plot_marxan_best_solution_scores_COR_and_APP}
\alias{plot_marxan_best_solution_scores_COR_and_APP}
\title{Plot marxan best solution scores COR and APP}
\usage{
plot_marxan_best_solution_scores_COR_and_APP(
  plot_output_dir,
  marxan_solution_scores_wrt_COR_reps_and_costs,
  best_solution_ID_according_to_marxan,
  marxan_solution_scores_wrt_APP_reps_and_costs
)
}
\arguments{
\item{plot_output_dir}{character string}

\item{marxan_solution_scores_wrt_COR_reps_and_costs}{data frame}

\item{best_solution_ID_according_to_marxan}{integer}

\item{marxan_solution_scores_wrt_APP_reps_and_costs}{data frame}
}
\value{
Returns nothing
}
\description{
Plot marxan best solution scores for both CORRECT and APPARENT
}
