% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_14_read_marxan_output_files.R
\name{read_marxan_output_files}
\alias{read_marxan_output_files}
\title{Read various marxan outputs into this program.}
\usage{
read_marxan_output_files(marxan_output_dir_path, all_correct_node_IDs, num_spp,
  cor_bpm, plot_output_dir, parameters, app_bpm, cor_PU_costs,
  correct_solution_cost, largest_PU_ID, largest_spp_ID, targets)
}
\description{
Should include these in the marxan package.
 Marxan's best solution.
 Marxan's votes for including each planning unit.
 Marxan output files to read from the Marxan output directory
 (need to look these file names up in the manual):
 output_best.csv
 output_mvbest.csv
 output_ssoln.csv
 output_solutionsmatrix.csv
 output_sum.csv
 output_penalty_planning_units.csv
 output_penalty.csv
 output_sen.dat
}
