% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_parse_Xu_set_cover_problem_file.R
\name{read_Xu_problem_from_Xu_file}
\alias{read_Xu_problem_from_Xu_file}
\title{FUNCTION_TITLE}
\usage{
read_Xu_problem_from_Xu_file(infile_name, correct_solution_cost)
}
\arguments{
\item{infile_name}{File path describing a Xu problem}

\item{correct_solution_cost}{Numeric value of cost of solution (since
problems read from one of Xu's benchmark files have their cost given
somewhere other than in the file)}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}

