% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{remove_base_spp_abundances_from_wrapping_distribution}
\alias{remove_base_spp_abundances_from_wrapping_distribution}
\title{Remove base problem's species from the wrapping problems set of species to
be distributed over the planning units}
\usage{
remove_base_spp_abundances_from_wrapping_distribution(Xu_PU_spp_table,
  trimmed_rounded_abund_per_spp, spp_col_name, allow_imperfect_wrap)
}
\arguments{
\item{Xu_PU_spp_table}{PU_spp table for original Xu problem being wrapped
around}

\item{trimmed_rounded_abund_per_spp}{vector of abundances of all species in
the full wrapped distribution, i.e., including the original Xu problem
abundances}

\item{spp_col_name}{string containing the name of the column in the PU_spp
table that contains the species IDs}
}
\value{
vector of abundances in the wrapping abundance distribution that are
    not also in the original Xu problem
}
\description{
When wrapping one distribution around another, the outside distribution
needs to contain the inside distribution as a proper subset, however,
when it's time to distribute the wrapping distribution over the landscape,
the inside distribution has already been distributed.  That means that the
inside distribution needs to be removed from full wrapping distribution
before its instances are spread across the landscape.  This function
strips that inside distribution out of the full wrapped distribution and
returns a set of abundances ready for spreading.
}
\details{
OLD COMMENT IN calling routine wrap_abundances_around_eligible_set().
Is the old comment correct or has the problem been fixed.
If it has been fixed, then it would be good to leave some comment in here
that describes what is being returned (as it's described in the old
comment's expectations).
 The calling routine assumes THAT the returned value, extra_spp_abund,
 IS A VECTOR CONTAINING THE ABUNDANCE OF EACH EXTRA SPECIES.
 AT THE MOMENT, extra_abund ACTUALLY CONTAINS THE NUMBER OF SPECIES
 HAVING EACH GIVEN ABUNDANCE.
 NEED TO CHANGE EXPAND WHAT'S RETURNED INTO THE EXPECTED VECTOR.
 CAN THAT BE DONE USING ONE OF THE PLYR THINGS LIKE MELT()?
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{base_abund_by_spp}{
\preformatted{
base_abund_by_spp : 'data.frame':	814 obs. of  2 variables:
 $ x   : int  1 2 3 4 5 6 7 8 9 10 ...
 $ freq: int  2 2 2 2 2 2 2 2 2 2 ...
}}
\subsection{base_abund_hist}{
\preformatted{
base_abund_hist : 'data.frame':	1 obs. of  2 variables:
 $ x   : int 2
 $ freq: int 814
}}
\subsection{cur_idx}{
\preformatted{
cur_idx :  int 6
}}
\subsection{end_idx}{
\preformatted{
end_idx :  num 463
}}
\subsection{extra_spp_abund}{
\preformatted{
extra_spp_abund :  num [1:463] 2 2 2 2 2 2 2 2 2 2 ...
}}
\subsection{num_abund_rows}{
\preformatted{
num_abund_rows :  int 6
}}
\subsection{num_extra_spp}{
\preformatted{
num_extra_spp :  num 463
}}
\subsection{spp_col_name}{
\preformatted{
spp_col_name :  chr "spp_ID"
}}
\subsection{start_idx}{
\preformatted{
start_idx :  num 464
}}
\subsection{trimmed_rounded_abund_per_spp}{
\preformatted{
trimmed_rounded_abund_per_spp :  num [1:1277] 2 2 3 3 2 2 2 2 2 2 ...
}}
\subsection{verbose_remove_base}{
\preformatted{
verbose_remove_base :  logi FALSE
}}
\subsection{wrapped_extra_spp_abund_hist}{
\preformatted{
wrapped_extra_spp_abund_hist : 'data.frame':	6 obs. of  2 variables:
 $ abund: num  2 3 4 5 6 7
 $ freq : num  77 311 57 15 2 1
}}
\subsection{wrapped_extra_spp_abund_merge}{
\preformatted{
wrapped_extra_spp_abund_merge : 'data.frame':	6 obs. of  3 variables:
 $ x     : num  2 3 4 5 6 7
 $ freq.x: int  891 311 57 15 2 1
 $ freq.y: num  814 0 0 0 0 0
}}
\subsection{wrapping_abund_hist}{
\preformatted{
wrapping_abund_hist : 'data.frame':	6 obs. of  2 variables:
 $ x   : num  2 3 4 5 6 7
 $ freq: int  891 311 57 15 2 1
}}
\subsection{Xu_PU_spp_table}{
\preformatted{
Xu_PU_spp_table : 'data.frame':	1628 obs. of  2 variables:
 $ PU_ID : int  1 2 3 4 5 6 7 8 9 10 ...
 $ spp_ID: int  1 1 2 2 3 3 4 4 5 5 ...
}}
}

