% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gurobi_analysis_and_output.R
\name{run_gurobi}
\alias{run_gurobi}
\title{Run gurobi linear programming optimizer}
\usage{
run_gurobi(num_spp, num_PUs, bpm, PU_costs, spp_rep_targets,
  use_gap_limit = FALSE, gap_limit = 0.005,
  use_given_time_as_limit = TRUE, time_limit = 60,
  use_marxan_time_as_limit = FALSE, marxan_elapsed_time = NA)
}
\arguments{
\item{num_spp}{integer number of species in the problem}

\item{num_PUs}{integer number of planning units}

\item{bpm}{bipartite matrix; integer matrix with one row for each species
and one column for each planning unit.  Each matrix entry specifies
whether that species occupies that planning unit; 1 indicates the
species does occupy the planning unit and 0 indicates it does not.
Same as occ_matrix.}

\item{PU_costs}{numeric vector of planning unit costs}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}

\item{use_gap_limit}{boolean indicating whether to set a maximum gap allowed
between the result of the optimization computation and the correct
optimum value; TRUE implies set the gap value, FALSE implies no gap
enforced}

\item{gap_limit}{numeric value for gap, measured in same units as the
objective function}

\item{use_given_time_as_limit}{boolean indicating whether to set a time limit on the
optimization computation; TRUE implies set the limit, FALSE implies no
time limit}

\item{time_limit}{numeric value for time limit measured in seconds}

\item{use_marxan_time_as_limit}{boolean indicating whether reserve selector
should use marxan's elapsed run time as the time limit for its own run}

\item{marxan_elapsed_time}{numeric value containing elapsed time for running
marxan (in seconds)}
}
\value{
Returns a gurobi model result list with named elements that contains,
    among other things, an element x which is the solution vector
}
\description{
Run gurobi linear programming optimizer
}
