% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_marxan_control_file_and_run_marxan.R
\name{run_marxan}
\alias{run_marxan}
\title{Run marxan}
\usage{
run_marxan(marxan_dir, marxan_executable_name)
}
\arguments{
\item{marxan_dir}{character string}

\item{marxan_executable_name}{character string}
}
\value{
Returns system return code from running marxan
}
\description{
Run marxan program
}
