% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_marxan_control_file_and_run_marxan.R
\name{run_marxan}
\alias{run_marxan}
\title{Title}
\usage{
run_marxan(marxan_dir, marxan_executable_name)
}
\arguments{
\item{marxan_executable_name}{}
}
\description{
Title
}
