% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{save_rsprob}
\alias{save_rsprob}
\title{Save bd problem to disk}
\usage{
save_rsprob(rsprob, exp_root_dir)
}
\arguments{
\item{rsprob}{a reserve selection problem object, e.g., a Xu_bd_problem}

\item{exp_root_dir}{character string path to root directory of an
experiment, usually parameters$fullOutputDir_NO_slash}
}
\value{
Returns input object with its checksum slot filled
}
\description{
After a problem is generated, its R representation is saved to disk so
that it can be archived and re-used in future experiments without having
to regenerate the problem.  Saving it is also useful for reproducibility
in that it allows re-creation of exactly the problem used in an experiment.
}
\details{
Writes to a file whose name contains:
\itemize{
 \item{the UUID of the problem}
 \item{whether it's a basic problem or a wrapped problem}
 \item{whether it's a correct or an apparent problem}
}
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{base_outdir}{
\preformatted{
base_outdir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/RSprob-COR-Base.d0729e1c-ea"| __truncated__
}}
\subsection{rsprob}{
\preformatted{
rsprob : Formal class 'Xu_bd_problem' [package "bdpg"] with 35 slots
}}
\subsection{exp_root_dir}{
\preformatted{
exp_root_dir :  chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
}

