% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{save_rsprob}
\alias{save_rsprob}
\title{Save bd problem to disk}
\usage{
save_rsprob(rsprob, starting_dir)
}
\description{
After a problem is generated, its R representation is saved to disk so
that it can be archived and re-used in future experiments without having
to regenerate the problem.  Saving it is also useful for reproducibility
in that it allows re-creation of exactly the problem used in an experiment.
}
\details{
Writes to a file whose name contains:
\itemize{
 \item{the UUID of the problem}
 \item{whether it's a basic problem or a wrapped problem}
 \item{whether it's a correct or an apparent problem}
}
}
