% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_15a_write_prob_results.R
\name{save_rsprob_results_data}
\alias{save_rsprob_results_data}
\title{Save problem description results for a Xu_bd_problem}
\usage{
save_rsprob_results_data(rsprob, exp_root_dir, parameters)
}
\arguments{
\item{rsprob}{a reserve selection problem object, e.g., a Xu_bd_problem}

\item{exp_root_dir}{character string path to root directory of an
experiment, usually parameters$fullOutputDir_NO_slash}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}
}
\description{
Create a list of all problem description values for a Xu_bd_problem
and write that list to a file as a 1 line data frame with column
headers.
}
