% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_error_to_spp_occupancy_data.R
\name{set_const_FP_and_FN_rates}
\alias{set_const_FP_and_FN_rates}
\title{Set constant FP and FN rates}
\usage{
set_const_FP_and_FN_rates(parameters, bdpg_error_codes)
}
\arguments{
\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{bdpg_error_codes}{list of error names and corresponding error codes}
}
\value{
Returns FP_and_FN_const_rates list
}
\description{
Set the False Positive and False Negative error rate to either a
given constant value or a constant value chosen from a uniform random
distribution whose upper and lower bounds are given.
}
\details{
As a result, the False
Positive error rate for every PU/spp pair in the problem will be identical.
The False Negative error rate for every PU/spp pair will also be identical
but not necessarily the same value as the False Positive rate.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{bdpg_error_codes}{
\preformatted{
bdpg_error_codes : List of 6
 $ ERROR_STATUS_num_inside_or_within_group_links_less_than_one: num 1001
 $ ERROR_STATUS_optimal_solution_is_not_optimal               : num 1002
 $ ERROR_STATUS_num_nodes_per_group_must_be_at_least_2        : num 1003
 $ ERROR_STATUS_duplicate_spp_in_Xu_input_file                : num 1004
 $ ERROR_STATUS_unknown_spp_occ_FP_error_type                 : num 1005
 $ ERROR_STATUS_unknown_spp_occ_FN_error_type                 : num 1006
}}
\subsection{FN_const_rate}{
\preformatted{
FN_const_rate :  num 0.1
}}
\subsection{FP_and_FN_const_rates}{
\preformatted{
FP_and_FN_const_rates : List of 2
 $ FP_const_rate: num 0.1
 $ FN_const_rate: num 0.1
}}
\subsection{FP_const_rate}{
\preformatted{
FP_const_rate :  num 0.1
}}
\subsection{parameters}{
\preformatted{
parameters : List of 66
 $ summary_without_run_id_filename                           : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress/prob_diff_results_with_0_ru"| __truncated__
 ...
 $ fullOutputDir_NO_slash                                    : chr "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1837_marxan_simulated_annealing.inprogress"
}}
\subsection{spp_occ_FN_error_type}{
\preformatted{
spp_occ_FN_error_type :  chr "CONSTANT"
}}
\subsection{spp_occ_FP_error_type}{
\preformatted{
spp_occ_FP_error_type :  chr "CONSTANT"
}}
}

