% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_marxan_control_file_and_run_marxan.R
\name{set_marxan_controls_and_run_marxan}
\alias{set_marxan_controls_and_run_marxan}
\title{Set marxan controls and run marxan}
\usage{
set_marxan_controls_and_run_marxan(marxan_input_dir, marxan_output_dir,
  marxan_IO_dir, parameters)
}
\arguments{
\item{marxan_input_dir}{character string}

\item{marxan_output_dir}{character string}

\item{marxan_IO_dir}{character string}

\item{parameters}{list}
}
\value{
Returns list
}
\description{
Set marxan controls and run marxan
}
