% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_gurobi_control_file_and_run_gurobi.R
\name{set_up_for_and_run_gurobi}
\alias{set_up_for_and_run_gurobi}
\title{Set up for and run gurobi}
\usage{
set_up_for_and_run_gurobi(num_spp, num_PUs, bpm, PU_costs, spp_rep_targets,
  use_gap_limit, gap_limit, use_given_time_as_limit, time_limit,
  use_marxan_time_as_limit, marxan_elapsed_time = NA)
}
\arguments{
\item{num_spp}{integer number of species in the problem}

\item{num_PUs}{integer number of planning units}

\item{bpm}{bipartite matrix; integer matrix with one row for each species
and one column for each planning unit.  Each matrix entry specifies
whether that species occupies that planning unit; 1 indicates the
species does occupy the planning unit and 0 indicates it does not.
Same as occ_matrix.}

\item{PU_costs}{numeric vector of planning unit costs}

\item{spp_rep_targets}{numeric vector with a target abundance for each
species in the problem}

\item{use_gap_limit}{boolean}

\item{gap_limit}{numeric}

\item{use_given_time_as_limit}{boolean}

\item{time_limit}{numeric}

\item{use_marxan_time_as_limit}{boolean indicating whether reserve selector
should use marxan's elapsed run time as the time limit for its own run}

\item{marxan_elapsed_time}{numeric value containing elapsed time for running
marxan (in seconds)}
}
\value{
Returns named list of gurobi control and return values
}
\description{
Set up for and run gurobi for either COR or APP
}
