% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_13_write_marxan_control_file_and_run_marxan.R
\name{set_up_for_and_run_marxan}
\alias{set_up_for_and_run_marxan}
\title{Title}
\usage{
set_up_for_and_run_marxan(PU_spp_pair_indices, PU_IDs, spp_IDs, PU_col_name,
  spp_col_name, marxan_input_dir, marxan_output_dir, marxan_IO_dir, parameters)
}
\arguments{
\item{marxan_IO_dir}{}
}
\description{
Title
}

