% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/gscp_13_write_marxan_control_file_and_run_marxan.R
\name{set_up_for_and_run_marxan}
\alias{set_up_for_and_run_marxan}
\title{Set up for and run marxan}
\usage{
set_up_for_and_run_marxan(PU_spp_pair_indices, PU_IDs, spp_IDs, PU_col_name,
  spp_col_name, rsrun, num_spp, parameters)
}
\arguments{
\item{PU_spp_pair_indices}{data frame}

\item{PU_IDs}{integer vector}

\item{spp_IDs}{integer vector}

\item{PU_col_name}{character string}

\item{spp_col_name}{character string}

\item{rsrun}{RSrun object}

\item{num_spp}{integer}

\item{parameters}{list}
}
\value{
Returns marxan control values list
}
\description{
Set up for and run marxan for either COR or APP
}
