% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_9a_create_Xu_graph.R
\name{sort_within_rows}
\alias{sort_within_rows}
\title{Sort a 2 column matrix within rows}
\usage{
sort_within_rows(a_2_col_matrix, decreasing = FALSE)
}
\arguments{
\item{a_2_col_matrix}{a 2 column numeric matrix}

\item{decreasing}{boolean flag indicating whether to sort in increasing or
decreasing order; TRUE implies sort in decreasing order, FALSE implies
sort in increasing order}
}
\value{
Returns sorted 2 column matrix
}
\description{
Sort a 2 column matrix within rows
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{a_2_col_matrix}{
\preformatted{
a_2_col_matrix :  int [1:1003, 1:2] 1 3 5 7 9 11 13 15 17 19 ...
}}
\subsection{decreasing}{
\preformatted{
decreasing :  logi FALSE
}}
\subsection{row}{
\preformatted{
row :  int 1003
}}
}

