% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{strip_trailing_slash}
\alias{strip_trailing_slash}
\title{Strip Trailing Slash Off Of Path If There Is One}
\usage{
strip_trailing_slash(path)
}
\arguments{
\item{path}{character string containing a path that may or may not end in a
slash}
}
\value{
character string containing original path argument with trailing
    slash removed if there was one, otherwise, same as original
}
\description{
Tzar puts a slash on the end of the output path and this causes problems
when using it to build full paths to file names.  This function is
primarily here to strip that off but can be used on any path.
}
