% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_pu_and_spp_histograms_from_marxan_pusvpr_file.R
\name{summarize_pu_and_spp_dist_data_from_marxan_puvspr_input_file}
\alias{summarize_pu_and_spp_dist_data_from_marxan_puvspr_input_file}
\title{Title}
\usage{
summarize_pu_and_spp_dist_data_from_marxan_puvspr_input_file(infile_path)
}
\arguments{
\item{infile_path}{character string giving path to the marxan puvspr.dat
file to be analyzed (including the name of the file itself), e.g.,
"~/marxan/input/puvspr.dat"}
}
\description{
Title
}
\examples{
\dontrun{
infile_path <- "/Users/bill/tzar/outputdata/biodivprobgen/default_runset/1834_marxan_simulated_annealing.completedTzarEmulation/RSrun_-COR-Wrap-Marxan_SA.408782d5-2bb6-4c09-8b18-f05ad3089b15/input/puvspr.dat"
pu_and_spp_dist_data <-
  summarize_pu_and_spp_dist_data_from_marxan_puvspr_input_file (infile_path)
}
}
