% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compute_pu_and_spp_histograms_from_marxan_pusvpr_file.R
\name{summarize_pu_and_spp_dist_data_from_pu_spp_amt_df}
\alias{summarize_pu_and_spp_dist_data_from_pu_spp_amt_df}
\title{Compute distributions for planning units and species from marxan input file}
\usage{
summarize_pu_and_spp_dist_data_from_pu_spp_amt_df(pu_spp_amt_df)
}
\arguments{
\item{pu_spp_amt_df}{data frame containing the marxan planning unit vs.
species data (i.e., data frame with columns for pu, species, and amount).}
}
\value{
Returns list containing 3 named elements, each of which is a data frame:
    num_spp_per_patch, num_PUs_per_spp, and amt_per_spp.
}
\description{
This function groups the data by planning unit to summarize the number of
species on each planning unit, then regroups the data by species to compute
the number of patches for each species and the total amount of each species
across all patches.  All 3 of these distributions are returned, each as a
distribution.  Each of them is also plotted.
}
