% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gurobi_analysis_and_output.R
\name{test_gurobi}
\alias{test_gurobi}
\title{Title}
\usage{
test_gurobi(seed = 456, num_spp = 10, num_PUs = 5)
}
\arguments{
\item{seed}{numeric}

\item{num_spp}{integer number of species in the problem}

\item{num_PUs}{integer number of planning units}
}
\value{
Returns a gurobi model result list with named elements that contains,
    among other things, an element x which is the solution vector
}
\description{
Title
}
