% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{touch}
\alias{touch}
\title{Imitate unix touch function to create an empty file.}
\usage{
touch(file_path_to_touch)
}
\arguments{
\item{file_path_to_touch}{character string}
}
\value{
Returns nothing
}
\description{
I couldn't find an existing function for this, but did find a
 stack overflow question that suggested using write.table() on an
 empty table, so that's what I've done.
}
