% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{trim_abundances}
\alias{trim_abundances}
\title{Drop spp from abund dist if on too many or too few patches}
\usage{
trim_abundances(rounded_abundances, min_abund = 2,
  max_abund = .Machine$double.xmax)
}
\arguments{
\item{rounded_abundances}{vector of abundances to be trimmed}

\item{min_abund}{lowest abundance to allow in the trimmed set}

\item{max_abund}{largest abundance to allow in the trimmed set}
}
\value{
vector of abundances whose values lie in the specified range
}
\description{
This function is primarily to get rid of species that occur on
 only 0 or 1 patches, however, some distribution generator might also
 generate species that are too common as well, so this routine will
 axe those too.
}
\details{
Abundances are expressed here as a vector of count values where each value
 is the count of the number of patches a particular species occurs on.
 For example, [1 1 1 2 6 6] would represent 3 species that occur
 on exactly one patch, 1 species that occurs on 2 patches, and 2 species
 that occur on 6 patches each.  Calling this routine with a min_abund of 2
 and a max_abund of 5 would return a 1 element vector, i.e., the vector [2].

This doesn't really need to be a function since it's only one line
 long, but I built an earlier version that was more verbose so that
 I could monitor how many things were being cut out.
 That version is in the git repository in case it needs to be
 resurrected for some reason.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{max_abund}{
\preformatted{
max_abund :  num 1.8e+308
}}
\subsection{min_abund}{
\preformatted{
min_abund :  num 2
}}
\subsection{rounded_abundances}{
\preformatted{
rounded_abundances :  num [1:1628] 2 2 1 1 3 1 3 2 2 1 ...
}}
\subsection{trimmed_rounded_abund_per_spp}{
\preformatted{
trimmed_rounded_abund_per_spp :  num [1:1277] 2 2 3 3 2 2 2 2 2 2 ...
}}
}

