% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_multi_bdprob.R
\name{wrap_abundance_dist_around_Xu_problem}
\alias{wrap_abundance_dist_around_Xu_problem}
\title{FUNCTION_TITLE}
\usage{
wrap_abundance_dist_around_Xu_problem(rounded_abundances, Xu_bdprob,
  dep_set_PUs_eligible, tot_num_PUs_in_landscape)
}
\arguments{
\item{rounded_abundances}{DESCRIPTION.}

\item{Xu_bdprob}{DESCRIPTION.}

\item{dep_set_PUs_eligible}{DESCRIPTION.}

\item{tot_num_PUs_in_landscape}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}

