% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{wrap_abundance_dist_around_Xu_problem}
\alias{wrap_abundance_dist_around_Xu_problem}
\title{Wrap abundance distribution around Xu problem}
\usage{
wrap_abundance_dist_around_Xu_problem(
  starting_dir,
  parameters,
  rounded_abundances,
  Xu_bdprob,
  dep_set_PUs_eligible,
  tot_num_PUs_in_landscape,
  seed_value_for_search_list,
  allow_imperfect_wrap,
  search_outfile_name_base,
  search_outfile_name,
  wrap_prob_name_stem = "wrap_prob",
  cor_dir_name_stem = "cor"
)
}
\arguments{
\item{starting_dir}{character string}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{rounded_abundances}{DESCRIPTION.}

\item{Xu_bdprob}{DESCRIPTION.}

\item{dep_set_PUs_eligible}{DESCRIPTION.}

\item{tot_num_PUs_in_landscape}{DESCRIPTION.}

\item{seed_value_for_search_list}{blah}

\item{allow_imperfect_wrap}{blah}

\item{search_outfile_name_base}{character string}

\item{search_outfile_name}{character string}

\item{wrap_prob_name_stem}{string}

\item{cor_dir_name_stem}{string}
}
\value{
Returns a Xu_wrapped_bd_problem
}
\description{
Note that this function doesn't care where you got the abundances,
 i.e., they don't have to have come from the lognormal generator.
 It can be any abundance set that you want, as long as it contains
 at least as many species sitting on exactly 2 patches as the base
 Xu problem has.  It can have more than the Xu problem, but not less.
}
