% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wrapped_bdprob.R
\name{wrap_abundances_around_eligible_set}
\alias{wrap_abundances_around_eligible_set}
\title{Wrap abundances around eligible set}
\usage{
wrap_abundances_around_eligible_set(dep_set, eligible_set,
  rounded_abund_per_spp, num_base_spp, Xu_PU_spp_table,
  allow_imperfect_wrap, min_allowed_abundance = 2,
  PU_col_name = "PU_ID", spp_col_name = "spp_ID",
  use_testing_only_rand_seed = FALSE, testing_only_rand_seed = 17)
}
\arguments{
\item{dep_set}{integer vector}

\item{eligible_set}{integer vector}

\item{rounded_abund_per_spp}{integer vector}

\item{num_base_spp}{integer}

\item{Xu_PU_spp_table}{data frame}

\item{allow_imperfect_wrap}{boolean}

\item{min_allowed_abundance}{integer}

\item{PU_col_name}{character string}

\item{spp_col_name}{character string}

\item{use_testing_only_rand_seed}{boolean}

\item{testing_only_rand_seed}{integer}
}
\value{
Returns PU_spp_table data frame
}
\description{
Take a given distribution (e.g., from a Xu problem) and add more species
(and probably more planning units too) to the problem so that the new
problem has a specified distribution, but retains the original problem's
solution.
}
