% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_marxan_input_files.R
\name{write_marxan_pu.dat_input_file_from_vectors}
\alias{write_marxan_pu.dat_input_file_from_vectors}
\title{Write marxan planning units input file (pu.dat)  given vectors of cost and status values.}
\usage{
write_marxan_pu.dat_input_file_from_vectors(PU_IDs, cost_values, status_values)
}
\arguments{
\item{PU_IDs}{A vector of planning unit IDs}

\item{cost_values}{A vector of costs}

\item{status_values}{A vector of status values}
}
\value{
nothing.
}
\description{
Write marxan planning units input file (pu.dat)  given vectors of cost and status values.
}
\examples{
\dontrun{
PU_IDs = 1:5
cost_values = c(2,3,8,1,4)
status_values = c(0,0,0,0,0)
write_marxan_spec.dat_input_file_from_vectors (PU_IDs, cost_values, status_values)
         }
}
