% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_marxan_input_files.R
\name{write_marxan_puvspr.dat_input_file}
\alias{write_marxan_puvspr.dat_input_file}
\title{Write marxan puvspr input file}
\usage{
write_marxan_puvspr.dat_input_file(spp_PU_amount_table)
}
\arguments{
\item{spp_PU_amount_table}{A data frame of species IDs vs planning unit IDs}
}
\value{
nothing.
}
\description{
Write planning unit vs. species input file (puvspr.dat).
}
\details{
The spp_PU_amount_table is expected to have 3 columns called (in left to
right order) "species",
"pu", and "amount".  Marxan requires this table to be sorted in increasing
order on the planning unit column.  I don't think the order within planning
unit matters on the
other columns.  See source code for the test function
\code{gen_random_spp_PU_amount_table} for an example of creating
this table.
}
\section{Local Variable Structures and examples}{

Here is the output of str() for each variable visible in the function.
Note that the particular counts and values given are just examples to show
what the data might look like.

\subsection{spp_PU_amount_table}{
\preformatted{
spp_PU_amount_table : 'data.frame':	3037 obs. of  3 variables:
 $ species: num  1 1 65 88 139 215 227 256 259 262 ...
 $ pu     : num  1 2 2 2 2 2 2 2 2 2 ...
 $ amount : num  1 1 1 1 1 1 1 1 1 1 ...
}}
}

\examples{
\dontrun{
num_PUs = 100
num_spp = 3
spp_PU_amount_table = gen_random_spp_PU_amount_table (num_PUs, num_spp)
write_marxan_puvspr.dat_input_file (spp_PU_amount_table)
         }
}
