% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_12_write_network_to_marxan_files.R
\name{write_network_to_marxan_files}
\alias{write_network_to_marxan_files}
\title{Write network to marxan input file}
\usage{
write_network_to_marxan_files(PU_spp_pair_indices, PU_IDs, spp_IDs,
  PU_col_name, spp_col_name, parameters, marxan_input_dir,
  marxan_output_dir, spf_const, targets = rep(1, length(spp_IDs)),
  costs = rep(1, length(PU_IDs)))
}
\arguments{
\item{PU_spp_pair_indices}{data frame}

\item{PU_IDs}{integer vector}

\item{spp_IDs}{integer vector}

\item{PU_col_name}{character string}

\item{spp_col_name}{character string}

\item{parameters}{lilst}

\item{marxan_input_dir}{character string}

\item{marxan_output_dir}{character string}

\item{spf_const}{numeric}

\item{targets}{numeric vector}

\item{costs}{numeric vector}
}
\value{
numeric spf_const
}
\description{
Write the network as a marxan input file.
}
