% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscp_12_write_network_to_marxan_files.R
\name{write_network_to_marxan_files}
\alias{write_network_to_marxan_files}
\title{Title}
\usage{
write_network_to_marxan_files(PU_spp_pair_indices, PU_IDs, spp_IDs, PU_col_name,
  spp_col_name, parameters, marxan_input_dir, marxan_output_dir, spf_const,
  targets = rep(1, length(spp_IDs)), costs = rep(1, length(PU_IDs)))
}
\arguments{
\item{costs}{}
}
\description{
Title
}
