% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivprobgen_utilities.R
\name{write_results_to_files}
\alias{write_results_to_files}
\title{Write results to files}
\usage{
write_results_to_files(csv_outfile_name, results_df, parameters, out_dir,
  tzar_run_id_field_name)
}
\arguments{
\item{csv_outfile_name}{Name of the csv file to write results to, not
including csv extension but not including path}

\item{results_df}{data frame containing the results}

\item{parameters}{parameters list for the run, usually derived from project.yaml
and can have a varying number and set of elements depending on the run}

\item{out_dir}{character string telling what directory to put the results
file in}

\item{cur_result_row}{integer row number in data frame where run_ID should
be added}
}
\value{
Returns nothing.
}
\description{
This function is used two different ways.
It's called when the program quits because there are too many species
or it's called when the program runs successfully.
}
